Core

Provide basic user's interfaces for the ACFlow toolkit.

Contents

Core

Contents

Index

Solvers

Parameters

Data

Grids

Meshes

Models

Kernels

Postprocessing

Index

ACFlow.kramers

ACFlow.make_data

ACFlow.make_grid

ACFlow.make_kernel

ACFlow.make_mesh

ACFlow.make_model

ACFlow.read_data

ACFlow.read_param

ACFlow.reprod

ACFlow.setup_param

ACFlow.solve

ACFlow.solve

ACFlow.solve

ACFlow.solve

Solvers

ACFlow.solve — Method

solve(grid::Vector{F64}, Gval::Vector{T}, Gerr::Vector{T})

Solve the analytic continuation problem. The arguments grid, Gval, and Gerr are the grid, value, and error bar, respectively.

ACFlow.solve — Method

solve(grid::Vector{F64}, Gval::Vector{T}, err::T)

Solve the analytic continuation problem. The arguments grid, Gval, and err are the grid, value, and error bar, respectively.

ACFlow.solve — Method

solve(grid::Vector{F64}, Gval::Vector{T})

Solve the analytic continuation problem. The arguments grid and Gval are the grid and value, respectively. Furthermore, the error bar is set to a fixed value 1.0e-4.

ACFlow.solve — Method

solve(rd::RawData)

Solve the analytic continuation problem. The input data are encapsulated in a RawData struct. This function call is the actual interface to the desired analytic continuation solvers.

See also: RawData.

Parameters

ACFlow.setup_param — Function

setup_param(C::Dict{String,Any}, S::Dict{String,Any}, reset::Bool = true)

Setup the configuration dictionaries via function call. Here C contains parameters for general setup, while S contains parameters for selected analytic continuation solver. If reset is true, then the configuration dictionaries will be reset to their default values at first. Later, C S will be used to customized the dictionaries further.

See also: read_param.

ACFlow.read_param — Function

read_param()

Setup the configuration dictionaries via an external file. The valid format of a configuration file is toml.

See also: read_param.

Data

ACFlow.read_data — Function

read_data(only_real_part::Bool = true)

Read data in imaginary axis and return a RawData struct.

See also: RawData.

ACFlow.make_data — Function

make_data(rd::RawData; T::DataType = F64)

Convert RawData struct to GreenData struct. Note that RawData is provided by the users directly, while GreenData is more suitable for various analytic continuation solvers and algorithms. Note that the GreenData struct is accessed and manipulated by this code internally, while the RawData struct is exposed to the users.

See also: RawData, GreenData.

Grids

ACFlow.make_grid — Function

make_grid(rd::RawData; T::DataType = F64)

Extract grid for input data from a RawData struct. It will return a sub-type of the AbstractGrid struct.

See also: RawData, AbstractGrid.

Meshes

ACFlow.make_mesh — Function

make_mesh(; T::DataType = F64)

Try to generate an uniform (linear) or non-uniform (non-linear) mesh for the spectral function in real axis. Notice that it supports arbitrary precision mesh. By default, the precision is F64. One can specify the precision by the argument T.

See also: LinearMesh, TangentMesh, LorentzMesh.

Models

ACFlow.make_model — Function

make_model(am::AbstractMesh)

Try to generate a default model function at given mesh am through various schemes.

See also: AbstractMesh.

Kernels

ACFlow.make_kernel — Function

make_kernel(am::AbstractMesh, ag::AbstractGrid)

Try to generate various kernel functions.

See also: AbstractMesh, AbstractGrid.

Postprocessing

ACFlow.reprod — Function

reprod(am::AbstractMesh, kernel::Matrix{F64}, A::Vector{F64})

Try to reproduce the input data using the calculated spectrum function A. kernel is the kernel function, and am is the mesh in which the spectrum is defined.

See also: AbstractMesh.

ACFlow.kramers — Function

kramers(am::AbstractMesh, A::Vector{F64})

Try to calculate the real part of the green's function from its imaginary part via the Kramers-Kronig relations.