Configuration

Read, parse, and extract the configuration parameters.

Contents

Configuration

Contents

Index

Parser

Validation

Extraction

Index

ACFlow._v

ACFlow.chk_dict

ACFlow.fil_dict

ACFlow.get_a

ACFlow.get_b

ACFlow.get_k

ACFlow.get_m

ACFlow.get_n

ACFlow.get_r

ACFlow.get_s

ACFlow.get_x

ACFlow.inp_toml

ACFlow.rev_dict_a

ACFlow.rev_dict_b

ACFlow.rev_dict_k

ACFlow.rev_dict_m

ACFlow.rev_dict_n

ACFlow.rev_dict_r

ACFlow.rev_dict_s

ACFlow.rev_dict_x

ACFlow.see_dict

Parser

ACFlow.inp_toml — Function

inp_toml(f::String, key::String, necessary::Bool)

Parse the configuration file (in toml format). It reads only parts of the configuration file, which depends on the value of key.

inp_toml(f::String, necessary::Bool)

Parse the configuration file (in toml format). It reads the whole file.

ACFlow.fil_dict — Function

fil_dict(cfg::Dict{String,Any})

Transfer configurations from dict cfg to internal dicts (including PBASE, PMaxEnt, PBarRat, PNevanAC, PStochAC, PStochSK, PStochOM and PStochPX etc).

ACFlow.see_dict — Function

see_dict()

Display all of the relevant configuration parameters to the terminal.

See also: fil_dict.

ACFlow.rev_dict_b — Function

rev_dict_b(BASE::Dict{String,Any})

Setup the configuration dictionary: PBASE.

See also: PBASE.

rev_dict_b(BASE::Dict{String,Vector{Any}})

Setup the configuration dictionary: PBASE.

See also: PBASE.

ACFlow.rev_dict_m — Function

rev_dict_m(S::MaxEntSolver, MaxEnt::Dict{String,Any})

Setup the configuration dictionary: PMaxEnt.

See also: PMaxEnt.

rev_dict_m(S::MaxEntSolver, MaxEnt::Dict{String,Vector{Any}})

Setup the configuration dictionary: PMaxEnt.

See also: PMaxEnt.

ACFlow.rev_dict_r — Function

rev_dict_r(S::BarRatSolver, BarRat::Dict{String,Any})

Setup the configuration dictionary: PBarRat.

See also: PBarRat.

rev_dict_r(S::BarRatSolver, BarRat::Dict{String,Vector{Any}})

Setup the configuration dictionary: PBarRat.

See also: PBarRat.

ACFlow.rev_dict_n — Function

rev_dict_n(S::NevanACSolver, NevanAC::Dict{String,Any})

Setup the configuration dictionary: PNevanAC.

See also: PNevanAC.

rev_dict_n(S::NevanACSolver, NevanAC::Dict{String,Vector{Any}})

Setup the configuration dictionary: PNevanAC.

See also: PNevanAC.

ACFlow.rev_dict_a — Function

rev_dict_a(S::StochACSolver, StochAC::Dict{String,Any})

Setup the configuration dictionary: PStochAC.

See also: PStochAC.

rev_dict_a(S::StochACSolver, StochAC::Dict{String,Vector{Any}})

Setup the configuration dictionary: PStochAC.

See also: PStochAC.

ACFlow.rev_dict_k — Function

rev_dict_k(S::StochSKSolver, StochSK::Dict{String,Any})

Setup the configuration dictionary: PStochSK.

See also: PStochSK.

rev_dict_k(S::StochSKSolver, StochSK::Dict{String,Vector{Any}})

Setup the configuration dictionary: PStochSK.

See also: PStochSK.

ACFlow.rev_dict_s — Function

rev_dict_s(S::StochOMSolver, StochOM::Dict{String,Any})

Setup the configuration dictionary: PStochOM.

See also: PStochOM.

rev_dict_s(S::StochOMSolver, StochOM::Dict{String,Vector{Any}})

Setup the configuration dictionary: PStochOM.

See also: PStochOM.

ACFlow.rev_dict_x — Function

rev_dict_x(S::StochPXSolver, StochPX::Dict{String,Any})

Setup the configuration dictionary: PStochPX.

See also: PStochPX.

rev_dict_x(S::StochPXSolver, StochPX::Dict{String,Vector{Any}})

Setup the configuration dictionary: PStochPX.

See also: PStochPX.

Validation

ACFlow.chk_dict — Function

chk_dict()

Validate the correctness and consistency of configurations.

See also: fil_dict, _v.

ACFlow._v — Function

_v(key::String, val::Array{Any,1})

Verify the value array. Called by chk_dict() function only.

See also: chk_dict.

Extraction

ACFlow.get_b — Function

get_b(key::String)

Extract configurations from dict: PBASE.

See also: PBASE.

ACFlow.get_m — Function

get_m(key::String)

Extract configurations from dict: PMaxEnt.

See also: PMaxEnt.

ACFlow.get_r — Function

get_r(key::String)

Extract configurations from dict: PBarRat.

See also: PBarRat.

ACFlow.get_n — Function

get_n(key::String)

Extract configurations from dict: PNevanAC.

See also: PNevanAC.

ACFlow.get_a — Function

get_a(key::String)

Extract configurations from dict: PStochAC.

See also: PStochAC.

ACFlow.get_k — Function

get_k(key::String)

Extract configurations from dict: PStochSK.

See also: PStochSK.

ACFlow.get_s — Function

get_s(key::String)

Extract configurations from dict: PStochOM.

See also: PStochOM.

ACFlow.get_x — Function

get_x(key::String)

Extract configurations from dict: PStochPX.

See also: PStochPX.