Grids

Grids on imaginary axis.

In the ACFlow package, the input correlators are defined on grid, while the calculated spectral functions are defined on mesh. The ACFlow package supports both imaginary-time and Matsubara frequency grids.

Contents

Grids

Contents

Index

Types

Constructors

Base.* Functions

Utilities

Index

ACFlow.AbstractGrid

ACFlow.BosonicImaginaryTimeGrid

ACFlow.BosonicImaginaryTimeGrid

ACFlow.BosonicMatsubaraGrid

ACFlow.BosonicMatsubaraGrid

ACFlow.FermionicImaginaryTimeGrid

ACFlow.FermionicImaginaryTimeGrid

ACFlow.FermionicMatsubaraGrid

ACFlow.FermionicMatsubaraGrid

ACFlow.rebuild

Base.eachindex

Base.eachindex

Base.eachindex

Base.eachindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.length

Base.length

Base.length

Base.length

Types

ACFlow.AbstractGrid — Type

AbstractGrid

An abstract type representing the imaginary axis. It is used to build the internal type system.

ACFlow.FermionicImaginaryTimeGrid — Type

FermionicImaginaryTimeGrid

Mutable struct. It represents the fermionic imaginary time grid.

Members

ntime -> Number of time slices.

β -> Inverse temperature.

τ -> Vector of grid points.

See also: FermionicMatsubaraGrid.

ACFlow.FermionicMatsubaraGrid — Type

FermionicMatsubaraGrid

Mutable struct. It represents the fermionic Matsubara frequency grid.

Members

nfreq -> Number of Matsubara frequency points.

β -> Inverse temperature.

ω -> Vector of grid points.

See also: FermionicImaginaryTimeGrid.

ACFlow.BosonicImaginaryTimeGrid — Type

BosonicImaginaryTimeGrid

Mutable struct. It represents the bosonic imaginary time grid.

Members

ntime -> Number of time slices.

β -> Inverse temperature.

τ -> Vector of grid points.

See also: BosonicMatsubaraGrid.

ACFlow.BosonicMatsubaraGrid — Type

BosonicMatsubaraGrid

Mutable struct. It represents the bosonic Matsubara frequency grid.

Members

nfreq -> Number of Matsubara frequency points.

β -> Inverse temperature.

ω -> Vector of grid points.

See also: BosonicImaginaryTimeGrid.

Constructors

ACFlow.FermionicImaginaryTimeGrid — Method

FermionicImaginaryTimeGrid(ntime::I64, β::F64)

A constructor for the FermionicImaginaryTimeGrid struct, which is defined in src/types.jl.

See also: FermionicImaginaryTimeGrid.

ACFlow.FermionicMatsubaraGrid — Method

FermionicMatsubaraGrid(nfreq::I64, β::F64)

A constructor for the FermionicMatsubaraGrid struct, which is defined in src/types.jl. The Matsubara grid is evaluated as ωₙ = (2n - 1) π / β.

See also: FermionicMatsubaraGrid.

ACFlow.BosonicImaginaryTimeGrid — Method

BosonicImaginaryTimeGrid(ntime::I64, β::F64)

A constructor for the BosonicImaginaryTimeGrid struct, which is defined in src/types.jl.

See also: BosonicImaginaryTimeGrid.

ACFlow.BosonicMatsubaraGrid — Method

BosonicMatsubaraGrid(nfreq::I64, β::F64)

A constructor for the BosonicMatsubaraGrid struct, which is defined in src/types.jl. The Matsubara grid is evaluated as ωₙ = (2n - 2) π / β.

See also: BosonicMatsubaraGrid.

Base.* Functions

Base.length — Method

Base.length(fg::FermionicImaginaryTimeGrid)

Return number of grid points in a FermionicImaginaryTimeGrid struct.

See also: FermionicImaginaryTimeGrid.

Base.length — Method

Base.length(fg::FermionicMatsubaraGrid)

Return number of grid points in a FermionicMatsubaraGrid struct.

See also: FermionicMatsubaraGrid.

Base.length — Method

Base.length(bg::BosonicImaginaryTimeGrid)

Return number of grid points in a BosonicImaginaryTimeGrid struct.

See also: BosonicImaginaryTimeGrid.

Base.length — Method

Base.length(bg::BosonicMatsubaraGrid)

Return number of grid points in a BosonicMatsubaraGrid struct.

See also: BosonicMatsubaraGrid.

Base.iterate — Method

Base.iterate(fg::FermionicImaginaryTimeGrid)

Advance the iterator of a FermionicImaginaryTimeGrid struct to obtain the next grid point.

See also: FermionicImaginaryTimeGrid.

Base.iterate — Method

Base.iterate(fg::FermionicMatsubaraGrid)

Advance the iterator of a FermionicMatsubaraGrid struct to obtain the next grid point.

See also: FermionicMatsubaraGrid.

Base.iterate — Method

Base.iterate(bg::BosonicImaginaryTimeGrid)

Advance the iterator of a BosonicImaginaryTimeGrid struct to obtain the next grid point.

See also: BosonicImaginaryTimeGrid.

Base.iterate — Method

Base.iterate(bg::BosonicMatsubaraGrid)

Advance the iterator of a BosonicMatsubaraGrid struct to obtain the next grid point.

See also: BosonicMatsubaraGrid.

Base.iterate — Method

Base.iterate(fg::FermionicImaginaryTimeGrid, i::I64)

This is the key method that allows a FermionicImaginaryTimeGrid struct to be iterated, yielding a sequences of grid points.

See also: FermionicImaginaryTimeGrid.

Base.iterate — Method

Base.iterate(fg::FermionicMatsubaraGrid, i::I64)

Create an iterable object for visiting each index of a FermionicMatsubaraGrid struct.

See also: FermionicMatsubaraGrid.

Base.iterate — Method

Base.iterate(bg::BosonicImaginaryTimeGrid, i::I64)

Create an iterable object for visiting each index of a BosonicImaginaryTimeGrid struct.

See also: BosonicImaginaryTimeGrid.

Base.iterate — Method

Base.iterate(bg::BosonicMatsubaraGrid, i::I64)

Create an iterable object for visiting each index of a BosonicMatsubaraGrid struct.

See also: BosonicMatsubaraGrid.

Base.eachindex — Method

Base.eachindex(fg::FermionicImaginaryTimeGrid)

Create an iterable object for visiting each index of a FermionicImaginaryTimeGrid struct.

See also: FermionicImaginaryTimeGrid.

Base.eachindex — Method

Base.eachindex(fg::FermionicMatsubaraGrid)

Create an iterable object for visiting each index of a FermionicMatsubaraGrid struct.

See also: FermionicMatsubaraGrid.

Base.eachindex — Method

Base.eachindex(bg::BosonicImaginaryTimeGrid)

Create an iterable object for visiting each index of a BosonicImaginaryTimeGrid struct.

See also: BosonicImaginaryTimeGrid.

Base.eachindex — Method

Base.eachindex(bg::BosonicMatsubaraGrid)

Create an iterable object for visiting each index of a BosonicMatsubaraGrid struct.

See also: BosonicMatsubaraGrid.

Base.firstindex — Method

Base.firstindex(fg::FermionicImaginaryTimeGrid)

Return the first index of a FermionicImaginaryTimeGrid struct.

See also: FermionicImaginaryTimeGrid.

Base.firstindex — Method

Base.firstindex(fg::FermionicMatsubaraGrid)

Return the first index of a FermionicMatsubaraGrid struct.

See also: FermionicMatsubaraGrid.

Base.firstindex — Method

Base.firstindex(bg::BosonicImaginaryTimeGrid)

Return the first index of a BosonicImaginaryTimeGrid struct.

See also: BosonicImaginaryTimeGrid.

Base.firstindex — Method

Base.firstindex(bg::BosonicMatsubaraGrid)

Return the first index of a BosonicMatsubaraGrid struct.

See also: BosonicMatsubaraGrid.

Base.lastindex — Method

Base.lastindex(fg::FermionicImaginaryTimeGrid)

Return the last index of a FermionicImaginaryTimeGrid struct.

See also: FermionicImaginaryTimeGrid.

Base.lastindex — Method

Base.lastindex(fg::FermionicMatsubaraGrid)

Return the last index of a FermionicMatsubaraGrid struct.

See also: FermionicMatsubaraGrid.

Base.lastindex — Method

Base.lastindex(bg::BosonicImaginaryTimeGrid)

Return the last index of a BosonicImaginaryTimeGrid struct.

See also: BosonicImaginaryTimeGrid.

Base.lastindex — Method

Base.lastindex(bg::BosonicMatsubaraGrid)

Return the last index of a BosonicMatsubaraGrid struct.

See also: BosonicMatsubaraGrid.

Base.getindex — Method

Base.getindex(fg::FermionicImaginaryTimeGrid, ind::I64)

Retrieve the value(s) stored at the given key or index within a FermionicImaginaryTimeGrid struct.

See also: FermionicImaginaryTimeGrid.

Base.getindex — Method

Base.getindex(fg::FermionicMatsubaraGrid, ind::I64)

Retrieve the value(s) stored at the given key or index within a FermionicMatsubaraGrid struct.

See also: FermionicMatsubaraGrid.

Base.getindex — Method

Base.getindex(bg::BosonicImaginaryTimeGrid, ind::I64)

Retrieve the value(s) stored at the given key or index within a BosonicImaginaryTimeGrid struct.

See also: BosonicImaginaryTimeGrid.

Base.getindex — Method

Base.getindex(bg::BosonicMatsubaraGrid, ind::I64)

Retrieve the value(s) stored at the given key or index within a BosonicMatsubaraGrid struct.

See also: BosonicMatsubaraGrid.

Base.getindex — Method

Base.getindex(fg::FermionicImaginaryTimeGrid, I::UnitRange{I64})

Return a subset of a FermionicImaginaryTimeGrid struct as specified by I.

See also: FermionicImaginaryTimeGrid.

Base.getindex — Method

Base.getindex(fg::FermionicMatsubaraGrid, I::UnitRange{I64})

Return a subset of a FermionicMatsubaraGrid struct as specified by I.

See also: FermionicMatsubaraGrid.

Base.getindex — Method

Base.getindex(bg::BosonicImaginaryTimeGrid, I::UnitRange{I64})

Return a subset of a BosonicImaginaryTimeGrid struct as specified by I.

See also: BosonicImaginaryTimeGrid.

Base.getindex — Method

Base.getindex(bg::BosonicMatsubaraGrid, I::UnitRange{I64})

Return a subset of a BosonicMatsubaraGrid struct as specified by I.

See also: BosonicMatsubaraGrid.

Utilities

ACFlow.rebuild — Function

rebuild(fg::FermionicImaginaryTimeGrid, ntime::I64, β::F64)

Rebuild the FermionicImaginaryTimeGrid struct via new ntime and β parameters.

See also: FermionicImaginaryTimeGrid.

rebuild(fg::FermionicMatsubaraGrid, nfreq::I64, β::F64)

Rebuild the FermionicMatsubaraGrid struct via new nfreq and β parameters.

See also: FermionicMatsubaraGrid.

rebuild(bg::BosonicImaginaryTimeGrid, ntime::I64, β::F64)

Rebuild the BosonicImaginaryTimeGrid struct via new ntime and β parameters.

See also: BosonicImaginaryTimeGrid.

rebuild(bg::BosonicMatsubaraGrid, nfreq::I64, β::F64)

Rebuild the BosonicMatsubaraGrid struct via new nfreq and β parameters.

See also: BosonicMatsubaraGrid.