Math

Define some essential mathematical functions.

Contents

Math

Contents

Index

Root Finding

Numerical Integrations

Interpolations

Structs

Constructors

Functions

Einstein Summation Convention

Curve Fitting

Structs

Constructors

Functions

Index

ACFlow.AbstractInterpolation

ACFlow.CubicSplineInterpolation

ACFlow.CubicSplineInterpolation

ACFlow.LinearInterpolation

ACFlow.LinearInterpolation

ACFlow.LsqFitResult

ACFlow.OnceDifferentiable

ACFlow.OnceDifferentiable

ACFlow.OptimizationResults

ACFlow.QuadraticInterpolation

ACFlow.QuadraticInterpolation

ACFlow._interp

ACFlow.curve_fit

ACFlow.jacobian

ACFlow.jacobian!

ACFlow.levenberg_marquardt

ACFlow.munge_data

ACFlow.newton

ACFlow.secant

ACFlow.simpson

ACFlow.trapz

ACFlow.value

ACFlow.value!

ACFlow.@einsum

Root Finding

ACFlow.secant — Function

secant(func, x0, args...)

It implements the well-known secant algorithm to locate root of a given polynomial function. Here, func means the function, x0 is the initial guess, and args... denotes the required arguments for function call to func. In addition, the maximum iterations and tolerance are controled by maxiter and tol, respectively. Be careful, func must be a single variable function.

See also: newton.

ACFlow.newton — Function

newton(fun::Function, guess, kwargs...;
       maxiter::I64 = 20000, mixing::F64 = 0.5)

It implements the well-known newton algorithm to locate root of a given polynomial function. Here, fun means the function, guess is the initial solution, and kwargs... denotes the required arguments for fun. Please be careful, func is a multiple variable function. It not only returns the value, but also the jacobian matrix of the function.

See also: secant.

Numerical Integrations

ACFlow.trapz — Function

trapz(x::AbstractMesh, y::AbstractVector{T})

Perform numerical integration by using the composite trapezoidal rule.

See also: simpson.

trapz(x::Vector{F64}, y::Vector{T}, linear::Bool = false)

Perform numerical integration by using the composite trapezoidal rule.

See also: simpson.

ACFlow.simpson — Function

simpson(x::AbstractVector{F64}, y::AbstractVector{T})

Perform numerical integration by using the simpson rule. Note that the length of x and y must be odd numbers. And x must be a linear and uniform mesh.

See also: trapz.

Interpolations

Structs

ACFlow.AbstractInterpolation — Type

AbstractInterpolation

It represents an abstract interpolation engine, which is used to build the internal type system.

ACFlow.LinearInterpolation — Type

LinearInterpolation

It represents the linear interpolation algorithm.

ACFlow.QuadraticInterpolation — Type

QuadraticInterpolation

It represents the quadratic interpolation algorithm.

ACFlow.CubicSplineInterpolation — Type

CubicSplineInterpolation

It represents the cubic spline interpolation algorithm.

Constructors

ACFlow.LinearInterpolation — Method

LinearInterpolation(u::AbstractVector, t::AbstractVector)

Create a LinearInterpolation struct. Note that u and t denote f(x) and x, respectively.

ACFlow.QuadraticInterpolation — Method

QuadraticInterpolation(u::AbstractVector, t::AbstractVector)

Create a QuadraticInterpolation struct. Note that u and t denote f(x) and x, respectively.

ACFlow.CubicSplineInterpolation — Method

CubicSplineInterpolation(u::AbstractVector, t::AbstractVector)

Create a CubicSplineInterpolation struct. Note that u and t denote f(x) and x, respectively.

Functions

ACFlow.munge_data — Function

munge_data(u::AbstractVector{<:Real}, t::AbstractVector{<:Real})

Preprocess the input data. Note that u and t denote f(x) and x, respectively.

munge_data(u::AbstractVector, t::AbstractVector)

Preprocess the input data. Note that u and t denote f(x) and x, respectively.

ACFlow._interp — Function

_interp(A::LinearInterpolation{<:AbstractVector}, t::Number)

To implement the linear interpolation algorithm.

See also: LinearInterpolation.

_interp(A::QuadraticInterpolation{<:AbstractVector}, t::Number)

To implement the quadratic interpolation algorithm.

See also: QuadraticInterpolation.

_interp(A::CubicSplineInterpolation{<:AbstractVector{<:Number}}, t::Number)

To implement the cubic spline interpolation algorithm.

See also: CubicSplineInterpolation.

Einstein Summation Convention

ACFlow.@einsum — Macro

@einsum(ex)

Perform Einstein summation like operations on Julia Arrays.

Examples

Basic matrix multiplication can be implemented as:

@einsum A[i, j] := B[i, k] * C[k, j]

If the destination array is preallocated, then use =:

A = ones(5, 6, 7) # Preallocated space
X = randn(5, 2)
Y = randn(6, 2)
Z = randn(7, 2)

# Store the result in A, overwriting as necessary:
@einsum A[i, j, k] = X[i, r] * Y[j, r] * Z[k, r]

If destination is not preallocated, then use := to automatically create a new array for the result:

X = randn(5, 2)
Y = randn(6, 2)
Z = randn(7, 2)

# Create new array B with appropriate dimensions:
@einsum B[i, j, k] := X[i, r] * Y[j, r] * Z[k, r]

Curve Fitting

Structs

ACFlow.OnceDifferentiable — Type

OnceDifferentiable

Mutable struct. It is used for objectives and solvers where the gradient is available/exists.

Members

ℱ! -> Objective. It is actually a function call and return objective.

𝒥! -> It is a function call as well and returns jacobian of objective.

𝐹 -> Cache for ℱ! output.

𝐽 -> Cache for 𝒥! output.

ACFlow.OptimizationResults — Type

OptimizationResults{T,N}

It is used to save the optimization results of the levenberg_marquardt algorithm.

Members

x₀ -> Initial guess for the solution.

minimizer -> Final results for the solution.

minimum -> Residual.

iterations -> Number of iterations.

xconv -> If the convergence criterion 1 is satisfied.

gconv -> If the convergence criterion 2 is satisfied.

ACFlow.LsqFitResult — Type

LsqFitResult

It encapsulates the results for curve fitting.

Members

param -> Fitted results, i.e, the fitting parameters.

resid -> Residuals.

jacobian -> Jacobian matrix.

converged -> If the curve-fitting algorithm is converged.

Constructors

ACFlow.OnceDifferentiable — Method

OnceDifferentiable(𝑓, p0::AbstractArray, 𝐹::AbstractArray)

Constructor for OnceDifferentiable struct. 𝑓 is the function, p0 is the inital guess, 𝐹 = 𝑓(p0) is the cache for 𝑓's output.

Functions

ACFlow.value — Function

value(obj::OnceDifferentiable)

Return obj.𝐹. obj will not be affected.

value(obj::OnceDifferentiable, 𝐹, x)

Return 𝑓(x). obj will not be affected, but 𝐹 is updated.

ACFlow.value! — Function

value!(obj::OnceDifferentiable, x)

Return 𝑓(x). obj.𝐹 will be updated and returned.

ACFlow.jacobian — Function

jacobian(obj::OnceDifferentiable)

Return obj.𝐽. obj will not be affected.

ACFlow.jacobian! — Function

jacobian!(obj::OnceDifferentiable, x)

Return jacobian. obj.𝐽 will be updated and returned.

ACFlow.levenberg_marquardt — Function

levenberg_marquardt(df::OnceDifferentiable, x₀::AbstractVector{T})

Returns the argmin over x of sum(f(x).^2) using the Levenberg-Marquardt algorithm. The function f is encoded in df. x₀ is an initial guess for the solution.

See also: OnceDifferentiable.

ACFlow.curve_fit — Function

curve_fit(model, x, y, p0)

Fit data to a non-linear model. p0 is an initial model parameter guess. The return object is a composite type (LsqFitResult).

See also: LsqFitResult.