Solvers

Define various solvers for the ACFlow package.

Contents

Solvers

Contents

Index

Abstract Structs

MaxEnt Solver

Structs

Functions

StochAC Solver

Structs

Functions

StochSK Solver

Structs

Functions

StochOM Solver

Structs

Functions

Index

ACFlow.AbstractMC

ACFlow.AbstractSolver

ACFlow.Box

ACFlow.MaxEntContext

ACFlow.MaxEntSolver

ACFlow.StochACContext

ACFlow.StochACElement

ACFlow.StochACMC

ACFlow.StochACSolver

ACFlow.StochOMContext

ACFlow.StochOMElement

ACFlow.StochOMMC

ACFlow.StochOMSolver

ACFlow.StochSKContext

ACFlow.StochSKElement

ACFlow.StochSKMC

ACFlow.StochSKSolver

ACFlow.Pdx

ACFlow.average

ACFlow.average

ACFlow.average

ACFlow.bryan

ACFlow.calc_alpha

ACFlow.calc_bayes

ACFlow.calc_bayes_offdiag

ACFlow.calc_chi2

ACFlow.calc_correlator

ACFlow.calc_delta

ACFlow.calc_entropy

ACFlow.calc_entropy_offdiag

ACFlow.calc_error

ACFlow.calc_fmesh

ACFlow.calc_fmesh

ACFlow.calc_goodness

ACFlow.calc_green

ACFlow.calc_hamil

ACFlow.calc_htau

ACFlow.calc_lambda

ACFlow.calc_norm

ACFlow.calc_phi

ACFlow.calc_theta

ACFlow.calc_xmesh

ACFlow.chi2kink

ACFlow.classic

ACFlow.constraints

ACFlow.constraints

ACFlow.constraints

ACFlow.f_and_J

ACFlow.f_and_J_offdiag

ACFlow.historic

ACFlow.init

ACFlow.init

ACFlow.init

ACFlow.init

ACFlow.init_context

ACFlow.init_element

ACFlow.init_element

ACFlow.init_element

ACFlow.init_iodata

ACFlow.init_iodata

ACFlow.init_iodata

ACFlow.init_mc

ACFlow.init_mc

ACFlow.init_mc

ACFlow.last

ACFlow.last

ACFlow.last

ACFlow.last

ACFlow.measure

ACFlow.measure

ACFlow.optimizer

ACFlow.precompute

ACFlow.prun

ACFlow.prun

ACFlow.prun

ACFlow.run

ACFlow.run

ACFlow.run

ACFlow.run

ACFlow.sample

ACFlow.sample

ACFlow.shuffle

ACFlow.solve

ACFlow.solve

ACFlow.solve

ACFlow.solve

ACFlow.svd_to_real

ACFlow.svd_to_real_offdiag

ACFlow.try_height

ACFlow.try_insert

ACFlow.try_merge

ACFlow.try_mov1

ACFlow.try_mov2

ACFlow.try_move_p

ACFlow.try_move_q

ACFlow.try_move_s

ACFlow.try_remove

ACFlow.try_shift

ACFlow.try_split

ACFlow.try_swap

ACFlow.try_width

ACFlow.update

ACFlow.warmup

ACFlow.warmup

Abstract Structs

ACFlow.AbstractSolver — Type

AbstractSolver

An abstract type representing the solver for analytical continuation problem. It is used to build the internal type system. All the other solvers are its sub-types.

ACFlow.AbstractMC — Type

AbstractMC

An abstract type representing the Monte Carlo engine. It is used to build the internal type system.

MaxEnt Solver

Structs

ACFlow.MaxEntSolver — Type

MaxEntSolver

It represents the analytical continuation solver that implements the maximum entropy method.

ACFlow.MaxEntContext — Type

MaxEntContext

Mutable struct. It is used within the MaxEnt solver only.

Members

Gᵥ -> Input data for correlator.

σ² -> Actually 1.0 / σ².

grid -> Grid for input data.

mesh -> Mesh for output spectrum.

model -> Default model function.

kernel -> Default kernel function.

Vₛ -> Matrix from singular value decomposition.

W₂ -> Precomputed array.

W₃ -> Precomputed array.

Bₘ -> Precomputed array.

Functions

ACFlow.solve — Method

solve(S::MaxEntSolver, rd::RawData)

Solve the analytical continuation problem by the maximum entropy method.

ACFlow.init — Method

init(S::MaxEntSolver, rd::RawData)

Initialize the MaxEnt solver and return a MaxEntContext struct.

ACFlow.run — Method

run(mec::MaxEntContext)

Perform maximum entropy simulation with different algorithms. Now it supports the historic, classic, bryan, and chi2kink algorithms.

ACFlow.last — Method

last(mec::MaxEntContext, svec::Vector, sol::Dict)

Postprocess the results generated during the maximum entropy simulations. Here sol is the final solution for the analytical continuation problem, while svec contains all the intermediate results (it is a vector of dictionary actually).

ACFlow.historic — Function

historic(mec::MaxEntContext)

Apply the historic algorithm to solve the analytical continuation problem. It choose α in a way that χ² ≈ N.

See also: MaxEntContext.

ACFlow.classic — Function

classic(mec::MaxEntContext)

Apply the classic algorithm to solve the analytical continuation problem.

Classic algorithm uses Bayes statistics to approximately determine the most probable value of α. We always start at a large value of α, where the optimization yields basically the default model, therefore u_vec is only a few steps away from 0 (= default model). And then we gradually decrease α, step by step moving away from the default model towards data fitting. Using u_vec as start for the next (smaller) α brings a great speedup into this procedure.

See also: MaxEntContext.

ACFlow.bryan — Function

bryan(mec::MaxEntContext)

Apply the bryan algorithm to solve the analytical continuation problem.

Bryan's maxent calculates an average of spectral functions, weighted by their Bayesian probability.

See also: MaxEntContext.

ACFlow.chi2kink — Function

chi2kink(mec::MaxEntContext)

Apply the chi2kink algorithm to solve the analytical continuation problem.

We start with an optimization at a large value of α, where we should get only the default model. And then, α is decreased step-by-step, until the minimal value of α is reached. Then, we fit a function

ϕ(x; a, b, c, d) = a + b / [1 + exp(-d*(x-c))],

from which the optimal α is determined by

x_opt = c - fit_position / d,

and

alpha_opt = 10^x_opt.

See also: MaxEntContext.

ACFlow.optimizer — Function

optimizer(mec::MaxEntContext, α::F64, us::Vector{F64}, use_bayes::Bool)

Optimization of maxent functional for a given value of α. Since a priori the best value of α is unknown, this function has to be called several times in order to find a good value.

α means a weight factor of the entropy. us is a vector in singular space. It is used as a starting value for the optimization. For the very first optimization, done at large α, we use zeros, which corresponds to the default model. Then we use the result of the previous optimization as a starting value. use_bayes determines whether to use the Bayesian inference parameters for α.

This function will return a dictionary object that holds the results of the optimization, e.g. spectral function, χ² deviation.

ACFlow.precompute — Function

precompute(Gᵥ::Vector{F64}, σ²::Vector{F64},
           am::AbstractMesh,
           D::Vector{F64},
           K::Matrix{F64})

Precompute some key coefficients. Here Gᵥ and σ² are input data, am is the mesh for spectrum, D is the default model, and K is the kernel function.

ACFlow.f_and_J — Function

f_and_J(u::Vector{F64}, mec::MaxEntContext, α::F64)

This function evaluates the function whose root we want to find. Here u is a singular-space vector that parametrizes the spectral function, and α is a (positive) weight factor of the entropy.

It returns f, value of the function whose zero we want to find, and J, jacobian at the current position.

See also: f_and_J_offdiag.

ACFlow.f_and_J_offdiag — Function

f_and_J_offdiag(u::Vector{F64}, mec::MaxEntContext, α::F64)

This function evaluates the function whose root we want to find. Here u is a singular-space vector that parametrizes the spectral function, and α is a (positive) weight factor of the entropy.

It returns f, value of the function whose zero we want to find, and J, jacobian at the current position.

This function is similar to f_and_J, but for offdiagonal elements.

See also: f_and_J.

ACFlow.svd_to_real — Function

svd_to_real(mec::MaxEntContext, u::Vector{F64})

Go from singular value space to real space. It will transform the singular space vector u into real-frequency space (to get the spectral function) by A(ω) = D(ω) eⱽᵘ, where D(ω) is the default model V is the matrix from the singular value decomposition. The argument u means a singular space vector that parametrizes the spectral function.

See also: svd_to_real_offdiag.

ACFlow.svd_to_real_offdiag — Function

svd_to_real_offdiag(mec::MaxEntContext, u::Vector{F64})

Go from singular value space to real space. It will transform the singular space vector u into real-frequency space in the case of an offdiagonal element. It will return the spectral function.

See also: svd_to_real.

ACFlow.calc_entropy — Function

calc_entropy(mec::MaxEntContext, A::Vector{F64}, u::Vector{F64})

It computes entropy for positive definite spectral function. Here the arguments A means spectral function and u means a singular space vector that parametrizes the spectral function.

See also: calc_entropy_offdiag.

ACFlow.calc_entropy_offdiag — Function

calc_entropy_offdiag(mec::MaxEntContext, A::Vector{F64})

It compute positive-negative entropy for spectral function with norm 0. Here the argument A means spectral function.

See also: calc_entropy.

ACFlow.calc_bayes — Function

calc_bayes(mec::MaxEntContext,
           A::Vector{F64},
           S::F64, χ²::F64, α::F64)

It calculates Bayesian convergence criterion (ng, tr, and conv) for classic maxent (maximum of probablility distribution) and then Bayesian a-posteriori probability (log_prob) for α after optimization of A.

Here, A is the spectral function, S the entropy, χ² the deviation, and α weight factor of the entropy.

See also: calc_bayes_offdiag.

ACFlow.calc_bayes_offdiag — Function

calc_bayes_offdiag(mec::MaxEntContext,
                   A::Vector{F64},
                   S::F64, χ²::F64, α::F64)

It calculates Bayesian convergence criterion (ng, tr, and conv) for classic maxent (maximum of probablility distribution) and then Bayesian a-posteriori probability (log_prob) for α after optimization of A.

Here, A is the spectral function, S the entropy, χ² the deviation, and α weight factor of the entropy.

It is just a offdiagonal version of calc_bayes().

See also: calc_bayes.

ACFlow.calc_chi2 — Function

calc_chi2(mec::MaxEntContext, A::Vector{F64})

It computes the χ²-deviation of the spectral function A.

StochAC Solver

Structs

ACFlow.StochACSolver — Type

StochACSolver

It represents the analytical continuation solver that implements the stochastic analytical continuation method (K. S. D. Beach's version).

ACFlow.StochACMC — Type

StochACMC

Mutable struct. It is used within the StochAC solver. It includes random number generator and some counters.

Members

rng -> Random number generator.

Macc -> Counter for move operation (accepted).

Mtry -> Counter for move operation (tried).

Sacc -> Counter for swap operation (accepted).

Stry -> Counter for swap operation (tried).

See also: StochACSolver.

ACFlow.StochACElement — Type

StochACElement

Mutable struct. It is used to record the field configurations, which will be sampled by Monte Carlo sweeping procedure.

Members

Γₐ -> It means the positions of the δ functions.

Γᵣ -> It means the weights / amplitudes of the δ functions.

ACFlow.StochACContext — Type

StochACContext

Mutable struct. It is used within the StochAC solver only.

Members

Gᵥ -> Input data for correlator.

σ¹ -> Actually 1.0 / σ¹.

allow -> Allowable indices.

grid -> Grid for input data.

mesh -> Mesh for output spectrum.

model -> Default model function.

kernel -> Default kernel function.

Aout -> Calculated spectral function, it is actually ⟨n(x)⟩.

Δ -> Precomputed δ functions.

hτ -> α-resolved h(τ).

Hα -> α-resolved Hc.

Uα -> α-resolved internal energy, it is actually ⟨Hα⟩.

αₗ -> Vector of the α parameters.

Functions

ACFlow.solve — Method

solve(S::StochACSolver, rd::RawData)

Solve the analytical continuation problem by the stochastic analytical continuation algorithm (K. S. D. Beach's version).

ACFlow.init — Method

init(S::StochACSolver, rd::RawData)

Initialize the StochAC solver and return the StochACMC, StochACElement, and StochACContext structs.

ACFlow.run — Method

run(MC::StochACMC, SE::StochACElement, SC::StochACContext)

Perform stochastic analytical continuation simulation, sequential version.

ACFlow.prun — Method

prun(S::StochACSolver,
     p1::Dict{String,Vector{Any}},
     p2::Dict{String,Vector{Any}},
     MC::StochACMC, SE::StochACElement, SC::StochACContext)

Perform stochastic analytical continuation simulation, parallel version. The arguments p1 and p2 are copies of PBASE and PStochAC, respectively.

ACFlow.average — Method

average(step::F64, SC::StochACContext)

Postprocess the results generated during the stochastic analytical continuation simulations. It will calculate the spectral functions, and internal energies.

ACFlow.last — Method

last(SC::StochACContext, Aout::Array{F64,2}, Uα::Vector{F64})

It will process and write the calculated results by the StochAC solver, including effective hamiltonian, final spectral function, reproduced correlator.

ACFlow.warmup — Method

warmup(MC::StochACMC, SE::StochACElement, SC::StochACContext)

Warmup the Monte Carlo engine to acheieve thermalized equilibrium.

ACFlow.sample — Method

sample(MC::StochACMC, SE::StochACElement, SC::StochACContext)

Perform Monte Carlo sweeps and sample the field configurations.

ACFlow.measure — Method

measure(SE::StochACElement, SC::StochACContext)

Measure the spectral functions and internal energies.

ACFlow.init_mc — Method

init_mc(S::StochACSolver)

Try to create a StochACMC struct.

See also: StochACMC.

ACFlow.init_element — Method

init_element(S::StochACSolver, rng::AbstractRNG, allow::Vector{I64})

Randomize the configurations for future Monte Carlo sampling. It will return a StochACElement object.

See also: StochACElement.

ACFlow.init_iodata — Method

init_iodata(S::StochACSolver, rd::RawData)

Preprocess the input data (rd), then allocate memory for the α-resolved spectral functions.

See also: RawData.

ACFlow.calc_fmesh — Method

calc_fmesh(S::StochACSolver)

Try to calculate very fine (dense) linear mesh in [wmin, wmax], which is used internally to build the kernel function.

See also: LinearMesh.

ACFlow.calc_xmesh — Function

calc_xmesh()

Try to calculate very fine (dense) linear mesh in [0, 1], which is used internally to build the δ functions.

See also: calc_delta.

ACFlow.calc_phi — Function

calc_phi(am::AbstractMesh, model::Vector{F64})

Try to calculate ϕ(ω) function. am is the mesh for calculated spectrum, and model means the default model function.

See also: AbstractMesh, calc_delta.

ACFlow.calc_delta — Function

calc_delta(xmesh::Vector{F64}, ϕ::Vector{F64})

Precompute the δ functions. xmesh is a very dense linear mesh in [0, 1] and ϕ is the ϕ function.

See also: calc_xmesh, calc_phi.

ACFlow.calc_hamil — Function

calc_hamil(Γₐ, Γᵣ, kernel, Gᵥ, σ¹)

Initialize h(τ) and H(α) using Eq.(35) and Eq.(36), respectively. Γₐ and Γᵣ represent n(x), kernel means the kernel function, Gᵥ is the correlator. Note that kernel and Gᵥ have been rotated into singular space. Please see comments in init() for more details.

See also: calc_htau.

ACFlow.calc_htau — Function

calc_htau(Γₐ, Γᵣ, kernel, Gᵥ)

Try to calculate α-dependent h(τ) via Eq.(36). Γₐ and Γᵣ represent n(x), kernel means the kernel function, Gᵥ is the correlator. Note that kernel and Gᵥ have been rotated into singular space. Please see comments in init() for more details.

See also: calc_hamil.

ACFlow.calc_alpha — Function

calc_alpha()

Generate a list for the α parameters

ACFlow.constraints — Method

constraints(S::StochACSolver)

Try to implement the constrained stochastic analytical continuation method. This function will return a collection. It contains all the allowable indices.

See also: StochACSolver.

ACFlow.try_mov1 — Function

try_mov1(i::I64, MC::StochACMC, SE::StochACElement, SC::StochACContext)

Select two δ functions and then change their weights. Here i means the index for α parameters.

See also: try_mov2.

ACFlow.try_mov2 — Function

try_mov2(i::I64, MC::StochACMC, SE::StochACElement, SC::StochACContext)

Select two δ functions and then change their positions. Here i means the index for α parameters.

See also: try_mov1.

ACFlow.try_swap — Function

try_swap(MC::StochACMC, SE::StochACElement, SC::StochACContext)

Try to exchange field configurations between two adjacent layers.

StochSK Solver

Structs

ACFlow.StochSKSolver — Type

StochSKSolver

It represents the analytical continuation solver that implements the stochastic analytical continuation method (A. W. Sandvik's version).

ACFlow.StochSKMC — Type

StochSKMC

Mutable struct. It is used within the StochSK solver. It includes random number generator and some counters.

Members

rng -> Random number generator.

Sacc -> Counter for single-updated operation (accepted).

Stry -> Counter for single-updated operation (tried).

Pacc -> Counter for pair-updated operation (accepted).

Ptry -> Counter for pair-updated operation (tried).

Qacc -> Counter for quadruple-updated operation (accepted).

Qtry -> Counter for quadruple-updated operation (tried).

See also: StochSKSolver.

ACFlow.StochSKElement — Type

StochSKElement

Mutable struct. It is used to record the field configurations, which will be sampled by Monte Carlo sweeping procedure.

Members

P -> It means the positions of the δ functions.

A -> It means the weights / amplitudes of the δ functions.

W -> It denotes the window that is used to shift the δ functions.

ACFlow.StochSKContext — Type

StochSKContext

Mutable struct. It is used within the StochSK solver only.

Members

Gᵥ -> Input data for correlator.

Gᵧ -> Generated correlator.

σ¹ -> Actually 1.0 / σ¹.

allow -> Allowable indices.

grid -> Grid for input data.

mesh -> Mesh for output spectrum.

kernel -> Default kernel function.

Aout -> Calculated spectral function.

χ² -> Current goodness function.

χ²min -> Mininum goodness function.

χ²vec -> Vector of goodness function.

Θ -> Current Θ parameter.

Θvec -> Vector of Θ parameter.

Functions

ACFlow.solve — Method

solve(S::StochSKSolver, rd::RawData)

Solve the analytical continuation problem by the stochastic analytical continuation algorithm (A. W. Sandvik's version).

ACFlow.init — Method

init(S::StochSKSolver, rd::RawData)

Initialize the StochSK solver and return the StochSKMC, StochSKElement, and StochSKContext structs.

ACFlow.run — Method

run(MC::StochSKMC, SE::StochSKElement, SC::StochSKContext)

Perform stochastic analytical continuation simulation, sequential version.

ACFlow.prun — Method

prun(S::StochSKSolver,
     p1::Dict{String,Vector{Any}},
     p2::Dict{String,Vector{Any}},
     MC::StochSKMC, SE::StochSKElement, SC::StochSKContext)

Perform stochastic analytical continuation simulation, parallel version. The arguments p1 and p2 are copies of PBASE and PStochSK, respectively.

ACFlow.average — Method

average(step::F64, SC::StochSKContext)

Postprocess the results generated during the stochastic analytical continuation simulations. It will generate the spectral functions.

ACFlow.last — Method

last(SC::StochSKContext,
     Asum::Vector{F64},
     χ²vec::Vector{F64}, Θvec::Vector{F64})

It will process and write the calculated results by the StochSK solver, including final spectral function and reproduced correlator.

ACFlow.warmup — Method

warmup(MC::StochSKMC, SE::StochSKElement, SC::StochSKContext)

Warmup the Monte Carlo engine to acheieve thermalized equilibrium. Then it will try to figure out the optimized Θ and the corresponding Monte Carlo field configuration.

ACFlow.sample — Method

sample(MC::StochSKMC, SE::StochSKElement, SC::StochSKContext)

Perform Monte Carlo sweeps and sample the field configurations.

ACFlow.measure — Method

measure(SE::StochSKElement, SC::StochSKContext)

Measure the final spectral functions.

See also: nearest.

ACFlow.shuffle — Method

shuffle(MC::StochSKMC, SE::StochSKElement, SC::StochSKContext)

Try to shuffle the Monte Carlo field configuration via the Metropolis algorithm. Then the window for shifting the δ functions is adjusted.

ACFlow.init_mc — Method

init_mc(S::StochSKSolver)

Try to create a StochSKMC struct.

See also: StochSKMC.

ACFlow.init_element — Method

init_element(S::StochSKSolver, rng::AbstractRNG, allow::Vector{I64})

Randomize the configurations for future Monte Carlo sampling. It will return a StochSKElement object.

See also: StochSKElement.

ACFlow.init_iodata — Method

init_iodata(S::StochSKSolver, rd::RawData)

Preprocess the input data (rd), then allocate memory for the calculated spectral functions.

See also: RawData.

ACFlow.calc_fmesh — Method

calc_fmesh(S::StochSKSolver)

Try to calculate very fine (dense) linear mesh in [wmin, wmax], which is used internally to build the kernel function.

See also: LinearMesh.

ACFlow.calc_correlator — Function

calc_correlator(SE::StochSKElement, kernel::Array{F64,2})

Try to calculate correlator with the kernel function and the Monte Carlo field configuration. This correlator will then be used to evaluate the goodness function.

See also: calc_goodness.

ACFlow.calc_goodness — Function

calc_goodness(Gₙ::Vector{F64,}, Gᵥ::Vector{F64})

Try to calculate the goodness function (i.e, χ²), which measures the distance between input and regenerated correlators.

See also: calc_correlator.

ACFlow.calc_theta — Function

calc_theta(len::I64, SC::StochSKContext)

Try to locate the optimal Θ and χ². This function implements the chi2min and chi2kink algorithms. Note that the chi2min algorithm is preferred.

ACFlow.constraints — Method

constraints(S::StochSKSolver)

Try to implement the constrained stochastic analytical continuation method. This function will return a collection. It contains all the allowable indices.

See also: StochSKSolver.

ACFlow.try_move_s — Function

try_move_s(MC::StochSKMC, SE::StochSKElement, SC::StochSKContext)

Try to update the Monte Carlo field configurations via the Metropolis algorithm. In each update, only single δ function is shifted.

See also: try_move_p.

ACFlow.try_move_p — Function

try_move_p(MC::StochSKMC, SE::StochSKElement, SC::StochSKContext)

Try to update the Monte Carlo field configurations via the Metropolis algorithm. In each update, only a pair of δ functions are shifted.

See also: try_move_s.

ACFlow.try_move_q — Function

try_move_q(MC::StochSKMC, SE::StochSKElement, SC::StochSKContext)

Try to update the Monte Carlo field configurations via the Metropolis algorithm. In each update, four different δ functions are shifted.

See also: try_move_s.

StochOM Solver

Structs

ACFlow.StochOMSolver — Type

StochOMSolver

It represents the analytical continuation solver that implements the stochastic optimization method.

ACFlow.StochOMMC — Type

StochOMMC

Mutable struct. It is used within the StochOM solver. It includes random number generator and some counters.

Members

rng -> Random number generator.

Macc -> Counter for move operation (accepted).

Mtry -> Counter for move operation (tried).

See also: StochOMSolver.

ACFlow.Box — Type

Box

Rectangle. The field configuration consists of many boxes. They exhibit various areas (width × height). We use the Metropolis important sampling algorithm to sample them and evaluate their contributions to the spectrum.

Members

h -> Height of the box.

w -> Width of the box.

c -> Position of the box.

ACFlow.StochOMElement — Type

StochOMElement

Mutable struct. It is used to record the field configurations, which will be sampled by Monte Carlo sweeping procedure.

Members

C -> Field configuration.

Λ -> Contributions of the field configuration to the correlator.

G -> Reproduced correlator.

Δ -> Difference between reproduced and raw correlators.

ACFlow.StochOMContext — Type

StochOMContext

Mutable struct. It is used within the StochOM solver only.

Members

Gᵥ -> Input data for correlator.

σ¹ -> Actually 1.0 / σ¹.

grid -> Grid for input data.

mesh -> Mesh for output spectrum.

Cᵥ -> It is used to record the field configurations for all attempts.

Δᵥ -> It is used to record the errors for all attempts.

Functions

ACFlow.solve — Method

solve(S::StochOMSolver, rd::RawData)

Solve the analytical continuation problem by the stochastic optimization method.

ACFlow.init — Method

init(S::StochOMSolver, rd::RawData)

Initialize the StochOM solver and return the StochOMMC and StochOMContext structs.

ACFlow.run — Method

run(MC::StochOMMC, SC::StochOMContext)

Perform stochastic optimization simulation, sequential version.

ACFlow.prun — Method

prun(S::StochOMSolver,
     p1::Dict{String,Vector{Any}},
     p2::Dict{String,Vector{Any}},
     MC::StochOMMC, SC::StochOMContext)

Perform stochastic optimization simulation, parallel version. The arguments p1 and p2 are copies of PBASE and PStochOM, respectively.

ACFlow.average — Method

average(SC::StochOMContext)

Postprocess the collected results after the stochastic optimization simulations. It will generate the spectral functions.

ACFlow.last — Method

last(SC::StochOMContext, Aout::Vector{F64})

It will process and write the calculated results by the StochOM solver, including final spectral function and reproduced correlator.

ACFlow.update — Method

update(MC::StochOMMC, SE::StochOMElement, SC::StochOMContext)

Using the Metropolis algorithm to update the field configuration, i.e, a collection of hundreds of boxes.

ACFlow.init_mc — Method

init_mc(S::StochOMSolver)

Try to create a StochOMMC struct.

See also: StochOMMC.

ACFlow.init_element — Method

init_element(MC::StochOMMC, SC::StochOMContext)

Try to initialize a StochOMElement struct.

See also: StochOMElement.

ACFlow.init_iodata — Method

init_iodata(S::StochOMSolver, rd::RawData)

Preprocess the input data (rd).

See also: RawData, GreenData.

ACFlow.init_context — Method

init_context(S::StochOMSolver)

Try to initialize the key members of a StochOMContext struct.

See also: StochOMContext.

ACFlow.calc_lambda — Function

calc_lambda(r::Box, grid::FermionicMatsubaraGrid)

Try to calculate the contribution of a given box r to the Λ function. This function works for FermionicMatsubaraGrid only.

See also: FermionicMatsubaraGrid.

calc_lambda(r::Box, grid::FermionicImaginaryTimeGrid)

Try to calculate the contribution of a given box r to the Λ function. This function works for FermionicImaginaryTimeGrid only.

See also: FermionicImaginaryTimeGrid.

calc_lambda(r::Box, grid::BosonicMatsubaraGrid)

Try to calculate the contribution of a given box r to the Λ function. This function works for BosonicMatsubaraGrid only.

See also: BosonicMatsubaraGrid.

calc_lambda(r::Box, grid::BosonicImaginaryTimeGrid)

Try to calculate the contribution of a given box r to the Λ function. This function works for BosonicImaginaryTimeGrid only.

See also: BosonicImaginaryTimeGrid.

ACFlow.calc_error — Function

calc_error(G::Vector{F64}, Gᵥ::Vector{F64}, σ¹::Vector{F64})

Try to calculate χ². Here Gᵥ and σ¹ denote the raw correlator and related standard deviation. G means the reproduced correlator.

See also: calc_green.

ACFlow.calc_green — Function

calc_green(Λ::Array{F64,2}, nk::I64)

Try to reconstruct the correlator via the field configuration.

See also: calc_error.

ACFlow.calc_norm — Function

calc_norm(C::Vector{Box})

Calculate the total area of all boxes.

ACFlow.constraints — Method

constraints(e₁::F64, e₂::F64)

This function is used to judege whether a given box overlapes with the forbidden zone. Here e₁ and e₂ denote the left and right boundaries of the box.

ACFlow.try_insert — Function

try_insert(MC::StochOMMC,
           SE::StochOMElement,
           SC::StochOMContext,
           dacc::F64)

Insert a new box into the field configuration.

ACFlow.try_remove — Function

try_remove(MC::StochOMMC,
           SE::StochOMElement,
           SC::StochOMContext,
           dacc::F64)

Remove an old box from the field configuration.

ACFlow.try_shift — Function

try_shift(MC::StochOMMC,
          SE::StochOMElement,
          SC::StochOMContext,
          dacc::F64)

Change the position of given box in the field configuration.

ACFlow.try_width — Function

try_width(MC::StochOMMC,
          SE::StochOMElement,
          SC::StochOMContext,
          dacc::F64)

Change the width and height of given box in the field configuration. Note that the box's area is kept.

ACFlow.try_height — Function

try_height(MC::StochOMMC,
           SE::StochOMElement,
           SC::StochOMContext,
           dacc::F64)

Change the heights of two given boxes in the field configuration.

ACFlow.try_split — Function

try_split(MC::StochOMMC,
          SE::StochOMElement,
          SC::StochOMContext,
          dacc::F64)

Split a given box into two boxes in the field configuration.

ACFlow.try_merge — Function

try_merge(MC::StochOMMC,
          SE::StochOMElement,
          SC::StochOMContext,
          dacc::F64)

Merge two given boxes into one box in the field configuration.

ACFlow.Pdx — Function

Pdx(xmin::F64, xmax::F64, rng::AbstractRNG)

Try to calculate the probability density function.