Types

Define some fundamental types and structs for the ACFlow toolkit.

Contents

Types

Contents

Index

Data Types

Dictionaries

Index

ACFlow.PBASE

ACFlow.PBarRat

ACFlow.PMaxEnt

ACFlow.PNevanAC

ACFlow.PStochAC

ACFlow.PStochOM

ACFlow.PStochPX

ACFlow.PStochSK

ACFlow.ADT

ACFlow.AbstractData

ACFlow.DType

ACFlow.GreenData

ACFlow.RawData

Data Types

ACFlow.DType — Type

Customized types. It is used to define the following dicts.

ACFlow.ADT — Type

Customized types. It is used to define the following dicts.

ACFlow.AbstractData — Type

AbstractData

An abstract type representing the input data in imaginary axis. It is used to build the internal type system.

ACFlow.RawData — Type

RawData

Mutable struct. It represent the raw input data. The datatype T of raw data may be Float64 or ComplexF64.

Members

_grid -> Raw grid for the input data, such as τ or iωₙ.

value -> Raw input data, such as G(τ), G(iωₙ), or Σ(iωₙ).

error -> Error bar (standard deviation) of raw input data, σ.

See also: GreenData.

ACFlow.GreenData — Type

GreenData

Mutable struct. It represents the preprocessed input data. Note that it should support arbitrary precision via T

Members

value -> Preprocessed input data.

error -> Preprocessed error bar.

covar -> Diagonal parts of the covariance matrix, σ².

See also: RawData.

Dictionaries

ACFlow.PBASE — Constant

PBASE

Dictionary for configuration parameters: general setup.

ACFlow.PMaxEnt — Constant

PMaxEnt

Dictionary for configuration parameters: the maximum entropy method.

ACFlow.PBarRat — Constant

PBarRat

Dictionary for configuration parameters: the barycentric rational function approximation method.

ACFlow.PNevanAC — Constant

PNevanAC

Dictionary for configuration parameters: the Nevanlinna analytical continuation method.

ACFlow.PStochAC — Constant

PStochAC

Dictionary for configuration parameters: the stochastic analytic continuation method (K. S. D. Beach's version).

ACFlow.PStochSK — Constant

PStochSK

Dictionary for configuration parameters: the stochastic analytic continuation method (A. W. Sandvik's version).

ACFlow.PStochOM — Constant

PStochOM

Dictionary for configuration parameters: the stochastic optimization method.

ACFlow.PStochPX — Constant

PStochPX

Dictionary for configuration parameters: the stochastic pole expansion method.