Utilities

Provide some utility macros and functions for the ACFlow package.

Contents

Utilities

Contents

Index

Color Constants

Macros

Query Runtime Environment

Colorful Outputs

Input/output Operations

Color Tools

Index

ACFlow.COLORS

ACFlow.MODES

ACFlow.colorize

ACFlow.goodbye

ACFlow.line_to_array

ACFlow.overview

ACFlow.prompt

ACFlow.query_args

ACFlow.require

ACFlow.setup_args

ACFlow.sorry

ACFlow.welcome

ACFlow.@cswitch

ACFlow.@pcs

ACFlow.@time_call

Color Constants

ACFlow.COLORS — Constant

COLORS

A global dict, which is used to specify the system colors.

ACFlow.MODES — Constant

MODES

A global dict, which is used to specify the mode for output characters.

Macros

ACFlow.@cswitch — Macro

@cswitch(constexpr, body)

Provides a C-like switch statement with the falling through behavior. This implementation was borrowed from the following github repository:

https://github.com/Gnimuc/CSyntax.jl

Examples

engine = get_d("engine")
@cswitch engine begin
    @case "vasp"
        just_do_it()
        break

    @default
        sorry()
        break
end

ACFlow.@time_call — Macro

@time_call(ex)

Evaluate a function call (ex), and then print the elapsed time (number of seconds) it took to execute.

This macro is a variation of the standard @elapsed macro.

ACFlow.@pcs — Macro

@pcs(x...)

Try to print colorful strings. Here x is a combination of strings and colors. Its format likes string1 color1 string2 color2 (repeat). For the supported colors, please check the global dict COLORS.

Examples

julia> @pcs "Hello world!" blue
julia> @pcs "Hello " red "world!" green

See also: COLORS, welcome.

Query Runtime Environment

ACFlow.require — Function

require()

Check the version of julia runtime environment. It should be higher than v1.6.x. One of the most important philosophies of the ACFlow toolkit is minimizing the dependence on the third-party libraries as far as possible. Note that the ACFlow toolkit relys on the TOML package to parse the *.toml file. Only in v1.6.0 and higher versions, julia includes the TOML package in its standard library.

ACFlow.setup_args — Function

setup_args(x::Vararg{String})

Setup ARGS manually. This function is used only in REPL environment. We can use this function to update ARGS, so that the query_args() and the other related functions can work correctly.

Examples

julia> setup_args("ac.toml")
1-element Array{String,1}:
 "ac.toml"

See also: query_args.

ACFlow.query_args — Function

query_args()

Check whether the configuration file (case.toml) is provided.

See also: setup_args.

Colorful Outputs

ACFlow.welcome — Function

welcome()

Print out the welcome messages to the screen.

ACFlow.overview — Function

overview()

Print out the overview of ACFlow to the screen.

ACFlow.goodbye — Function

goodbye()

Print the goodbye messages to the screen.

ACFlow.sorry — Function

sorry()

Print an error message to the screen.

ACFlow.prompt — Function

prompt(msg::String)

Print a stylized ACFlow message to the screen.

Input/output Operations

ACFlow.line_to_array — Function

line_to_array(io::IOStream)

Convert a line (reading from an IOStream) to a string array.

line_to_array(str::AbstractString)

Convert a string (AbstractString) to a string array.

Color Tools

ACFlow.colorize — Function

colorize(c::String, s::String; bg::String = "default", m::String="default")

Return some escape sequences, which will be displayed as colorized texts in the terminal.

colorize(c::String, s::String; bg::String = "default", m::String="default")

Return some escape sequences, which will be displayed as colorized texts in the terminal.