ACTest.jl

Declare the module and public members.

Modules

ACTest — Module

ACTest

ACTest is an automatical generator for fermionic or bosonic correlation functions. It is used to generate a lot of correlators and corresponding spectral functions. They are then employed to test the newly developed analytic continuation methods or tools.

Now ACTest is used together with the ACFlow toolkit. But it is easily to be interfaced with the other analytic continuation codes.

For more details about how to obtain, install and use the ACTest toolkit, please visit the following website:

https://huangli712.github.io/projects/actest/index.html

Any suggestions, comments, and feedbacks are welcome. Enjoy it!

Initialization

ACTest._precompile — Function

_precompile()

Here, we would like to precompile the whole ACTest toolkit to reduce the runtime latency and speed up the successive calculations.

ACTest.__init__ — Function

__init__()

This function would be executed immediately after the module is loaded at runtime for the first time. It works at the REPL mode only.