Input and Output

Define input and output functions for the ACTest toolkit.

Contents

Input and Output

Contents

Index

Write Data

Index

ACTest.write_backward

ACTest.write_spectrum

Write Data

ACTest.write_spectrum — Function

write_spectrum(am::AbstractMesh, Aout::Vector{F64})

Write spectrum A(ω) to image.data. The grid is defined in am, and the spectral data are contained in Aout.

Arguments

am -> Real frequency mesh, ω.

Aout -> Spectral function, A(ω).

Returns

N/A

write_spectrum(ind::I64, sf::SpectralFunction)

Write spectrum A(ω) to image.data.i. All information about the spectral function is included in sf.

Arguments

ind -> Index for the spectral function.

sf -> A SpectralFunction struct.

Returns

N/A

ACTest.write_backward — Function

write_backward(ag::AbstractGrid, G::Vector{F64})

We can use the calculated spectrum in real axis to generate the Green's function data in imaginary axis. This function will write the data to green.data, which can be fed into the analytic continuation tools. Here, G is the constructed Green's function data.

Arguments

ag -> Grid for input data.

G -> Constructed Green's function.

Returns

N/A

See also: reprod.

write_backward(ind::I64, gf::GreenFunction)

Write the Green's function data to green.data.i. All information about the Green's function is included in gf.

Arguments

ind -> Index for the Green's function.

gf -> A GreenFunction struct.

Returns

N/A

See also: reprod.