Kernels

Build kernel functions.

The ACTest toolkit supports twelve types of kernel functions. They are:

Fermionic imaginary time kernel (ktype = "fermi", grid = "ftime")

Fermionic Matsubara kernel (ktype = "fermi", grid = "ffreq")

Bosonic imaginary time kernel (ktype = "boson", grid = "btime")

Bosonc Matsubara kernel (ktype = "boson", grid = "bfreq")

Symmetric bosonic imaginary time kernel (ktype = "bsymm", grid = "btime")

Symmetric bosonic Matsubara kernel (ktype = "bsymm", grid = "bfreq")

Contents

Kernels

Contents

Index

Making Kernels

Index

ACTest.build_kernel

ACTest.build_kernel_symm

Making Kernels

ACTest.build_kernel — Function

build_kernel(am::AbstractMesh, fg::FermionicImaginaryTimeGrid)

Try to build fermionic kernel function in imaginary time axis.

Arguments

am -> Real frequency mesh.

fg -> Imaginary time grid.

Returns

kernel -> Kernel function, K(τ,ω).

See also: AbstractMesh, FermionicImaginaryTimeGrid.

build_kernel(am::AbstractMesh, fg::FermionicMatsubaraGrid)

Try to build fermionic kernel function in Matsubara frequency axis.

Arguments

am -> Real frequency mesh.

fg -> Matsubara frequency grid.

Returns

kernel -> Kernel function, K(iωₙ,ω).

See also: AbstractMesh, FermionicMatsubaraGrid.

build_kernel(am::AbstractMesh, bg::BosonicImaginaryTimeGrid)

Try to build bosonic kernel function in imaginary time axis.

Arguments

am -> Real frequency mesh.

bg -> Imaginary time grid.

Returns

kernel -> Kernel function, K(τ,ω).

See also: AbstractMesh, BosonicImaginaryTimeGrid.

build_kernel(am::AbstractMesh, bg::BosonicMatsubaraGrid)

Try to build bosonic kernel function in Matsubara frequency axis.

Arguments

am -> Real frequency mesh.

bg -> Matsubara frequency grid.

Returns

kernel -> Kernel function, K(iωₙ,ω).

See also: AbstractMesh, BosonicMatsubaraGrid.

ACTest.build_kernel_symm — Function

build_kernel_symm(am::AbstractMesh, bg::BosonicImaginaryTimeGrid)

Try to build bosonic kernel function in imaginary time axis (just for correlator of Hermitian operator only).

Arguments

am -> Real frequency mesh.

bg -> Imaginary time grid.

Returns

kernel -> Kernel function, K(τ,ω).

See also: AbstractMesh, BosonicImaginaryTimeGrid.

build_kernel_symm(am::AbstractMesh, bg::BosonicMatsubaraGrid)

Try to build bosonic kernel function in Matsubara frequency axis (just for correlator of Hermitian operator only).

Arguments

am -> Real frequency mesh.

bg -> Matsubara frequency grid.

Returns

kernel -> Kernel function, K(iωₙ,ω).

See also: AbstractMesh, BosonicMatsubaraGrid.