Configuration

To extract, parse, verify, and print the configuration parameters. They are stored in external files (case.toml) or dictionaries.

Source: config.jl

Contents

Configuration

Contents

Index

Functions

Index

ZenCore._v

ZenCore.cat_c

ZenCore.cat_d

ZenCore.cat_i

ZenCore.cat_m

ZenCore.cat_s

ZenCore.chk_dict

ZenCore.exhibit

ZenCore.get_c

ZenCore.get_d

ZenCore.get_i

ZenCore.get_m

ZenCore.get_s

ZenCore.inp_toml

ZenCore.rev_dict

ZenCore.setup

ZenCore.str_c

ZenCore.str_d

ZenCore.str_i

ZenCore.str_m

ZenCore.str_s

Functions

ZenCore.setup — Function

setup()

Read parameters from configuration file, and then setup the related dicts.

See also: inp_toml, rev_dict, chk_dict.

ZenCore.inp_toml — Function

inp_toml(f::String, key::String, necessary::Bool)

Parse the configuration file (in toml format). It reads only parts of the configuration file, which depends on the value of key.

See also: setup.

inp_toml(f::String, necessary::Bool)

Parse the configuration file (in toml format). It reads the whole file.

See also: setup.

ZenCore.rev_dict — Function

rev_dict(cfg::Dict{String,Any})

Transfer configurations from dict cfg to internal dicts (including PCASE, PDFT, PDMFT, PIMP, and PSOLVER).

See also: chk_dict.

ZenCore.chk_dict — Function

chk_dict()

Validate the correctness and consistency of configurations.

See also: rev_dict, _v.

ZenCore.exhibit — Function

exhibit()

Display the configuration parameters for reference.

See also: setup.

ZenCore._v — Function

_v(key::String, val::Array{Any,1})

Verify the value array. Called by chk_dict() function only.

See also: chk_dict.

ZenCore.cat_c — Function

cat_c()

Print the configuration parameters to stdout: for PCASE dict.

See also: get_c, str_c.

ZenCore.cat_d — Function

cat_d()

Print the configuration parameters to stdout: for PDFT dict.

See also: get_d, str_d.

ZenCore.cat_m — Function

cat_m()

Print the configuration parameters to stdout: for PDMFT dict.

See also: get_m, str_m.

ZenCore.cat_i — Function

cat_i()

Print the configuration parameters to stdout: for PIMP dict.

See also: get_i, str_i.

ZenCore.cat_s — Function

cat_s()

Print the configuration parameters to stdout: for PSOLVER dict.

See also: get_s, str_s.

ZenCore.get_c — Function

get_c(key::String)

Extract configurations from dict: PCASE.

See also: cat_c, str_c.

ZenCore.get_d — Function

get_d(key::String)

Extract configurations from dict: PDFT.

See also: cat_d, str_d.

ZenCore.get_m — Function

get_m(key::String)

Extract configurations from dict: PDMFT.

See also: cat_m, str_m.

ZenCore.get_i — Function

get_i(key::String)

Extract configurations from dict: PIMP.

See also: cat_i, str_i.

ZenCore.get_s — Function

get_s(key::String)

Extract configurations from dict: PSOLVER.

See also: cat_s, str_s.

ZenCore.str_c — Function

str_c(key::String)

Extract configurations from dict: PCASE, convert them into strings.

See also: cat_c, get_c.

ZenCore.str_d — Function

str_d(key::String)

Extract configurations from dict: PDFT, convert them into strings.

See also: cat_d, get_d.

ZenCore.str_m — Function

str_m(key::String)

Extract configurations from dict: PDMFT, convert them into strings.

See also: cat_m, get_m.

ZenCore.str_i — Function

str_i(key::String)

Extract configurations from dict: PIMP, convert them into strings.

See also: cat_i, get_i.

ZenCore.str_s — Function

str_s(key::String)

Extract configurations from dict: PSOLVER, convert them into strings.

See also: cat_s, get_s.