Intermediate Representation

Tools for the intermediate representation format (adaptor).

Source: ir.jl

Contents

Intermediate Representation

Contents

Index

Functions

Index

ZenCore.ir_adaptor

ZenCore.ir_save

ZenCore.irio_charge

ZenCore.irio_eigen

ZenCore.irio_fermi

ZenCore.irio_groups

ZenCore.irio_kmesh

ZenCore.irio_lattice

ZenCore.irio_maps

ZenCore.irio_params

ZenCore.irio_projs

ZenCore.irio_tetra

ZenCore.irio_windows

Functions

ZenCore.ir_adaptor — Function

ir_adaptor(D::Dict{Symbol,Any})

Write the Kohn-Sham dataset to specified files using the IR format. Note that the Kohn-Sham dataset are encapsulated in the D dict.

See also: vasp_adaptor, plo_adaptor.

ZenCore.ir_save — Function

ir_save(it::IterInfo)

Backup the files outputed by the adaptor.

See also: ir_adaptor.

ZenCore.irio_params — Function

irio_params(f::String, D::Dict{Symbol,Any})

Write the key parameters extracted from the Kohn-Sham dataset. Here f means only the directory that we want to use.

See also: PrGroup, PrWindow.

ZenCore.irio_maps — Function

irio_maps(f::String, MAP::Mapping)

Write the information contained in Mapping struct. Here f means only the directory that we want to use.

See also: Mapping.

ZenCore.irio_groups — Function

irio_groups(f::String, PG::Array{PrGroup,1})

Write the information contained in PrGroup struct. Here f means only the directory that we want to use.

See also: PrGroup.

ZenCore.irio_windows — Function

irio_windows(f::String, PW::Array{PrWindow,1})

Write the information contained in PrWindow struct. Here f means only the directory that we want to use.

See also: PrWindow.

ZenCore.irio_lattice — Function

irio_lattice(f::String, latt::Lattice)

Write the lattice information to lattice.ir using the IR format. Here f means only the directory that we want to use.

See also: vaspio_lattice.

ZenCore.irio_kmesh — Function

irio_kmesh(f::String, kmesh::Array{F64,2}, weight::Array{F64,1})

Write the kmesh and weight information to kmesh.ir using the IR format. Here f means only the directory that we want to use.

See also: vaspio_kmesh.

ZenCore.irio_tetra — Function

irio_tetra(f::String, volt::F64, itet::Array{I64,2})

Write the tetrahedra information to tetra.ir using the IR format. Here f means only the directory that we want to use.

See also: vaspio_tetra.

ZenCore.irio_eigen — Function

irio_eigen(f::String, enk::Array{F64,3}, occupy::Array{F64,3})

Write the eigenvalues to eigen.ir using the IR format. Here f means only the directory that we want to use.

See also: vaspio_eigen.

ZenCore.irio_projs — Function

irio_projs(f::String, chipsi::Array{C64,4})

Write the projectors to projs.ir using the IR format. Here f means only the directory that we want to use.

The projectors are original data. They have not been modified.

See also: vaspio_projs.

irio_projs(f::String, chipsi::Array{Array{C64,4},1})

Write the projectors to projs.ir using the IR format. Here f means only the directory that we want to use.

The projectors have been processed to fulfill the requirement of the DMFT engine.

See also: vaspio_projs.

ZenCore.irio_fermi — Function

irio_fermi(f::String, fermi::F64)

Write the fermi level to fermi.ir using the IR format. Here f means only the directory that we want to use.

See also: vaspio_fermi.

ZenCore.irio_charge — Function

irio_charge(f::String)

Write the charge density to charge.ir using the IR format. Here f means only the directory that we want to use.

See also: vaspio_charge.