Analytical tetrahedron method

Implementation of the analytical tetrahedron method.

Source: tetra.jl

Contents

Analytical tetrahedron method

Contents

Index

Structs

Functions

Index

ZenCore.TetraWeight

ZenCore.bzint

ZenCore.fermi_weight

ZenCore.gauss_weight

ZenCore.tetra_p_ek1

ZenCore.tetra_p_ek12

ZenCore.tetra_p_ek23

ZenCore.tetra_p_ek34

ZenCore.tetra_p_ek4

ZenCore.tetra_weight

Structs

ZenCore.TetraWeight — Type

TetraWeight

Struct. Integration weights for analytical tetrahedron algorithm.

Members

cw -> Blochl corrections for dw.

dw -> Density of states weights at the four corners of a given tetrahedron.

tw -> Integration weights at the four corners of a given tetrahedron.

Functions

ZenCore.bzint — Function

bzint(z::F64, itet::Array{I64,2}, enk::Array{F64,3})

Compute tetrahedron integrated weights for Brillouin zone integration. It is used to calculate (partial) density of states.

See also: gauss_weight, fermi_weight, tetra_weight.

ZenCore.gauss_weight — Function

gauss_weight(z::F64, e::Array{F64,1})

Gaussian broadening algorithm for (integrated) density of states and relevant integration weights.

See also: TetraWeight.

ZenCore.fermi_weight — Function

fermi_weight(z::F64, e::Array{F64,1})

Fermi-Dirac broadening algorithm for (integrated) density of states and relevant integration weights.

See also: TetraWeight.

ZenCore.tetra_weight — Function

tetra_weight(z::F64, e::Array{F64,1})

Peter E. Blochl algorithm for (integrated) density of states and relevant integration weights. Blochl corrections are taken into considersions as well. See Phys. Rev. B, 49, 16223 (1994) for more details.

See also: TetraWeight.

ZenCore.tetra_p_ek1 — Function

tetra_p_ek1()

Blochl algorithm, case 1, for fully unoccupied tetrahedron.

See also: tetra_weight.

ZenCore.tetra_p_ek12 — Function

tetra_p_ek12(z::F64, e::Array{F64,1})

Blochl algorithm, case 2, for partially occupied tetrahedron.

See also: tetra_weight.

ZenCore.tetra_p_ek23 — Function

tetra_p_ek23(z::F64, e::Array{F64,1})

Blochl algorithm, case 3, for partially occupied tetrahedron.

See also: tetra_weight.

ZenCore.tetra_p_ek34 — Function

tetra_p_ek34(z::F64, e::Array{F64,1})

Blochl algorithm, case 4, for partially occupied tetrahedron.

See also: tetra_weight.

ZenCore.tetra_p_ek4 — Function

tetra_p_ek4()

Blochl algorithm, case 5, for fully occupied tetrahedron.

See also: tetra_weight.