Projected Local Orbitals

Tools for the projection on localized orbitals scheme (adaptor).

Source: plo.jl

Contents

Projected Local Orbitals

Contents

Index

Functions

Index

ZenCore.calc_dm

ZenCore.calc_dos

ZenCore.calc_hamk

ZenCore.calc_level

ZenCore.calc_ovlp

ZenCore.get_win1

ZenCore.get_win2

ZenCore.get_win3

ZenCore.plo_adaptor

ZenCore.plo_check

ZenCore.plo_fermi

ZenCore.plo_filter

ZenCore.plo_group

ZenCore.plo_map

ZenCore.plo_orthog

ZenCore.plo_rotate

ZenCore.plo_window

ZenCore.try_blk1

ZenCore.try_blk2

ZenCore.try_diag

ZenCore.view_dm

ZenCore.view_dos

ZenCore.view_hamk

ZenCore.view_level

ZenCore.view_ovlp

Functions

ZenCore.plo_adaptor — Function

plo_adaptor(D::Dict{Symbol,Any}, ai::Array{Impurity,1})

Adaptor support. It will preprocess the raw projector matrix. The dict D contains all of the necessary Kohn-Sham data, which will be modified in this function.

If it is in the REPL mode and the file case.test is present in the current directory, this function will try to calculate some physical observables, such as the density matrix, overlap matrix, partial density of states, and local hamiltonian, which will be written to external files or terminal for further reference.

See also: vasp_adaptor, ir_adaptor.

ZenCore.plo_check — Function

plo_check(D::Dict{Symbol,Any})

Generate some key physical quantities by using the projectors and the Kohn-Sham band structures. It is used for debug only.

ZenCore.plo_map — Function

plo_map(PG::Array{PrGroup,1}, ai::Array{Impurity,1})

Create connections / mappings between projectors (or band windows) and quantum impurity problems. Return a Mapping struct.

See also: PrGroup, PrWindow, Mapping.

ZenCore.plo_fermi — Function

plo_fermi(enk::Array{F64,3}, fermi::F64)

Calibrate the band structure to enforce the fermi level to be zero.

See also: vaspio_fermi, irio_fermi.

ZenCore.plo_group — Function

plo_group(MAP::Mapping, PG::Array{PrGroup,1})

Use the information contained in the Mapping struct to further complete the PrGroup struct.

See also: PIMP, Mapping, PrGroup.

ZenCore.plo_window — Function

plo_window(PG::Array{PrGroup,1}, enk::Array{F64,3})

Calibrate the band window to filter the Kohn-Sham eigenvalues.

See also: PrWindow, get_win1, get_win2.

ZenCore.plo_rotate — Function

plo_rotate(PG::Array{PrGroup,1}, chipsi::Array{C64,4})

Perform global rotations or transformations for the projectors. In this function, the projectors will be classified into different groups, and then they will be rotated group by group.

See also: PrGroup, plo_filter, plo_orthog.

ZenCore.plo_filter — Function

plo_filter(PW::Array{PrWindow,1}, chipsi::Array{Array{C64,4},1}}

Filter the projector matrix according to band window.

See also: PrWindow, plo_rotate, plo_orthog.

ZenCore.plo_orthog — Function

plo_orthog(PW::Array{PrWindow,1}, chipsi::Array{Array{C64,4},1})

Orthogonalize and normalize the projectors.

See also: PrWindow, plo_rotate, plo_filter.

ZenCore.get_win1 — Function

get_win1(enk::Array{F64,3}, bwin::Tuple{I64,I64})

Return momentum- and spin-dependent band window (case 1). The users provide only the band indices for the window.

See also: plo_window.

ZenCore.get_win2 — Function

get_win2(enk::Array{F64,3}, bwin::Tuple{F64,F64})

Return momentum- and spin-dependent band window (case 2). The users provide only the maximum and minimum energies for the window.

See also: plo_window.

ZenCore.get_win3 — Function

get_win3(chipsi::Array{C64,4})

Return momentum- and spin-dependent band window (case 3). This function implement a smart algorithm to determine the band window. It does not need any external parameters. This algorithm is quite simple. First, it uses the projection matrix (chipsi) to calculate the 𝑘- and σ-dependent projectibility. Then it calculates the effective projectibility by using 𝑘-summation. Next, it filters the projectibility to figure out which bands have large contributions. At last, the band window is composed of these selected bands. Note that the obtained band window is actually momentum- and spin-independent.

See also: plo_window.

ZenCore.try_blk1 — Function

try_blk1(PW::Array{PrWindow,1}, chipsi::Array{Array{C64,4},1})

Try to orthogonalize and normalize the projectors as a whole.

See also: PrWindow, try_blk2, plo_orthog.

ZenCore.try_blk2 — Function

try_blk2(PW::Array{PrWindow,1}, chipsi::Array{Array{C64,4},1})

Try to orthogonalize the projectors group by group.

See also: PrWindow, try_blk1, plo_orthog.

ZenCore.try_diag — Function

try_diag(M::AbstractArray{C64,2})

Orthogonalize the given matrix.

See also: try_blk1, try_blk2.

ZenCore.calc_ovlp — Function

calc_ovlp(chipsi::Array{C64,4}, weight::Array{F64,1})

Calculate the overlap matrix out of projectors. For raw projectors only.

See also: view_ovlp.

calc_ovlp(PW::Array{PrWindow,1},
          chipsi::Array{Array{C64,4},1},
          weight::Array{F64,1})

Calculate the overlap matrix out of projectors. For normalized projectors only.

See also: view_ovlp.

ZenCore.calc_dm — Function

calc_dm(chipsi::Array{C64,4},
        weight::Array{F64,1},
        occupy::Array{F64,3})

Calculate the density matrix out of projectors. For raw projectors only.

See also: view_dm.

calc_dm(PW::Array{PrWindow,1},
        chipsi::Array{Array{C64,4},1},
        weight::Array{F64,1},
        occupy::Array{F64,3})

Calculate the density matrix out of projectors. For normalized projectors only.

See also: view_dm.

ZenCore.calc_level — Function

calc_level(chipsi::Array{C64,4},
           weight::Array{F64,1},
           enk::Array{F64,3})

Try to build the effective band level in local basis. For raw projectors only.

See also: view_level.

calc_level(PW::Array{PrWindow,1},
           chipsi::Array{Array{C64,4},1},
           weight::Array{F64,1},
           enk::Array{F64,3})

Try to build the effective band level in local basis. For normalized projectors only.

See also: view_level.

ZenCore.calc_hamk — Function

calc_hamk(chipsi::Array{C64,4}, enk::Array{F64,3})

Try to build the hamiltonian matrix in local basis (momentum-dependent). For raw projectors only.

See also: view_hamk.

calc_hamk(PW::Array{PrWindow,1},
          chipsi::Array{Array{C64,4},1},
          enk::Array{F64,3})

Try to build the hamiltonian matrix in local basis (momentum-dependent). For normalized projectors only.

See also: view_hamk.

ZenCore.calc_dos — Function

calc_dos(PW::Array{PrWindow,1},
         chipsi::Array{Array{C64,4},1},
         itet::Array{I64,2},
         enk::Array{F64,3})

Try to calculate the partial density of states using the analytical tetrahedron method.

See also: view_dos.

ZenCore.view_ovlp — Function

view_ovlp(ovlp::Array{F64,3})

Output the overlap matrix to ovlp.raw. For raw projectors only.

See also: calc_ovlp.

view_ovlp(PG::Array{PrGroup,1}, ovlp::Array{Array{F64,3},1})

Output the overlap matrix to ovlp.nor. For normalized projectors only.

See also: calc_ovlp.

ZenCore.view_dm — Function

view_dm(dm::Array{F64,3})

Output the density matrix to dm.raw. For raw projectors only.

See also: calc_dm.

view_dm(PG::Array{PrGroup,1}, dm::Array{Array{F64,3},1})

Output the density matrix to dm.nor. For normalized projectors only.

See also: calc_dm.

ZenCore.view_level — Function

view_level(level::Array{C64,3})

Output the effective atomic level to level.raw. For raw projectors only.

See also: calc_level.

view_level(PG::Array{PrGroup,1}, level::Array{Array{C64,3},1})

Output the effective atomic level to level.nor. For normalized projectors only.

See also: calc_level.

ZenCore.view_hamk — Function

view_hamk(hamk::Array{C64,4})

Output the hamiltonian matrix in local basis to hamk.raw. For raw projectors only.

See also: calc_hamk.

view_hamk(hamk::Array{Array{C64,4},1})

Output the hamiltonian matrix in local basis to hamk.nor.i. For normalized projectors only.

See also: calc_hamk.

ZenCore.view_dos — Function

view_dos(mesh::Array{Array{F64,1},1}, dos::Array{Array{F64,3},1})

Output the density of states to dos.nor. For normalized projectors only.

See also: calc_dos.