Quantum Impurity Solvers

Wrapper for various quantum impurity solvers. Now only the CT-HYB₁, CT-HYB₂, HIA, and NORG quantum impurity solvers are supported.

Source: solver.jl

Contents

Quantum Impurity Solvers

Contents

Index

Functions

Index

ZenCore.CatImpurity

ZenCore.FixImpurity

ZenCore.GetEdmft

ZenCore.GetImpurity

ZenCore.GetNimpx

ZenCore.GetSigma

ZenCore.GetSymmetry

ZenCore.ctqmc_atomx

ZenCore.ctqmc_delta

ZenCore.ctqmc_edmft

ZenCore.ctqmc_eimpx

ZenCore.ctqmc_nimpx

ZenCore.ctqmc_setup

ZenCore.ctqmc_sigma

ZenCore.norg_delta

ZenCore.norg_edmft

ZenCore.norg_eimpx

ZenCore.norg_nimpx

ZenCore.norg_setup

ZenCore.norg_sigma

ZenCore.s_hub1_copy

ZenCore.s_hub1_exec

ZenCore.s_hub1_init

ZenCore.s_hub1_save

ZenCore.s_norg_copy

ZenCore.s_norg_exec

ZenCore.s_norg_init

ZenCore.s_norg_save

ZenCore.s_qmc1_copy

ZenCore.s_qmc1_exec

ZenCore.s_qmc1_init

ZenCore.s_qmc1_save

ZenCore.s_qmc2_copy

ZenCore.s_qmc2_exec

ZenCore.s_qmc2_init

ZenCore.s_qmc2_save

ZenCore.solver_call

ZenCore.solver_copy

ZenCore.solver_edmft

ZenCore.solver_nimpx

ZenCore.solver_sigma

Functions

ZenCore.solver_call — Function

solver_call(::NULLSolver, it::IterInfo, imp::Impurity)
solver_call(::CTHYB₁Solver, it::IterInfo, imp::Impurity)
solver_call(::CTHYB₂Solver, it::IterInfo, imp::Impurity)
solver_call(::HIASolver, it::IterInfo, imp::Impurity)
solver_call(::NORGSolver, it::IterInfo, imp::Impurity)
solver_call(::ATOMSolver, it::IterInfo, imp::Impurity)

Try to solve the quantum impurity problems by using various quantum impurity solvers. It acts as a dispatcher. Now it supports CTHYB₁ (ct_hyb1), CTHYB₂ (ct_hyb2), HIA (hia), NORG (norg), and ATOM (atomic) solvers.

See also: _solver_.

ZenCore.solver_copy — Function

solver_copy(::NULLSolver, it::IterInfo, imp₁::Impurity, imp₂::Impurity)
solver_copy(::CTHYB₁Solver, it::IterInfo, imp₁::Impurity, imp₂::Impurity)
solver_copy(::CTHYB₂Solver, it::IterInfo, imp₁::Impurity, imp₂::Impurity)
solver_copy(::HIASolver, it::IterInfo, imp₁::Impurity, imp₂::Impurity)
solver_copy(::NORGSolver, it::IterInfo, imp₁::Impurity, imp₂::Impurity)
solver_copy(::ATOMSolver, it::IterInfo, imp₁::Impurity, imp₂::Impurity)

Try to solve a quantum impurity problem by copying solution from another equivalent quantum impurity problem. It acts as a dispatcher.

See also: _solver_.

ZenCore.solver_sigma — Function

solver_sigma(::NULLSolver, imp::Impurity)
solver_sigma(::CTHYB₁Solver, imp::Impurity)
solver_sigma(::CTHYB₂Solver, imp::Impurity)
solver_sigma(::HIASolver, imp::Impurity)
solver_sigma(::NORGSolver, imp::Impurity)
solver_sigma(::ATOMSolver, imp::Impurity)

Try to extract self-energy function from the output data of quantum impurity solver. It acts as a dispatcher.

See also: _solver_.

ZenCore.solver_nimpx — Function

solver_nimpx(::NULLSolver, imp::Impurity)
solver_nimpx(::CTHYB₁Solver, imp::Impurity)
solver_nimpx(::CTHYB₂Solver, imp::Impurity)
solver_nimpx(::HIASolver, imp::Impurity)
solver_nimpx(::NORGSolver, imp::Impurity)
solver_nimpx(::ATOMSolver, imp::Impurity)

Try to extract impurity occupancy from the output data of quantum impurity solver. It acts as a dispatcher.

See also: _solver_.

ZenCore.solver_edmft — Function

solver_edmft(::NULLSolver)
solver_edmft(::CTHYB₁Solver)
solver_edmft(::CTHYB₂Solver)
solver_edmft(::HIASolver)
solver_edmft(::NORGSolver)
solver_edmft(::ATOMSolver)

Try to extract interaction energy from the output data of quantum impurity solver. It acts as a dispatcher.

See also: _solver_.

ZenCore.s_qmc1_init — Function

s_qmc1_init(it::IterInfo, imp::Impurity)

Check runtime environment of the CTHYB₁ quantum impurity solver. Prepare the necessary input files.

This quantum impurity solver is from the iQIST software package.

See also: s_qmc1_exec, s_qmc1_save.

ZenCore.s_qmc1_exec — Function

s_qmc1_exec(it::IterInfo)

Launch the CTHYB₁ quantum impurity solver.

This quantum impurity solver is from the iQIST software package.

See also: s_qmc1_init, s_qmc1_save.

ZenCore.s_qmc1_save — Function

s_qmc1_save(it::IterInfo, imp::Impurity)

Backup output files of the CTHYB₁ quantum impurity solver.

This quantum impurity solver is from the iQIST software package.

See also: s_qmc1_init, s_qmc1_exec.

ZenCore.s_qmc1_copy — Function

s_qmc1_copy(it::IterInfo, imp₁::Impurity, imp₂::Impurity)

Duplicate output files of the CTHYB₁ quantum impurity solver. We just copy selected output files from impurity.1 to impurity.2. Be careful, now we are already in directory impurity.2.

This quantum impurity solver is from the iQIST software package.

See also: s_qmc1_init, s_qmc1_exec.

ZenCore.s_qmc2_init — Function

s_qmc2_init(it::IterInfo)

Check runtime environment of the CTHYB₂ quantum impurity solver. Prepare the necessary input files.

This quantum impurity solver is from the iQIST software package.

See also: s_qmc2_exec, s_qmc2_save.

ZenCore.s_qmc2_exec — Function

s_qmc2_exec(it::IterInfo)

Launch the CTHYB₂ quantum impurity solver.

This quantum impurity solver is from the iQIST software package.

See also: s_qmc2_init, s_qmc2_save.

ZenCore.s_qmc2_save — Function

s_qmc2_save(it::IterInfo)

Backup output files of the CTHYB₂ quantum impurity solver.

This quantum impurity solver is from the iQIST software package.

See also: s_qmc2_init, s_qmc2_exec.

ZenCore.s_qmc2_copy — Function

s_qmc2_copy(it::IterInfo, imp₁::Impurity, imp₂::Impurity)

Duplicate output files of the CTHYB₂ quantum impurity solver. We just copy selected output files from impurity.1 to impurity.2. Be careful, now we are already in directory impurity.2.

This quantum impurity solver is from the iQIST software package.

See also: s_qmc2_init, s_qmc2_exec.

ZenCore.s_hub1_init — Function

s_hub1_init(it::IterInfo)

Check runtime environment of the HIA quantum impurity solver. Prepare the necessary input files.

See also: s_hub1_exec, s_hub1_save.

ZenCore.s_hub1_exec — Function

s_hub1_exec(it::IterInfo)

Launch the HIA quantum impurity solver.

See also: s_hub1_init, s_hub1_save.

ZenCore.s_hub1_save — Function

s_hub1_save(it::IterInfo)

Backup output files of the HIA quantum impurity solver.

See also: s_hub1_init, s_hub1_exec.

ZenCore.s_hub1_copy — Function

s_hub1_copy(it::IterInfo, imp₁::Impurity, imp₂::Impurity)

Duplicate output files of the HIA quantum impurity solver. We just copy selected output files from impurity.1 to impurity.2. Be careful, now we are already in directory impurity.2.

See also: s_hub1_init, s_hub1_exec.

ZenCore.s_norg_init — Function

s_norg_init(it::IterInfo, imp::Impurity)

Check runtime environment of the NORG quantum impurity solver. Prepare the necessary input files.

This quantum impurity solver is from the RUC Team.

See also: s_norg_exec, s_norg_save.

ZenCore.s_norg_exec — Function

s_norg_exec(it::IterInfo)

Launch the NORG quantum impurity solver.

This quantum impurity solver is from the RUC Team.

See also: s_norg_init, s_norg_save.

ZenCore.s_norg_save — Function

s_norg_save(it::IterInfo, imp::Impurity)

Backup output files of the NORG quantum impurity solver.

This quantum impurity solver is from the RUC Team.

See also: s_norg_init, s_norg_exec.

ZenCore.s_norg_copy — Function

s_norg_copy(it::IterInfo, imp₁::Impurity, imp₂::Impurity)

Duplicate output files of the NORG quantum impurity solver. We just copy selected output files from impurity.1 to impurity.2. Be careful, now we are already in directory impurity.2.

This quantum impurity solver is from the RUC Team.

See also: s_norg_init, s_norg_exec.

ZenCore.ctqmc_setup — Function

ctqmc_setup(imp::Impurity)

Generate default configuration file (solver.ctqmc.in) for the CTHYB quantum impurity solvers automatically (according to the information encoded in the Impurity struct).

See also: Impurity, ctqmc_atomx.

ZenCore.ctqmc_atomx — Function

ctqmc_atomx(imp::Impurity)

Generate configuration file for the atomic problem solver.

See also: Impurity, ctqmc_setup.

ZenCore.ctqmc_delta — Function

ctqmc_delta(fmesh::Array{F64,1}, Delta::Array{C64,4})

Write the hybridization functions to the solver.hyb.in file, which is suitable for the CTHYB quantum impurity solver.

See also: ctqmc_eimpx.

ZenCore.ctqmc_eimpx — Function

ctqmc_eimpx(Eimpx::Array{C64,3})

Write the local impurity levels to the solver.eimp.in file, which is suitable for the CTHYB quantum impurity solver.

See also: ctqmc_delta.

ZenCore.ctqmc_sigma — Function

ctqmc_sigma(imp::Impurity)

Parse the solver.sgm.dat file, which is generated by the CTHYB quantum impurity solver, to extract the bare self-energy functions.

In the sigma_gather() function, these data will be combined to generate the sigma.bare file, which is essential for the DMFT engine.

See also: GetSigma, solver_sigma.

ZenCore.ctqmc_nimpx — Function

ctqmc_nimpx(imp::Impurity)

Parse the solver.nmat.dat file, which is generated by the CTHYB quantum impurity solver, to extract the impurity occupancy. Then the fields nup, ndown, and occup in Impurity struct will be updated.

See also: GetNimpx, solver_nimpx.

ZenCore.ctqmc_edmft — Function

ctqmc_edmft()

Parse the solver.paux.dat file, which is generated by the CTHYB quantum impurity solver, to extract the interaction energy.

See also: GetEdmft, solver_edmft.

ZenCore.norg_setup — Function

norg_setup(imp::Impurity)

Generate default configuration file (solver.norg.in) for the NORG quantum impurity solvers automatically (according to the information encoded in the Impurity struct).

See also: Impurity, ctqmc_atomx.

ZenCore.norg_delta — Function

norg_delta(fmesh::Array{F64,1}, Delta::Array{C64,4})

Write the hybridization functions to the solver.hyb.in file, which is suitable for the NORG quantum impurity solver.

See also: norg_eimpx.

ZenCore.norg_eimpx — Function

norg_eimpx(Eimpx::Array{C64,3})

Write the local impurity levels to the solver.eimp.in file, which is suitable for the NORG quantum impurity solver.

See also: norg_delta.

ZenCore.norg_sigma — Function

norg_sigma(imp::Impurity)

Parse the seimp.txt file, which is generated by the NORG quantum impurity solver, to extract the bare self-energy functions.

In the sigma_gather() function, these data will be combined to generate the sigma.bare file, which is essential for the DMFT engine.

See also: GetSigma, solver_sigma.

ZenCore.norg_nimpx — Function

norg_nimpx(imp::Impurity)

Parse the nmat.txt file, which is generated by the NORG quantum impurity solver, to extract the impurity occupancy. Then the fields nup, ndown, and occup in Impurity struct will be updated.

See also: GetNimpx, solver_nimpx.

ZenCore.norg_edmft — Function

norg_edmft()

Parse the solver.out file, which is generated by the NORG quantum impurity solver, to extract the interaction energy.

See also: GetEdmft, solver_edmft.

ZenCore.GetSigma — Function

GetSigma(imp::Impurity)

Extract self-energy functions from the output files of various quantum impurity solvers. The data will be combined in the sigma_gather() function. Then they will be fed back to the DMFT engine. The working directory of this function must be the root folder.

See also: Impurity, solver_sigma.

ZenCore.GetNimpx — Function

GetNimpx(imp::Impurity)

Extract impurity occupancy from the output files of various quantum impurity solvers. Then the field occup in the Impurity struct will be updated, which will then be used to evaluate the double counting term for self-energy functions. The working directory of this function must be the root folder.

The argument imp may be modified in this function.

See also: Impurity, solver_nimpx.

ZenCore.GetEdmft — Function

GetEdmft(imp::Impurity)

Extract interaction energy (i.e potential energy) from the output files of various quantum impurity solvers. The input Impurity struct won't be modified. The working directory of this function must be the root folder.

See also: Impurity, solver_edmft.

ZenCore.GetSymmetry — Function

GetSymmetry(Eimpx::Array{C64,3})

Analyze the symmetry according to the diagonal elements of the matrix of the local impurity levels.

See also: ctqmc_eimpx.

ZenCore.GetImpurity — Function

GetImpurity()

Return an array of Impurity struct, which encapsulates useful information about the quantum impurity problems.

See also: Impurity.

ZenCore.CatImpurity — Function

CatImpurity(imp::Impurity)

Display the Impurity struct that need to be solved.

See also: Impurity.

ZenCore.FixImpurity — Function

FixImpurity(ai::Array{Impurity,1})

Update the quantum impurity problems encapsulated in ai according to the configuration parameters.

See also: Impurity.