Declare the module and public members.

Source: ZenCore.jl

ZenCore

ZenCore.__init__

ZenCore._precompile

Modules

ZenCore — Module

Zen and ZenCore

Zen is a modern DFT + DMFT computation framework. It can be used to study the correlated electronic structures of a wide range of strongly correlated materials. Now this framework is under heavy development. PLEASE USE IT AT YOUR OWN RISK.

Zen supports the following DFT backends:

VASP

QUANTUM ESPRESSO (Actually only the PWSCF code)

Zen supports the following schemes for defining local orbitals:

PLO

WANNIER (Need support from the wannier90 code)

Zen supports the following quantum impurity solvers:

CTHYB

HIA (Unavailable)

NORG (Experimental)

Zen consists of several components, including:

ZenCore (Core library)

ZenApps (Major applications)

ZenTools (Auxiliary tools and plugins)

Dyson (A density functional theory + dynamical mean-field theory engine)

iQIST (An interacting quantum impurity solver toolkit)

Flink (A reusable fortran numerical library)

ACFlow (A modern toolkit for analytical continuation problems)

ZenCore implements the core library of the Zen DFT + DMFT computation framework. It connects various components of Zen, and drive them to work together. It provides an easy-to-use, flexible, efficient, and robust user interface (UI) and application programming interface (API).

For more details about how to obtain, install and use the Zen framework and the ZenCore library, please visit the following website:

https://huangli712.github.io/projects/zen/index.html

Any suggestions, comments, and feedbacks are welcome. Enjoy it!

Functions

ZenCore._precompile — Function

_precompile()

Here, we would like to precompile the whole ZenCore package to reduce the runtime latency and speed up the successive calculations.

ZenCore.__init__ — Function

__init__()

This function would be executed immediately after the module is loaded at runtime for the first time. It works at the REPL mode only.