Library

Contents

Library

Contents

Index

Modules

Numerical Types

String Constants

Utilities

Widgets: Menu

Widgets: Window

Index

ZenGui

ZenGui.__AUTHORS__

ZenGui.__LIBNAME__

ZenGui.__RELEASE__

ZenGui.__VERSION__

ZenGui.APC

ZenGui.APF

ZenGui.API

ZenGui.APN

ZenGui.APR

ZenGui.C32

ZenGui.C64

ZenGui.F32

ZenGui.F64

ZenGui.I32

ZenGui.I64

ZenGui.N32

ZenGui.N64

ZenGui.R32

ZenGui.R64

ZenGui.authors

ZenGui.create_app_about

ZenGui.create_menu

ZenGui.set_menu_edit

ZenGui.set_menu_file

ZenGui.set_menu_help

ZenGui.set_menu_style

Modules

ZenGui — Module

ZenGui

ZenGui is a general-purpose graphic user interface for ab initio dynamical mean-field theory codes. It can be used to prepare necessary configuration files for them. Now it supports the following codes:

All-in-one DFT + DMFT package (Zen)

Quantum impurity solver toolkit (iQIST)

Analytic continuation tools (ACFlow and ACTest)

Dynamical mean-field theory engines (Dyson and DFermion)

This code is under heavy development. PLEASE USE IT AT YOUR OWN RISK.

For more details about how to obtain, install and use the ZenGui code, please visit the following website:

https://huangli712.github.io/projects/zengui/index.html

Any suggestions, comments, and feedbacks are welcome. Enjoy it!

Numerical Types

ZenGui.I32 — Type

I32

Alias of Integer type (32 bit).

See also: R32, N32.

ZenGui.I64 — Type

I64

Alias of Integer type (64 bit).

See also: R64, N64.

ZenGui.API — Type

API

Alias of Integer type (Arbitrary Precision Integer).

See also: APF, APC.

ZenGui.F32 — Type

F32

Alias of Float type (32 bit).

See also: R32, N32.

ZenGui.F64 — Type

F64

Alias of Float type (64 bit).

See also: R64, N64.

ZenGui.APF — Type

APF

Alias of Float type (Arbitrary Precision Float).

See also: API, APC.

ZenGui.C32 — Type

C32

Alias of Complex type (32 bit).

See also: R32, N32.

ZenGui.C64 — Type

C64

Alias of Complex type (64 bit).

See also: R64, N64.

ZenGui.APC — Type

APC

Alias of Complex type (Arbitrary Precision Complex).

See also: API, APF.

ZenGui.R32 — Type

R32

Alias of Integer and Float types (32 bit). Here R means Real.

See also: N32, N64, APN.

ZenGui.R64 — Type

R64

Alias of Integer and Float types (64 bit). Here R means Real.

See also: N32, N64, APN.

ZenGui.APR — Type

APR

Alias of Integer and Float types (Arbitrary Precision). Here R means Real.

See also: N32, N64, APN.

ZenGui.N32 — Type

N32

Alias of Integer, Float, and Complex types (32 bit). Here N means Number.

See also: R32, R64, APR.

ZenGui.N64 — Type

N64

Alias of Integer, Float, and Complex types (64 bit). Here N means Number.

See also: R32, R64, APR.

ZenGui.APN — Type

APN

Alias of Integer, Float, and Complex types (Arbitrary Precision). Here N means Number.

See also: R32, R64, APR.

String Constants

ZenGui.__LIBNAME__ — Constant

__LIBNAME__

Name of this julia toolkit.

See also: __VERSION__.

ZenGui.__VERSION__ — Constant

__VERSION__

Version of this julia toolkit.

See also: __RELEASE__.

ZenGui.__RELEASE__ — Constant

__RELEASE__

Release date of this julia toolkit.

See also: __AUTHORS__.

ZenGui.__AUTHORS__ — Constant

__AUTHORS__

Core authors of this julia toolkit.

See also: __LIBNAME__.

Utilities

ZenGui.authors — Function

authors()

Print authors / contributors of the ZenGui toolkit.

See also: __AUTHORS__.

Widgets: Menu

ZenGui.create_menu — Function

create_menu()

Generate menu in the main window.

ZenGui.set_menu_file — Function

set_menu_file()

Setup items in menu ``File''.

ZenGui.set_menu_edit — Function

setmenuedit()

Setup items in menu ``Edit''.

ZenGui.set_menu_style — Function

setmenustyle()

Setup items in menu ``Style''.

ZenGui.set_menu_help — Function

setmenuhelp()

Setup items in menu ``Help''.

Widgets: Window

ZenGui.create_app_about — Function

create_app_about(p_open::Ref{Bool})

Display the About window, which is used to show some userful information for users.