ACFlow.jl

Declare the module and public members.

Modules

ACFlow — Module

ACFlow

ACFlow is a modern software toolkit for solving the many-body analytical continuation problem. It is usually used to convert the single-particle or two-particle correlators from imaginary axis to real axis. Now this toolkit is under heavy development. PLEASE USE IT AT YOUR OWN RISK.

Nowadays the ACFlow toolkit supports the following algorithms:

Maximum Entropy Method (MaxEnt solver)

Stochastic Analytical Continuation (StochAC and StochSK solvers)

Stochastic Optimization Method (StochOM solver)

Stochastic Pole eXpansion (StochPX solver, experimental)

More algorithms will be implemented in the future.

Note that ACFlow toolkit has been designed to be integrated into the Zen package. Actually, it is also compatible with the iQIST toolkit. In the other words, the end user can use it to analytically continue the imaginary time (or Matsubara frequency) data generated by the various quantum impurity solvers in the iQIST toolkit. Of course, it is quite easy to implement some kinds of interfaces for the other quantum impurity solvers.

For more details about how to obtain, install and use the ACFlow toolkit, please visit the following website:

https://huangli712.github.io/projects/acflow/index.html

Any suggestions, comments, and feedbacks are welcome. Enjoy it!

Initialization

ACFlow._precompile — Function

_precompile()

Here, we would like to precompile the whole ACFlow toolkit to reduce the runtime latency and speed up the successive calculations.

ACFlow.__init__ — Function

__init__()

This function would be executed immediately after the module is loaded at runtime for the first time. It works at the REPL mode only.