Grids

Grids on imaginary axis.

In the ACFlow toolkit, the input correlators are defined on grid, while the calculated spectral functions are defined on mesh. The ACFlow toolkit supports both imaginary time and Matsubara frequency grids. Note that for Matsubara frequency grid, the bosonic and fermionic grids are different.

Contents

Grids

Contents

Index

Types

Constructors

Base.* Functions

Utilities

Index

ACFlow.AbstractGrid

ACFlow.BosonicFragmentMatsubaraGrid

ACFlow.BosonicFragmentMatsubaraGrid

ACFlow.BosonicFragmentTimeGrid

ACFlow.BosonicFragmentTimeGrid

ACFlow.BosonicImaginaryTimeGrid

ACFlow.BosonicImaginaryTimeGrid

ACFlow.BosonicMatsubaraGrid

ACFlow.BosonicMatsubaraGrid

ACFlow.FermionicFragmentMatsubaraGrid

ACFlow.FermionicFragmentMatsubaraGrid

ACFlow.FermionicFragmentTimeGrid

ACFlow.FermionicFragmentTimeGrid

ACFlow.FermionicImaginaryTimeGrid

ACFlow.FermionicImaginaryTimeGrid

ACFlow.FermionicMatsubaraGrid

ACFlow.FermionicMatsubaraGrid

ACFlow.rebuild!

Base.eachindex

Base.eachindex

Base.eachindex

Base.eachindex

Base.eachindex

Base.eachindex

Base.eachindex

Base.eachindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.length

Base.length

Base.length

Base.length

Base.length

Base.length

Base.length

Base.length

Base.resize!

Base.reverse!

Types

ACFlow.AbstractGrid — Type

AbstractGrid

An abstract type representing the imaginary axis. It is used to build the internal type system.

ACFlow.FermionicImaginaryTimeGrid — Type

FermionicImaginaryTimeGrid

Mutable struct. It represents the fermionic imaginary time grid.

Members

ntime -> Number of time slices.

β -> Inverse temperature.

τ -> Vector of grid points， τᵢ.

See also: FermionicFragmentTimeGrid.

ACFlow.FermionicFragmentTimeGrid — Type

FermionicFragmentTimeGrid

Mutable struct. It represents part of the fermionic imaginary time grid. In other words, the grid might be fragmentary。

Members

ntime -> Number of time slices.

β -> Inverse temperature.

τ -> Vector of grid points, τᵢ.

See also: FermionicImaginaryTimeGrid.

ACFlow.FermionicMatsubaraGrid — Type

FermionicMatsubaraGrid

Mutable struct. It represents the fermionic Matsubara frequency grid.

Members

nfreq -> Number of Matsubara frequency points.

β -> Inverse temperature.

ω -> Vector of grid points, iωₙ.

See also: FermionicFragmentMatsubaraGrid.

ACFlow.FermionicFragmentMatsubaraGrid — Type

FermionicFragmentMatsubaraGrid

Mutable struct. It represents part of the fermionic Matsubara frequency grid. In other words, the grid might be fragmentary。

Members

nfreq -> Number of Matsubara frequency points.

β -> Inverse temperature.

ω -> Vector of grid points, iωₙ.

See also: FermionicMatsubaraGrid.

ACFlow.BosonicImaginaryTimeGrid — Type

BosonicImaginaryTimeGrid

Mutable struct. It represents the bosonic imaginary time grid.

Members

ntime -> Number of time slices.

β -> Inverse temperature.

τ -> Vector of grid points, τᵢ.

See also: BosonicFragmentTimeGrid.

ACFlow.BosonicFragmentTimeGrid — Type

BosonicFragmentTimeGrid

Mutable struct. It represents part of the bosonic imaginary time grid. In other words, the grid might be fragmentary。

Members

ntime -> Number of time slices.

β -> Inverse temperature.

τ -> Vector of grid points, τᵢ.

See also: BosonicImaginaryTimeGrid.

ACFlow.BosonicMatsubaraGrid — Type

BosonicMatsubaraGrid

Mutable struct. It represents the bosonic Matsubara frequency grid.

Members

nfreq -> Number of Matsubara frequency points.

β -> Inverse temperature.

ω -> Vector of grid points, iωₙ.

See also: BosonicFragmentMatsubaraGrid.

ACFlow.BosonicFragmentMatsubaraGrid — Type

BosonicFragmentMatsubaraGrid

Mutable struct. It represents part of the bosonic Matsubara frequency grid. In other words, the grid might be fragmentary。 However, the first frequency point should be present (ωₙ ≡ 0.0).

Members

nfreq -> Number of Matsubara frequency points.

β -> Inverse temperature.

ω -> Vector of grid points, iωₙ.

See also: BosonicMatsubaraGrid.

Constructors

ACFlow.FermionicImaginaryTimeGrid — Method

FermionicImaginaryTimeGrid(ntime::I64, β::T) where {T}

A constructor for the FermionicImaginaryTimeGrid struct, which is defined in src/types.jl.

Arguments

ntime -> Number of time slices in imaginary axis.

β -> Inverse temperature.

Returns

grid -> A FermionicImaginaryTimeGrid struct.

See also: FermionicImaginaryTimeGrid.

ACFlow.FermionicFragmentTimeGrid — Method

FermionicFragmentTimeGrid(β::T, τ::Vector{T}) where {T}

A constructor for the FermionicFragmentTimeGrid struct, which is defined in src/types.jl.

Arguments

β -> Inverse temperature.

τ -> Given imaginary time points.

Returns

grid -> A FermionicFragmentTimeGrid struct.

See also: FermionicFragmentTimeGrid.

ACFlow.FermionicMatsubaraGrid — Method

FermionicMatsubaraGrid(nfreq::I64, β::T) where {T}

A constructor for the FermionicMatsubaraGrid struct, which is defined in src/types.jl. The Matsubara grid is evaluated as ωₙ = (2n - 1) π / β.

Arguments

nfreq -> Number of Matsubara frequencies.

β -> Inverse temperature.

Returns

grid -> A FermionicMatsubaraGrid struct.

See also: FermionicMatsubaraGrid.

ACFlow.FermionicFragmentMatsubaraGrid — Method

FermionicFragmentMatsubaraGrid(β::T, ω::Vector{T}) where {T}

A constructor for the FermionicFragmentMatsubaraGrid struct, which is defined in src/types.jl. The Matsubara grid is from input.

Arguments

β -> Inverse temperature.

ω -> Given Matsubara frequency points.

Returns

grid -> A FermionicFragmentMatsubaraGrid struct.

See also: FermionicFragmentMatsubaraGrid.

ACFlow.BosonicImaginaryTimeGrid — Method

BosonicImaginaryTimeGrid(ntime::I64, β::T)

A constructor for the BosonicImaginaryTimeGrid struct, which is defined in src/types.jl.

Arguments

ntime -> Number of time slices in imaginary axis.

β -> Inverse temperature.

Returns

grid -> A BosonicImaginaryTimeGrid struct.

See also: BosonicImaginaryTimeGrid.

ACFlow.BosonicFragmentTimeGrid — Method

BosonicFragmentTimeGrid(β::T, τ::Vector{T}) where {T}

A constructor for the BosonicFragmentTimeGrid struct, which is defined in src/types.jl.

Arguments

β -> Inverse temperature.

τ -> Given imaginary time points.

Returns

grid -> A BosonicFragmentTimeGrid struct.

See also: BosonicFragmentTimeGrid.

ACFlow.BosonicMatsubaraGrid — Method

BosonicMatsubaraGrid(nfreq::I64, β::T) where {T}

A constructor for the BosonicMatsubaraGrid struct, which is defined in src/types.jl. The Matsubara grid is evaluated as ωₙ = (2n - 2) π / β.

Arguments

nfreq -> Number of Matsubara frequencies.

β -> Inverse temperature.

Returns

grid -> A BosonicMatsubaraGrid struct.

See also: BosonicMatsubaraGrid.

ACFlow.BosonicFragmentMatsubaraGrid — Method

BosonicFragmentMatsubaraGrid(β::T, ω::Vector{T}) where {T}

A constructor for the BosonicFragmentMatsubaraGrid struct, which is defined in src/types.jl. The Matsubara grid is from input.

Arguments

β -> Inverse temperature.

ω -> Given Matsubara frequency points.

Returns

grid -> A BosonicFragmentMatsubaraGrid struct.

See also: BosonicFragmentMatsubaraGrid.

Base.* Functions

Base.length — Method

Base.length(fg::FermionicImaginaryTimeGrid)

Return number of grid points in a FermionicImaginaryTimeGrid struct.

See also: FermionicImaginaryTimeGrid.

Base.length — Method

Base.length(fg::FermionicFragmentTimeGrid)

Return number of grid points in a FermionicFragmentTimeGrid struct.

See also: FermionicFragmentTimeGrid.

Base.length — Method

Base.length(fg::FermionicMatsubaraGrid)

Return number of grid points in a FermionicMatsubaraGrid struct.

See also: FermionicMatsubaraGrid.

Base.length — Method

Base.length(fg::FermionicFragmentMatsubaraGrid)

Return number of grid points in a FermionicFragmentMatsubaraGrid struct.

See also: FermionicFragmentMatsubaraGrid.

Base.length — Method

Base.length(bg::BosonicImaginaryTimeGrid)

Return number of grid points in a BosonicImaginaryTimeGrid struct.

See also: BosonicImaginaryTimeGrid.

Base.length — Method

Base.length(bg::BosonicFragmentTimeGrid)

Return number of grid points in a BosonicFragmentTimeGrid struct.

See also: BosonicFragmentTimeGrid.

Base.length — Method

Base.length(bg::BosonicMatsubaraGrid)

Return number of grid points in a BosonicMatsubaraGrid struct.

See also: BosonicMatsubaraGrid.

Base.length — Method

Base.length(bg::BosonicFragmentMatsubaraGrid)

Return number of grid points in a BosonicFragmentMatsubaraGrid struct.

See also: BosonicFragmentMatsubaraGrid.

Base.iterate — Method

Base.iterate(fg::FermionicImaginaryTimeGrid)

Advance the iterator of a FermionicImaginaryTimeGrid struct to obtain the next grid point.

See also: FermionicImaginaryTimeGrid.

Base.iterate — Method

Base.iterate(fg::FermionicFragmentTimeGrid)

Advance the iterator of a FermionicFragmentTimeGrid struct to obtain the next grid point.

See also: FermionicFragmentTimeGrid.

Base.iterate — Method

Base.iterate(fg::FermionicMatsubaraGrid)

Advance the iterator of a FermionicMatsubaraGrid struct to obtain the next grid point.

See also: FermionicMatsubaraGrid.

Base.iterate — Method

Base.iterate(fg::FermionicFragmentMatsubaraGrid)

Advance the iterator of a FermionicFragmentMatsubaraGrid struct to obtain the next grid point.

See also: FermionicFragmentMatsubaraGrid.

Base.iterate — Method

Base.iterate(bg::BosonicImaginaryTimeGrid)

Advance the iterator of a BosonicImaginaryTimeGrid struct to obtain the next grid point.

See also: BosonicImaginaryTimeGrid.

Base.iterate — Method

Base.iterate(bg::BosonicFragmentTimeGrid)

Advance the iterator of a BosonicFragmentTimeGrid struct to obtain the next grid point.

See also: BosonicFragmentTimeGrid.

Base.iterate — Method

Base.iterate(bg::BosonicMatsubaraGrid)

Advance the iterator of a BosonicMatsubaraGrid struct to obtain the next grid point.

See also: BosonicMatsubaraGrid.

Base.iterate — Method

Base.iterate(bg::BosonicFragmentMatsubaraGrid)

Advance the iterator of a BosonicFragmentMatsubaraGrid struct to obtain the next grid point.

See also: BosonicFragmentMatsubaraGrid.

Base.iterate — Method

Base.iterate(fg::FermionicImaginaryTimeGrid, i::I64)

This is the key method that allows a FermionicImaginaryTimeGrid struct to be iterated, yielding a sequences of grid points.

See also: FermionicImaginaryTimeGrid.

Base.iterate — Method

Base.iterate(fg::FermionicFragmentTimeGrid, i::I64)

This is the key method that allows a FermionicFragmentTimeGrid struct to be iterated, yielding a sequences of grid points.

See also: FermionicFragmentTimeGrid.

Base.iterate — Method

Base.iterate(fg::FermionicMatsubaraGrid, i::I64)

Create an iterable object for visiting each index of a FermionicMatsubaraGrid struct.

See also: FermionicMatsubaraGrid.

Base.iterate — Method

Base.iterate(fg::FermionicFragmentMatsubaraGrid, i::I64)

Create an iterable object for visiting each index of a FermionicFragmentMatsubaraGrid struct.

See also: FermionicFragmentMatsubaraGrid.

Base.iterate — Method

Base.iterate(bg::BosonicImaginaryTimeGrid, i::I64)

Create an iterable object for visiting each index of a BosonicImaginaryTimeGrid struct.

See also: BosonicImaginaryTimeGrid.

Base.iterate — Method

Base.iterate(bg::BosonicFragmentTimeGrid, i::I64)

Create an iterable object for visiting each index of a BosonicFragmentTimeGrid struct.

See also: BosonicFragmentTimeGrid.

Base.iterate — Method

Base.iterate(bg::BosonicMatsubaraGrid, i::I64)

Create an iterable object for visiting each index of a BosonicMatsubaraGrid struct.

See also: BosonicMatsubaraGrid.

Base.iterate — Method

Base.iterate(bg::BosonicFragmentMatsubaraGrid, i::I64)

Create an iterable object for visiting each index of a BosonicFragmentMatsubaraGrid struct.

See also: BosonicFragmentMatsubaraGrid.

Base.eachindex — Method

Base.eachindex(fg::FermionicImaginaryTimeGrid)

Create an iterable object for visiting each index of a FermionicImaginaryTimeGrid struct.

See also: FermionicImaginaryTimeGrid.

Base.eachindex — Method

Base.eachindex(fg::FermionicFragmentTimeGrid)

Create an iterable object for visiting each index of a FermionicFragmentTimeGrid struct.

See also: FermionicFragmentTimeGrid.

Base.eachindex — Method

Base.eachindex(fg::FermionicMatsubaraGrid)

Create an iterable object for visiting each index of a FermionicMatsubaraGrid struct.

See also: FermionicMatsubaraGrid.

Base.eachindex — Method

Base.eachindex(fg::FermionicFragmentMatsubaraGrid)

Create an iterable object for visiting each index of a FermionicFragmentMatsubaraGrid struct.

See also: FermionicFragmentMatsubaraGrid.

Base.eachindex — Method

Base.eachindex(bg::BosonicImaginaryTimeGrid)

Create an iterable object for visiting each index of a BosonicImaginaryTimeGrid struct.

See also: BosonicImaginaryTimeGrid.

Base.eachindex — Method

Base.eachindex(bg::BosonicFragmentTimeGrid)

Create an iterable object for visiting each index of a BosonicFragmentTimeGrid struct.

See also: BosonicFragmentTimeGrid.

Base.eachindex — Method

Base.eachindex(bg::BosonicMatsubaraGrid)

Create an iterable object for visiting each index of a BosonicMatsubaraGrid struct.

See also: BosonicMatsubaraGrid.

Base.eachindex — Method

Base.eachindex(bg::BosonicFragmentMatsubaraGrid)

Create an iterable object for visiting each index of a BosonicFragmentMatsubaraGrid struct.

See also: BosonicFragmentMatsubaraGrid.

Base.firstindex — Method

Base.firstindex(fg::FermionicImaginaryTimeGrid)

Return the first index of a FermionicImaginaryTimeGrid struct.

See also: FermionicImaginaryTimeGrid.

Base.firstindex — Method

Base.firstindex(fg::FermionicFragmentTimeGrid)

Return the first index of a FermionicFragmentTimeGrid struct.

See also: FermionicFragmentTimeGrid.

Base.firstindex — Method

Base.firstindex(fg::FermionicMatsubaraGrid)

Return the first index of a FermionicMatsubaraGrid struct.

See also: FermionicMatsubaraGrid.

Base.firstindex — Method

Base.firstindex(fg::FermionicFragmentMatsubaraGrid)

Return the first index of a FermionicFragmentMatsubaraGrid struct.

See also: FermionicFragmentMatsubaraGrid.

Base.firstindex — Method

Base.firstindex(bg::BosonicImaginaryTimeGrid)

Return the first index of a BosonicImaginaryTimeGrid struct.

See also: BosonicImaginaryTimeGrid.

Base.firstindex — Method

Base.firstindex(bg::BosonicFragmentTimeGrid)

Return the first index of a BosonicFragmentTimeGrid struct.

See also: BosonicFragmentTimeGrid.

Base.firstindex — Method

Base.firstindex(bg::BosonicMatsubaraGrid)

Return the first index of a BosonicMatsubaraGrid struct.

See also: BosonicMatsubaraGrid.

Base.firstindex — Method

Base.firstindex(bg::BosonicFragmentMatsubaraGrid)

Return the first index of a BosonicFragmentMatsubaraGrid struct.

See also: BosonicFragmentMatsubaraGrid.

Base.lastindex — Method

Base.lastindex(fg::FermionicImaginaryTimeGrid)

Return the last index of a FermionicImaginaryTimeGrid struct.

See also: FermionicImaginaryTimeGrid.

Base.lastindex — Method

Base.lastindex(fg::FermionicFragmentTimeGrid)

Return the last index of a FermionicFragmentTimeGrid struct.

See also: FermionicFragmentTimeGrid.

Base.lastindex — Method

Base.lastindex(fg::FermionicMatsubaraGrid)

Return the last index of a FermionicMatsubaraGrid struct.

See also: FermionicMatsubaraGrid.

Base.lastindex — Method

Base.lastindex(fg::FermionicFragmentMatsubaraGrid)

Return the last index of a FermionicFragmentMatsubaraGrid struct.

See also: FermionicFragmentMatsubaraGrid.

Base.lastindex — Method

Base.lastindex(bg::BosonicImaginaryTimeGrid)

Return the last index of a BosonicImaginaryTimeGrid struct.

See also: BosonicImaginaryTimeGrid.

Base.lastindex — Method

Base.lastindex(bg::BosonicFragmentTimeGrid)

Return the last index of a BosonicFragmentTimeGrid struct.

See also: BosonicFragmentTimeGrid.

Base.lastindex — Method

Base.lastindex(bg::BosonicMatsubaraGrid)

Return the last index of a BosonicMatsubaraGrid struct.

See also: BosonicMatsubaraGrid.

Base.lastindex — Method

Base.lastindex(bg::BosonicFragmentMatsubaraGrid)

Return the last index of a BosonicFragmentMatsubaraGrid struct.

See also: BosonicFragmentMatsubaraGrid.

Base.getindex — Method

Base.getindex(fg::FermionicImaginaryTimeGrid, ind::I64)

Retrieve the value(s) stored at the given key or index within a FermionicImaginaryTimeGrid struct.

See also: FermionicImaginaryTimeGrid.

Base.getindex — Method

Base.getindex(fg::FermionicFragmentTimeGrid, ind::I64)

Retrieve the value(s) stored at the given key or index within a FermionicFragmentTimeGrid struct.

See also: FermionicFragmentTimeGrid.

Base.getindex — Method

Base.getindex(fg::FermionicMatsubaraGrid, ind::I64)

Retrieve the value(s) stored at the given key or index within a FermionicMatsubaraGrid struct.

See also: FermionicMatsubaraGrid.

Base.getindex — Method

Base.getindex(fg::FermionicFragmentMatsubaraGrid, ind::I64)

Retrieve the value(s) stored at the given key or index within a FermionicFragmentMatsubaraGrid struct.

See also: FermionicFragmentMatsubaraGrid.

Base.getindex — Method

Base.getindex(bg::BosonicImaginaryTimeGrid, ind::I64)

Retrieve the value(s) stored at the given key or index within a BosonicImaginaryTimeGrid struct.

See also: BosonicImaginaryTimeGrid.

Base.getindex — Method

Base.getindex(bg::BosonicFragmentTimeGrid, ind::I64)

Retrieve the value(s) stored at the given key or index within a BosonicFragmentTimeGrid struct.

See also: BosonicFragmentTimeGrid.

Base.getindex — Method

Base.getindex(bg::BosonicMatsubaraGrid, ind::I64)

Retrieve the value(s) stored at the given key or index within a BosonicMatsubaraGrid struct.

See also: BosonicMatsubaraGrid.

Base.getindex — Method

Base.getindex(bg::BosonicFragmentMatsubaraGrid, ind::I64)

Retrieve the value(s) stored at the given key or index within a BosonicFragmentMatsubaraGrid struct.

See also: BosonicFragmentMatsubaraGrid.

Base.getindex — Method

Base.getindex(fg::FermionicImaginaryTimeGrid, I::UnitRange{I64})

Return a subset of a FermionicImaginaryTimeGrid struct as specified by I.

See also: FermionicImaginaryTimeGrid.

Base.getindex — Method

Base.getindex(fg::FermionicFragmentTimeGrid, I::UnitRange{I64})

Return a subset of a FermionicFragmentTimeGrid struct as specified by I.

See also: FermionicFragmentTimeGrid.

Base.getindex — Method

Base.getindex(fg::FermionicMatsubaraGrid, I::UnitRange{I64})

Return a subset of a FermionicMatsubaraGrid struct as specified by I.

See also: FermionicMatsubaraGrid.

Base.getindex — Method

Base.getindex(fg::FermionicFragmentMatsubaraGrid, I::UnitRange{I64})

Return a subset of a FermionicFragmentMatsubaraGrid struct as specified by I.

See also: FermionicFragmentMatsubaraGrid.

Base.getindex — Method

Base.getindex(bg::BosonicImaginaryTimeGrid, I::UnitRange{I64})

Return a subset of a BosonicImaginaryTimeGrid struct as specified by I.

See also: BosonicImaginaryTimeGrid.

Base.getindex — Method

Base.getindex(bg::BosonicFragmentTimeGrid, I::UnitRange{I64})

Return a subset of a BosonicFragmentTimeGrid struct as specified by I.

See also: BosonicFragmentTimeGrid.

Base.getindex — Method

Base.getindex(bg::BosonicMatsubaraGrid, I::UnitRange{I64})

Return a subset of a BosonicMatsubaraGrid struct as specified by I.

See also: BosonicMatsubaraGrid.

Base.getindex — Method

Base.getindex(bg::BosonicFragmentMatsubaraGrid, I::UnitRange{I64})

Return a subset of a BosonicFragmentMatsubaraGrid struct as specified by I.

See also: BosonicFragmentMatsubaraGrid.

Utilities

ACFlow.rebuild! — Function

rebuild!(fg::FermionicImaginaryTimeGrid, ntime::I64, β::T) where {T}

Rebuild the FermionicImaginaryTimeGrid struct via new ntime and β parameters.

Arguments

fg -> A FermionicImaginaryTimeGrid struct.

ntime -> Number of time slices.

β -> Inverse temperature.

Returns

N/A

See also: FermionicImaginaryTimeGrid.

rebuild!(fg::FermionicFragmentTimeGrid, ntime::I64, β::T) where {T}

Rebuild the FermionicFragmentTimeGrid struct via new ntime and β parameters. Now its imaginary time points are continuous and complete.

Arguments

fg -> A FermionicFragmentTimeGrid struct.

ntime -> Number of time slices.

β -> Inverse temperature.

Returns

N/A

See also: FermionicFragmentTimeGrid.

rebuild!(fg::FermionicMatsubaraGrid, nfreq::I64, β::T) where {T}

Rebuild the FermionicMatsubaraGrid struct via new nfreq and β parameters.

Arguments

fg -> A FermionicMatsubaraGrid struct.

nfreq -> Number of Matsubara frequencies.

β -> Inverse temperature.

Returns

N/A

See also: FermionicMatsubaraGrid.

rebuild!(fg::FermionicFragmentMatsubaraGrid, nfreq::I64, β::T) where {T}

Rebuild the FermionicFragmentMatsubaraGrid struct via new nfreq and β parameters. Now its Matsubara frequency points are continuous and complete.

Arguments

fg -> A FermionicFragmentMatsubaraGrid struct.

nfreq -> Number of Matsubara frequencies.

β -> Inverse temperature.

Returns

N/A

See also: FermionicFragmentMatsubaraGrid.

rebuild!(bg::BosonicImaginaryTimeGrid, ntime::I64, β::T) where {T}

Rebuild the BosonicImaginaryTimeGrid struct via new ntime and β parameters.

Arguments

bg -> A BosonicImaginaryTimeGrid struct.

ntime -> Number of time slices.

β -> Inverse temperature.

Returns

N/A

See also: BosonicImaginaryTimeGrid.

rebuild!(bg::BosonicFragmentTimeGrid, ntime::I64, β::T) where {T}

Rebuild the BosonicFragmentTimeGrid struct via new ntime and β parameters. Now its imaginary time points are continuous and complete.

Arguments

bg -> A BosonicFragmentTimeGrid struct.

ntime -> Number of time slices.

β -> Inverse temperature.

Returns

N/A

See also: BosonicFragmentTimeGrid.

rebuild!(bg::BosonicMatsubaraGrid, nfreq::I64, β::T) where {T}

Rebuild the BosonicMatsubaraGrid struct via new nfreq and β parameters.

Arguments

bg -> A BosonicMatsubaraGrid struct.

nfreq -> Number of Matsubara frequencies.

β -> Inverse temperature.

Returns

N/A

See also: BosonicMatsubaraGrid.

rebuild!(bg::BosonicFragmentMatsubaraGrid, nfreq::I64, β::T) where {T}

Rebuild the BosonicFragmentMatsubaraGrid struct via new nfreq and β parameters. Now its Matsubara frequency points are continuous and complete.

Arguments

bg -> A BosonicFragmentMatsubaraGrid struct.

nfreq -> Number of Matsubara frequencies.

β -> Inverse temperature.

Returns

N/A

See also: BosonicFragmentMatsubaraGrid.

Base.resize! — Function

Base.resize!(fg::FermionicMatsubaraGrid, nfreq::I64)

Reduce the size of the fermionic Matsubara grid. Note that nfreq should be smaller than or equal to fg.nfreq. This function is called by the NevanAC solver only.

Arguments

fg -> A FermionicMatsubaraGrid struct.

nfreq -> Number of Matsubara frequencies.

Returns

N/A

See also: FermionicMatsubaraGrid.

Base.resize!(fg::FermionicFragmentMatsubaraGrid, nfreq::I64)

Reduce the size of the fermionic fragment Matsubara grid. Note that nfreq should be smaller than or equal to fg.nfreq. This function is called by the NevanAC solver only.

Arguments

fg -> A FermionicFragmentMatsubaraGrid struct.

nfreq -> Number of Matsubara frequencies.

Returns

N/A

See also: FermionicFragmentMatsubaraGrid.

Base.resize!(bg::BosonicMatsubaraGrid, nfreq::I64)

Reduce the size of the bosonic Matsubara grid. Note that nfreq should be smaller than or equal to bg.nfreq. This function is called by the NevanAC solver only.

Arguments

bg -> A BosonicMatsubaraGrid struct.

nfreq -> Number of Matsubara frequencies.

Returns

N/A

See also: BosonicMatsubaraGrid.

Base.resize!(bg::BosonicFragmentMatsubaraGrid, nfreq::I64)

Reduce the size of the bosonic fragment Matsubara grid. Note that nfreq should be smaller than or equal to bg.nfreq. This function is called by the NevanAC solver only.

Arguments

bg -> A BosonicFragmentMatsubaraGrid struct.

nfreq -> Number of Matsubara frequencies.

Returns

N/A

See also: BosonicFragmentMatsubaraGrid.

Base.reverse! — Function

Base.reverse!(fg::FermionicMatsubaraGrid)

Reverse the fermionic Matsubara grid. This function is called by the NevanAC solver only.

See also: FermionicMatsubaraGrid.

Base.reverse!(fg::FermionicFragmentMatsubaraGrid)

Reverse the fermionic fragment Matsubara grid. This function is called by the NevanAC solver only.

See also: FermionicFragmentMatsubaraGrid.

Base.reverse!(bg::BosonicMatsubaraGrid)

Reverse the bosonic Matsubara grid. This function is called by the NevanAC solver only.

See also: BosonicMatsubaraGrid.

Base.reverse!(bg::BosonicFragmentMatsubaraGrid)

Reverse the bosonic fragment Matsubara grid. This function is called by the NevanAC solver only.

See also: BosonicFragmentMatsubaraGrid.