Input and output

Define input and output functions for the ACFlow toolkit.

Contents

Input and output

Contents

Index

Read Data

Write Data

Index

ACFlow.read_cmplx_data

ACFlow.read_real_data

ACFlow.write_backward

ACFlow.write_complete

ACFlow.write_goodness

ACFlow.write_hamiltonian

ACFlow.write_misfit

ACFlow.write_model

ACFlow.write_passed

ACFlow.write_pole

ACFlow.write_probability

ACFlow.write_spectrum

ACFlow.write_statistics

Read Data

ACFlow.read_real_data — Function

read_real_data(finput::AbstractString, ngrid::I64)

Read input data. This function is used for imaginary time data. The input file should contain three columns. The first column is the imaginary time grid, the second column is the value, the third column is the standard deviation σ. Here, ngrid specifies the number of grid points.

See also: read_cmplx_data.

ACFlow.read_cmplx_data — Function

read_cmplx_data(finput::AbstractString, ngrid::I64)

Read input data. This function is used for Matsubara frequency data. The input should contain four columns or five columns. The first column is the Matsubara freqency grid, the second and third columns are the values (real part and imaginary part), the four and fifth columns are the standard deviations σ for the real and imaginary parts, respectively. If there are only four columns, it means that the real and imaginary parts share the same standard deviations.

See also: read_real_data.

read_cmplx_data(finput::AbstractString, ngrid::I64, only_real_part::Bool)

Read input data. This function is used for Matsubara frequency data. The input file only contains three columns. The first column is the Matsubara frequency grid, the second column is the real part or imaginary part of the data (which is specified by the argument only_real_part), and the third column is the standard deviation σ. This function is for bosonic correlation function.

See also: read_real_data.

Write Data

ACFlow.write_spectrum — Function

write_spectrum(am::AbstractMesh, Aout::Vector{F64})

Write spectrum A(ω) to Aout.data. The grid is defined in am, and the spectral data are contained in Aout.

write_spectrum(am::AbstractMesh, αₗ::Vector{F64}, Aout::Array{F64,2})

Write α-resolved spectrum A(ω) to Aout.data.alpha. The grid is defined in am, the α-resolved spectrum is contained in Aout, αₗ is the list for the α parameters. This function is called by the StochAC solver.

ACFlow.write_backward — Function

write_backward(ag::AbstractGrid, G::Vector{F64})

We can use the calculated spectrum in real axis to reproduce the input data in imaginary axis. This function will write the reproduced data to repr.data, which can be compared with the original data. Here, G is the reproduced data.

See also: reprod.

ACFlow.write_complete — Function

write_complete(am::AbstractMesh, G::Vector{C64})

Write the full data at real axis to Gout.data. am denotes the real axis, G is the calculated green's function data. Note that its real part is obtained via the so-called Kramers-Kronig transformation.

See also: kramers.

ACFlow.write_misfit — Function

write_misfit(α_vec::Vector{F64}, χ²_vec::Vector{F64})

Write log10(α)-log10(χ²) data to chi2.data, which could be used to judge whether the obtained optimal α parameter is reasonable.

See also: write_goodness.

ACFlow.write_goodness — Function

write_goodness(Θ_vec::Vector{F64}, χ²_vec::Vector{F64})

Write log10(Θ)-log10(χ²) data to goodness.data, which could be used to judge whether the obtained optimal Θ parameter is reasonable. This function is only useful for the StochSK solver.

See also: write_misfit.

ACFlow.write_model — Function

write_model(am::AbstractMesh, D::Vector{F64})

Write the default model function to model.data.

ACFlow.write_hamiltonian — Function

write_hamiltonian(α_vec::Vector{F64}, Uα::Vector{F64})

Write α-U(α) data to hamil.data, which could be used to judge whether the obtained optimal α parameter is reasonable. This function is only useful for the StochAC solver.

ACFlow.write_passed — Function

write_passed(passed::Vector{I64}, med::F64, αgood::F64)

Write indices of selected solutions which should be used to calculate the averaged spectrum. Here, passed means the indices, med is the median value of χ², and αgood is the factor that is used to filter the solutions. This function is only useful for the StochOM and the StochPX solvers.

ACFlow.write_pole — Function

write_pole(Pᵥ::Vector{Vector{I64}},
           Aᵥ::Vector{Vector{F64}},
           𝕊ᵥ::Vector{Vector{F64}},
           χ²::Vector{F64},
           fmesh::AbstractMesh)

Write positions, amplitudes, and signs of poles to pole.data. This function is only useful for the StochPX solver.

ACFlow.write_probability — Function

write_probability(α_vec::Vector{F64}, p_vec::Vector{F64})

Write p(α) data to prob.data. This function is only useful for the MaxEnt solver (bryan algorithm).

ACFlow.write_statistics — Function

write_statistics(MC::StochACMC)

Write Monte Carlo statistical information for the StochAC solver. Note that the StochAC solver is based on a stochastic approach.

See also: PStochAC, StochACMC.

write_statistics(MC::StochSKMC)

Write Monte Carlo statistical information for the StochSK solver. Note that the StochSK solver is based on a stochastic approach.

See also: PStochSK, StochSKMC.

write_statistics(MC::StochOMMC)

Write Monte Carlo statistical information for the StochOM solver. Note that the StochOM solver is based on a stochastic approach.

See also: PStochOM, StochOMMC.

write_statistics(MC::StochPXMC)

Write Monte Carlo statistical information for the StochPX solver. Note that the StochPX solver is based on a stochastic approach.

See also: PStochPX, StochPXMC.