Solvers

Define various solvers for the ACFlow toolkit.

Now the ACFlow toolkit supports seven analytic continuation solvers. They are:

MaxEnt (Maximum Entropy Method, see maxent.jl)

BarRat (Barycentric Rational Function Approximation, see rfa.jl)

NevanAC (Nevanlinna Analytical Continuation, see nac.jl)

StochAC (Stochastic Analytic Continuation, see sac.jl)

StochSK (Stochastic Analytic Continuation, see san.jl)

StochOM (Stochastic Optimization Method, see som.jl)

StochPX (Stochastic Pole Expansion, see spx.jl)

Note

The StochAC solver is based on the Beach's variant, while the StochSK solver is based on the Sandvik's variant.

Contents

Solvers

Contents

Index

Abstract Structs

MaxEnt Solver

Structs

Functions

BarRat Solver

Structs

Functions

NevanAC Solver

Structs

Functions

StochAC Solver

Structs

Functions

StochSK Solver

Structs

Functions

StochOM Solver

Structs

Functions

StochPX Solver

Structs

Functions

Index

ACFlow.AbstractMC

ACFlow.AbstractSolver

ACFlow.BarRatContext

ACFlow.BarRatSolver

ACFlow.BarycentricFunction

ACFlow.Box

ACFlow.MaxEntContext

ACFlow.MaxEntSolver

ACFlow.NevanACContext

ACFlow.NevanACSolver

ACFlow.PronyApproximation

ACFlow.StochACContext

ACFlow.StochACElement

ACFlow.StochACMC

ACFlow.StochACSolver

ACFlow.StochOMContext

ACFlow.StochOMElement

ACFlow.StochOMMC

ACFlow.StochOMSolver

ACFlow.StochPXContext

ACFlow.StochPXElement

ACFlow.StochPXMC

ACFlow.StochPXSolver

ACFlow.StochSKContext

ACFlow.StochSKElement

ACFlow.StochSKMC

ACFlow.StochSKSolver

ACFlow.Pdx

ACFlow.aaa

ACFlow.average

ACFlow.average

ACFlow.average

ACFlow.average

ACFlow.bc_degree

ACFlow.bc_nodes

ACFlow.bc_poles

ACFlow.bc_values

ACFlow.bc_weights

ACFlow.bryan

ACFlow.calc_abcd

ACFlow.calc_alpha

ACFlow.calc_bayes

ACFlow.calc_bayes_od

ACFlow.calc_chi2

ACFlow.calc_chi2

ACFlow.calc_correlator

ACFlow.calc_delta

ACFlow.calc_entropy

ACFlow.calc_entropy_od

ACFlow.calc_error

ACFlow.calc_fmesh

ACFlow.calc_fmesh

ACFlow.calc_fmesh

ACFlow.calc_goodness

ACFlow.calc_green

ACFlow.calc_green

ACFlow.calc_green

ACFlow.calc_green

ACFlow.calc_green

ACFlow.calc_hamil

ACFlow.calc_hbasis

ACFlow.calc_hmatrix

ACFlow.calc_hmin!

ACFlow.calc_hopt!

ACFlow.calc_htau

ACFlow.calc_inv_mobius

ACFlow.calc_lambda

ACFlow.calc_lambda

ACFlow.calc_mobius

ACFlow.calc_noptim

ACFlow.calc_norm

ACFlow.calc_phi

ACFlow.calc_phis

ACFlow.calc_pick

ACFlow.calc_theta

ACFlow.calc_theta

ACFlow.check_causality

ACFlow.check_pick

ACFlow.chi2kink

ACFlow.classic

ACFlow.constraints

ACFlow.constraints

ACFlow.constraints

ACFlow.constraints

ACFlow.eval_lambda

ACFlow.f_and_J

ACFlow.f_and_J_od

ACFlow.hardy_optimize!

ACFlow.historic

ACFlow.init

ACFlow.init

ACFlow.init

ACFlow.init

ACFlow.init

ACFlow.init

ACFlow.init

ACFlow.init_context

ACFlow.init_context

ACFlow.init_element

ACFlow.init_element

ACFlow.init_element

ACFlow.init_element

ACFlow.init_iodata

ACFlow.init_iodata

ACFlow.init_iodata

ACFlow.init_iodata

ACFlow.init_mc

ACFlow.init_mc

ACFlow.init_mc

ACFlow.init_mc

ACFlow.last

ACFlow.last

ACFlow.last

ACFlow.last

ACFlow.last

ACFlow.last

ACFlow.last

ACFlow.measure

ACFlow.measure

ACFlow.measure

ACFlow.optimizer

ACFlow.poles!

ACFlow.precompute

ACFlow.precompute

ACFlow.prony_data

ACFlow.prony_gamma

ACFlow.prony_idx

ACFlow.prony_omega

ACFlow.prony_svd

ACFlow.prony_v

ACFlow.prun

ACFlow.prun

ACFlow.prun

ACFlow.prun

ACFlow.reset_context

ACFlow.reset_element

ACFlow.reset_mc

ACFlow.run

ACFlow.run

ACFlow.run

ACFlow.run

ACFlow.run

ACFlow.run

ACFlow.run

ACFlow.sample

ACFlow.sample

ACFlow.sample

ACFlow.shuffle

ACFlow.smooth_norm

ACFlow.solve

ACFlow.solve

ACFlow.solve

ACFlow.solve

ACFlow.solve

ACFlow.solve

ACFlow.solve

ACFlow.svd_to_real

ACFlow.svd_to_real_od

ACFlow.try_height

ACFlow.try_insert

ACFlow.try_merge

ACFlow.try_move_a

ACFlow.try_move_a

ACFlow.try_move_p

ACFlow.try_move_p

ACFlow.try_move_p

ACFlow.try_move_q

ACFlow.try_move_s

ACFlow.try_move_s

ACFlow.try_move_x

ACFlow.try_move_x

ACFlow.try_remove

ACFlow.try_shift

ACFlow.try_split

ACFlow.try_width

ACFlow.update

ACFlow.warmup

ACFlow.warmup

Abstract Structs

ACFlow.AbstractSolver — Type

AbstractSolver

An abstract type representing the solver for analytic continuation problem. It is used to build the internal type system. All the other solvers are its sub-types.

ACFlow.AbstractMC — Type

AbstractMC

An abstract type representing the Monte Carlo engine. It is used to build the internal type system.

MaxEnt Solver

Structs

ACFlow.MaxEntSolver — Type

MaxEntSolver

It represents the analytic continuation solver that implements the maximum entropy method.

ACFlow.MaxEntContext — Type

MaxEntContext

Mutable struct. It is used within the MaxEnt solver only.

Members

Gᵥ -> Input data for correlator.

σ² -> Actually 1.0 / σ².

grid -> Grid for input data.

mesh -> Mesh for output spectrum.

model -> Default model function.

kernel -> Default kernel function.

Vₛ -> Matrix from singular value decomposition.

W₂ -> Precomputed array.

W₃ -> Precomputed array.

Bₘ -> Precomputed array.

Functions

ACFlow.solve — Method

solve(S::MaxEntSolver, rd::RawData)

Solve the analytic continuation problem by the maximum entropy method.

ACFlow.init — Method

init(S::MaxEntSolver, rd::RawData)

Initialize the MaxEnt solver and return a MaxEntContext struct.

ACFlow.run — Method

run(mec::MaxEntContext)

Perform maximum entropy simulation with different algorithms. Now it supports the historic, classic, bryan, and chi2kink algorithms.

ACFlow.last — Method

last(mec::MaxEntContext, svec::Vector, sol::Dict)

Postprocess the results generated during the maximum entropy simulations. Here sol is the final solution for the analytic continuation problem, while svec contains all the intermediate results (it is a vector of dictionary actually).

ACFlow.historic — Function

historic(mec::MaxEntContext)

Apply the historic algorithm to solve the analytic continuation problem. It choose α in a way that χ² ≈ N.

For the historic algorithm, alpha is usually 10⁶, and ratio is 10.0. It is compatible with the Bayesian Reconstruction entropy.

See also: MaxEntContext.

ACFlow.classic — Function

classic(mec::MaxEntContext)

Apply the classic algorithm to solve the analytic continuation problem.

Classic algorithm uses Bayes statistics to approximately determine the most probable value of α. We always start at a large value of α, where the optimization yields basically the default model, therefore u_vec is only a few steps away from 0 (= default model). And then we gradually decrease α, step by step moving away from the default model towards data fitting. Using u_vec as start for the next (smaller) α brings a great speedup into this procedure.

For the classic algorithm, alpha is usually 10⁶, and ratio is 10.0. It is incompatible with the Bayesian Reconstruction entropy.

See also: MaxEntContext.

ACFlow.bryan — Function

bryan(mec::MaxEntContext)

Apply the bryan algorithm to solve the analytic continuation problem.

Bryan's maxent calculates an average of spectral functions, weighted by their Bayesian probability.

For the bryan algorithm, alpha is usually 500, and ratio is 1.1. It is incompatible with the Bayesian Reconstruction entropy.

See also: MaxEntContext.

ACFlow.chi2kink — Function

chi2kink(mec::MaxEntContext)

Apply the chi2kink algorithm to solve the analytic continuation problem.

We start with an optimization at a large value of α, where we should get only the default model. And then, α is decreased step-by-step, until the minimal value of α is reached. Then, we fit a function

ϕ(x; a, b, c, d) = a + b / [1 + exp(-d*(x-c))],

from which the optimal α is determined by

x_opt = c - fit_position / d,

and

alpha_opt = 10^x_opt.

For the chi2kink algorithm, alpha is usually 10⁹, ratio is 10.0, the number of alpha parameters is 12. It is compatible with the Bayesian Reconstruction entropy.

See also: MaxEntContext.

ACFlow.optimizer — Function

optimizer(
    mec::MaxEntContext,
    α::F64,
    us::Vector{F64},
    use_bayes::Bool
    )

Optimization of maxent functional for a given value of α. Since a priori the best value of α is unknown, this function has to be called several times in order to find a good value.

α means a weight factor of the entropy. us is a vector in singular space. It is used as a starting value for the optimization. For the very first optimization, done at large α, we use zeros, which corresponds to the default model. Then we use the result of the previous optimization as a starting value. use_bayes determines whether to use the Bayesian inference parameters for α.

This function will return a dictionary object that holds the results of the optimization, e.g. spectral function, χ² deviation.

ACFlow.precompute — Method

precompute(
    Gᵥ::Vector{F64},
    σ²::Vector{F64},
    am::AbstractMesh,
    D::Vector{F64},
    K::Matrix{F64}
    )

Precompute some key coefficients. Here Gᵥ and σ² are input data, am is the mesh for spectrum, D is the default model, and K is the kernel function.

ACFlow.f_and_J — Function

f_and_J(u::Vector{F64}, mec::MaxEntContext, α::F64)

This function evaluates the function whose root we want to find. Here u is a singular space vector that parametrizes the spectral function, and α is a (positive) weight factor of the entropy.

It returns f, value of the function whose zero we want to find, and J, jacobian at the current position.

See also: f_and_J_od.

ACFlow.f_and_J_od — Function

f_and_J_od(u::Vector{F64}, mec::MaxEntContext, α::F64)

This function evaluates the function whose root we want to find. Here u is a singular space vector that parametrizes the spectral function, and α is a (positive) weight factor of the entropy.

It returns f, value of the function whose zero we want to find, and J, jacobian at the current position.

This function is similar to f_and_J, but for offdiagonal elements.

See also: f_and_J.

ACFlow.svd_to_real — Function

svd_to_real(mec::MaxEntContext, u::Vector{F64})

Go from singular value space to real space. It will transform the singular space vector u into real-frequency space (to get the spectral function) by A(ω) = D(ω) eⱽᵘ, where D(ω) is the default model V is the matrix from the singular value decomposition. The argument u means a singular space vector that parametrizes the spectral function.

See also: svd_to_real_od.

ACFlow.svd_to_real_od — Function

svd_to_real_od(mec::MaxEntContext, u::Vector{F64})

Go from singular value space to real space. It will transform the singular space vector u into real-frequency space in the case of an offdiagonal element. It will return the spectral function.

See also: svd_to_real.

ACFlow.calc_entropy — Function

calc_entropy(mec::MaxEntContext, A::Vector{F64}, u::Vector{F64})

It computes entropy for positive definite spectral function. Here the arguments A means spectral function and u means a singular space vector that parametrizes the spectral function.

See also: calc_entropy_od.

ACFlow.calc_entropy_od — Function

calc_entropy_od(mec::MaxEntContext, A::Vector{F64})

It compute positive-negative entropy for spectral function with norm 0. Here the argument A means spectral function.

See also: calc_entropy.

ACFlow.calc_bayes — Function

calc_bayes(
    mec::MaxEntContext,
    A::Vector{F64},
    S::F64,
    χ²::F64,
    α::F64
    )

It calculates Bayesian convergence criterion (ng, tr, and conv) for classic maxent (maximum of probablility distribution) and then Bayesian a-posteriori probability (log_prob) for α after optimization of A.

Here, A is the spectral function, S the entropy, χ² the deviation, and α weight factor of the entropy.

See also: calc_bayes_od.

ACFlow.calc_bayes_od — Function

calc_bayes_od(
    mec::MaxEntContext,
    A::Vector{F64},
    S::F64,
    χ²::F64,
    α::F64
    )

It calculates Bayesian convergence criterion (ng, tr, and conv) for classic maxent (maximum of probablility distribution) and then Bayesian a-posteriori probability (log_prob) for α after optimization of A.

Here, A is the spectral function, S the entropy, χ² the deviation, and α weight factor of the entropy.

It is just a offdiagonal version of calc_bayes().

See also: calc_bayes.

ACFlow.calc_chi2 — Method

calc_chi2(mec::MaxEntContext, A::Vector{F64})

It computes the χ²-deviation of the spectral function A.

BarRat Solver

Structs

ACFlow.BarRatSolver — Type

BarRatSolver

It represents the analytic continuation solver that implements the barycentric rational function method.

ACFlow.BarRatContext — Type

BarRatContext

Mutable struct. It is used within the BarRat solver only.

Members

Gᵥ -> Input data for correlator.

grid -> Grid for input data.

mesh -> Mesh for output spectrum.

𝒫 -> Prony approximation for the input data.

ℬ -> Barycentric rational function approximation for the input data.

ℬP -> It means the positions of the poles.

ℬA -> It means the weights / amplitudes of the poles.

ACFlow.BarycentricFunction — Type

BarycentricFunction

Mutable struct. Barycentric representation of a rational function.

Members

nodes -> Nodes of the rational function, $z_i$.

values -> Values of the rational function, $r(z_i)$.

weights -> Weights of the rational function, $w_i$.

wtimesf -> Weighted values of the rational function, $w_i f_i$.

ACFlow.PronyApproximation — Type

PronyApproximation

Mutable struct. Prony approximation to a complex-valued Matsubara function.

Members

𝑁ₚ -> Number of nodes for Prony approximation.

ωₚ -> Non-negative Matsubara frequency.

𝐺ₚ -> Complex values at ωₚ.

Γₚ -> Nodes for Prony approximation, $γ_i$.

Ωₚ -> Weights for Prony approximation, $w_i$.

Functions

ACFlow.solve — Method

solve(S::BarRatSolver, rd::RawData)

Solve the analytic continuation problem by the barycentric rational function method.

ACFlow.init — Method

init(S::BarRatSolver, rd::RawData)

Initialize the BarRat solver and return a BarRatContext struct.

ACFlow.run — Method

run(brc::BarRatContext)

At first, it will try to construct a Prony approximation for the input Matsubara data. Then the Prony approximation is used to build smooth data set (data denoising). Finally, the barycentric rational function for this data set is constructed. The member ℬ of the BarRatContext object (brc) should be updated in this function.

ACFlow.last — Method

last(brc::BarRatContext)

It will process and write the calculated results by the BarRat solver, including correlator at real axis, final spectral function, reproduced correlator. The information about Prony approximation and barycentric rational function approximation will be written as well.

Arguments

brc -> A BarRatContext object.

ACFlow.aaa — Function

aaa(z, y)

Adaptively compute a barycentric rational interpolant.

Arguments

z::AbstractVector{<:Number} -> Interpolation nodes.

y::AbstractVector{<:Number} -> Values at nodes.

max_degree::Integer=150 -> Maximum numerator/denominator degree to use.

float_type::Type=F64 -> Floating point type to use for the computation.

tol::Real=1000*eps(float_type) -> Tolerance for stopping.

lookahead::Integer=10 -> Number of iterations to determines stagnation.

stats::Bool=false -> Return convergence statistics.

Returns

r::BarycentricFunction -> The rational interpolant.

stats::NamedTuple -> Convergence statistics, if keyword stats = true.

Examples

julia> z = 1im * range(-10, 10, 500);

julia> y = @. exp(z);

julia> r = aaa(z, y);

julia> bc_degree(r) # both numerator and denominator
12

julia> first(nodes(r), 4)
4-element Vector{ComplexF64}:
 0.0 - 6.272545090180361im
 0.0 + 9.43887775551102im
 0.0 - 1.1022044088176353im
 0.0 + 4.909819639278557im

julia> r(1im * π / 2)
-2.637151617496356e-15 + 1.0000000000000002im

ACFlow.poles! — Function

poles!(brc::BarRatContext)

Convert the barycentric rational function approximation to the classic pole representation. Note that this feature is only suitable for the atype = "delta" case. In such case, the barycenteric algorithm can find the accurate positions for the poles via the bc_poles() function. But it seems that the weights for these poles are wrong. In this function, we just use the BFGS method to solve this optimization problem to get the correct weights for the poles. And then the positions and weights of these poles will be stored in brc, a BarRatContext object.

ACFlow.bc_nodes — Function

bc_nodes(r::BarycentricFunction)

Returns the nodes of the rational interpolant r as a vector. Actually, they are $z_i$.

ACFlow.bc_values — Function

bc_values(r::BarycentricFunction)

Returns the nodal values of the rational interpolant r as a vector. They are $r(z_i) = f_i$.

ACFlow.bc_weights — Function

bc_weights(r::BarycentricFunction)

Returns the weights of the rational interpolant r as a vector. Actually, they are $w_i$.

ACFlow.bc_degree — Function

bc_degree(r::BarycentricFunction)

Returns the degree of the numerator and denominator of the rational r.

ACFlow.bc_poles — Function

bc_poles(r::BarycentricFunction)

Return the poles of the rational function r.

ACFlow.prony_data — Function

prony_data(ω₁::Vector{F64}, 𝐺₁::Vector{C64})

Prepare essential data for the later Prony approximation. It will return the number of nodes, frequency mesh ωₚ, and Green's function data 𝐺ₚ at this mesh.

ACFlow.prony_svd — Function

prony_svd(𝑁ₚ::I64, 𝐺ₚ::Vector{C64})

Perform singular value decomposition for the matrix ℋ that is constructed from 𝐺ₚ. It will return the singular values S and orthogonal matrix V.

ACFlow.prony_idx — Function

prony_idx(S::Vector{F64}, ε::F64)

The diagonal matrix (singular values) S is used to test whether the threshold ε is reasonable and figure out the index for extracting v from V.

See also: prony_v.

prony_idx(S::Vector{F64})

The diagonal matrix (singular values) S is used to figure out the index for extracting v from V. This function is try to evaluate the maximum index for the exponentially decaying region of S.

See also: prony_v.

ACFlow.prony_v — Function

prony_v(V::Adjoint{C64, Matrix{C64}}, idx::I64)

Extract suitable vector v from orthogonal matrix V according to the threshold ε.

ACFlow.prony_gamma — Function

prony_gamma(v::Vector{C64}, Λ::F64)

Try to calculate Γₚ. Actually, Γₚ are eigenvalues of a matrix constructed by v. Λ is a cutoff for Γₚ. Only those Γₚ that are smaller than Λ are kept.

ACFlow.prony_omega — Function

prony_omega(𝐺ₚ::Vector{C64}, Γₚ::Vector{C64})

Try to calculate Ωₚ.

NevanAC Solver

Structs

ACFlow.NevanACSolver — Type

NevanACSolver

It represents the analytic continuation solver that implements the Nevanlinna analytical continuation (It doesn't support the analytic confinuations for matrix-valued Green's functions).

ACFlow.NevanACContext — Type

NevanACContext

Mutable struct. It is used within the NevanAC solver only.

Members

Gᵥ -> Input data for correlator.

grid -> Grid for input data.

mesh -> Mesh for output spectrum.

Φ -> Φ vector in Schur algorithm.

𝒜 -> Coefficients matrix abcd in Schur algorithm.

ℋ -> Hardy matrix for Hardy basis optimization.

𝑎𝑏 -> Coefficients matrix for expanding θ with Hardy basis.

hmin -> Minimal value of the order of Hardy basis functions.

hopt -> Optimal value of the order of Hardy basis functions.

Functions

ACFlow.solve — Method

solve(S::NevanACSolver, rd::RawData)

Solve the analytic continuation problem by the Nevanlinna analytical continuation method.

ACFlow.init — Method

init(S::NevanACSolver, rd::RawData)

Initialize the NevanAC solver and return a NevanACContext struct.

ACFlow.run — Method

run(nac::NevanACContext)

Perform Hardy basis optimization to smooth the spectrum. the members ℋ, 𝑎𝑏, hmin, and hopt of the NevanACContext object (nac) should be updated in this function.

ACFlow.last — Method

last(nac::NevanACContext)

Postprocess the results generated during the Nevanlinna analytical continuation simulations.

Arguments

nac -> A NevanACContext object.

ACFlow.precompute — Method

precompute(
    grid::AbstractGrid,
    mesh::AbstractMesh,
    Gᵥ::Vector{APC}
    )

Precompute some key quantities, such as Φ, 𝒜, ℋ, and 𝑎𝑏. Note that Φ and 𝒜 won't be changed any more. But ℋ and 𝑎𝑏 should be updated by the Hardy basis optimization to get a smooth spectrum. Here Gᵥ is input data, grid is the grid for input data, and mesh is the mesh for output spectrum.

ACFlow.calc_mobius — Function

calc_mobius(z::Vector{APC})

A direct Mobius transformation.

ACFlow.calc_inv_mobius — Function

calc_inv_mobius(z::Vector{APC})

An inverse Mobius transformation.

ACFlow.calc_pick — Function

calc_pick(k::I64, ℎ::Vector{APC}, λ::Vector{APC})

Try to calculate the Pick matrix, anc check whether it is a positive semidefinite matrix. See Eq. (5) in Fei's NAC paper.

Arguments

k -> Size of the Pick matrix.

ℎ -> Vector ℎ. It is actually 𝑧.

λ -> Vector λ. It is actually 𝒢(𝑧).

ACFlow.calc_phis — Function

calc_phis(grid::AbstractGrid, Gᵥ::Vector{APC})

Try to calculate the Φ vector, which is used to calculate the 𝒜 matrix. Note that Φ should not be changed anymore once it has been established.

Arguments

grid -> Grid in imaginary axis for input Green's function.

Gᵥ -> Input Green's function.

ACFlow.calc_abcd — Function

calc_abcd(grid::AbstractGrid, mesh::AbstractMesh, Φ::Vector{APC})

Try to calculate the coefficients matrix abcd (here it is called 𝒜), which is then used to calculate θ. See Eq. (8) in Fei's NAC paper.

Arguments

grid -> Grid in imaginary axis for input Green's function.

mesh -> Real frequency mesh.

Φ -> Φ vector calculated by calc_phis().

ACFlow.calc_hbasis — Function

calc_hbasis(z::APC, k::I64)

Try to calculate the Hardy basis $f^k(z)$.

Arguments

z -> A complex variable.

k -> Current order for the Hardy basis.

ACFlow.calc_hmatrix — Function

calc_hmatrix(mesh::AbstractMesh, H::I64)

Try to calculate $[f^k(z), f^k(z)^*]$ for 0 ≤ 𝑘 ≤ 𝐻-1, which is called the hardy matrix (ℋ) and is used to evaluate θₘ₊₁.

Arguments

mesh -> Real frequency mesh.

H -> Maximum order for the Hardy basis.

ACFlow.calc_theta — Method

calc_theta(𝒜::Array{APC,3}, ℋ::Array{APC,2}, 𝑎𝑏::Vector{C64})

Try to calculate the contractive function θ(z). 𝒜 is the coefficients matrix abcd, ℋ is the Hardy matrix, and 𝑎𝑏 are complex coefficients for expanding θₘ₊₁. See Eq. (7) in Fei's NAC paper.

Arguments

𝒜 -> Matrix 𝑎𝑏𝑐𝑑.

ℋ -> Hardy matrix.

𝑎𝑏 -> Expansion coefficients 𝑎 and 𝑏 for the contractive function θ.

ACFlow.calc_green — Method

calc_green(𝒜::Array{APC,3}, ℋ::Array{APC,2}, 𝑎𝑏::Vector{C64})

Firstly we try to calculate θ. Then θ is back transformed to a Nevanlinna interpolant via the inverse Mobius transform. Here, 𝒜 (abcd matrix), ℋ (Hardy matrix), and 𝑎𝑏 are used to evaluate θ.

Arguments

𝒜 -> Matrix 𝑎𝑏𝑐𝑑.

ℋ -> Hardy matrix.

𝑎𝑏 -> Expansion coefficients 𝑎 and 𝑏 for the contractive function θ.

ACFlow.calc_noptim — Function

calc_noptim(ωₙ::Vector{APC}, Gₙ::Vector{APC})

Evaluate the optimal value for the size of input data (how may frequency points are actually used in the analytic continuation simulations) via the Pick criterion.

Arguments

ωₙ -> Matsubara frequency points (the 𝑖 unit is not included).

Gₙ -> Matsubara Green's function.

ACFlow.calc_hmin! — Function

calc_hmin!(nac::NevanACContext)

Try to perform Hardy basis optimization. Such that the Hardy matrix ℋ and the corresponding coefficients 𝑎𝑏 are updated. They are used to calculate θ, which is then back transformed to generate smooth G (i.e., the spectrum) at real axis.

This function will determine the minimal value of H (hmin). Of course, ℋ and 𝑎𝑏 in NevanACContext object are also changed.

ACFlow.calc_hopt! — Function

calc_hopt!(nac::NevanACContext)

Try to perform Hardy basis optimization. Such that the Hardy matrix ℋ and the corresponding coefficients 𝑎𝑏 are updated. They are used to calculate θ, which is then back transformed to generate smooth G (i.e., the spectrum) at real axis.

This function will determine the optimal value of H (hopt). Of course, ℋ and 𝑎𝑏 in NevanACContext object are also changed.

ACFlow.hardy_optimize! — Function

hardy_optimize!(
    nac::NevanACContext,
    ℋ::Array{APC,2},
    𝑎𝑏::Vector{C64},
    H::I64
    )

For given Hardy matrix ℋ, try to update the expanding coefficients 𝑎𝑏 by minimizing the smooth norm.

Arguments

nac -> A NevanACContext object.

ℋ -> Hardy matrix, which contains the Hardy basis.

𝑎𝑏 -> Expansion coefficients 𝑎 and 𝑏 for the contractive function θ.

H -> Maximum order of the Hardy basis.

ACFlow.smooth_norm — Function

smooth_norm(nac::NevanACContext, ℋ::Array{APC,2}, 𝑎𝑏::Vector{C64})

Establish the smooth norm, which is used to improve the smoothness of the output spectrum.

Arguments

nac -> A NevanACContext object.

ℋ -> Hardy matrix, which contains the Hardy basis.

𝑎𝑏 -> Expansion coefficients 𝑎 and 𝑏 for the contractive function θ.

ACFlow.check_pick — Function

check_pick(wn::Vector{APC}, gw::Vector{APC}, Nopt::I64)

Check whether the input data are valid (the Pick criterion is satisfied). Here, wn is the Matsubara frequency, gw is the Matsubara function, and Nopt is the optimized number of Matsubara data points.

ACFlow.check_causality — Function

check_causality(ℋ::Array{APC,2}, 𝑎𝑏::Vector{C64})

Check causality of the Hardy coefficients 𝑎𝑏.

StochAC Solver

Structs

ACFlow.StochACSolver — Type

StochACSolver

It represents the analytic continuation solver that implements the stochastic analytic continuation method (K. S. D. Beach's version).

ACFlow.StochACMC — Type

StochACMC

Mutable struct. It is used within the StochAC solver. It includes random number generator and some counters.

Members

rng -> Random number generator.

Macc -> Counter for move operation (accepted).

Mtry -> Counter for move operation (tried).

Sacc -> Counter for swap operation (accepted).

Stry -> Counter for swap operation (tried).

See also: StochACSolver.

ACFlow.StochACElement — Type

StochACElement

Mutable struct. It is used to record the field configurations, which will be sampled by Monte Carlo sweeping procedure.

Members

Γₚ -> It means the positions of the δ functions.

Γₐ -> It means the weights / amplitudes of the δ functions.

ACFlow.StochACContext — Type

StochACContext

Mutable struct. It is used within the StochAC solver only.

Members

Gᵥ -> Input data for correlator.

σ¹ -> Actually 1.0 / σ¹.

allow -> Allowable indices.

grid -> Grid for input data.

mesh -> Mesh for output spectrum.

model -> Default model function.

kernel -> Default kernel function.

Aout -> Calculated spectral function, it is actually ⟨n(x)⟩.

Δ -> Precomputed δ functions.

hτ -> α-resolved h(τ).

Hα -> α-resolved Hc.

Uα -> α-resolved internal energy, it is actually ⟨Hα⟩.

αₗ -> Vector of the α parameters.

Functions

ACFlow.solve — Method

solve(S::StochACSolver, rd::RawData)

Solve the analytic continuation problem by the stochastic analytic continuation algorithm (K. S. D. Beach's version).

ACFlow.init — Method

init(S::StochACSolver, rd::RawData)

Initialize the StochAC solver and return the StochACMC, StochACElement, and StochACContext structs.

ACFlow.run — Method

run(MC::StochACMC, SE::StochACElement, SC::StochACContext)

Perform stochastic analytic continuation simulation, sequential version.

ACFlow.prun — Method

prun(
    S::StochACSolver,
    p1::Dict{String,Vector{Any}},
    p2::Dict{String,Vector{Any}},
    MC::StochACMC,
    SE::StochACElement,
    SC::StochACContext
    )

Perform stochastic analytic continuation simulation, parallel version. The arguments p1 and p2 are copies of PBASE and PStochAC, respectively.

ACFlow.average — Method

average(step::F64, SC::StochACContext)

Postprocess the results generated during the stochastic analytic continuation simulations. It will calculate the spectral functions, and internal energies.

Arguments

step -> How many steps are there in the Monte Carlo samplings.

SC -> A StochACContext object.

ACFlow.last — Method

last(SC::StochACContext, Aout::Array{F64,2}, Uα::Vector{F64})

It will process and write the calculated results by the StochAC solver, including effective hamiltonian, final spectral function, reproduced correlator.

Arguments

SC -> A StochACContext object.

Aout -> α-dependent spectral functions.

Uα -> α-dependent internal energies.

ACFlow.warmup — Method

warmup(MC::StochACMC, SE::StochACElement, SC::StochACContext)

Warmup the Monte Carlo engine to acheieve thermalized equilibrium.

ACFlow.sample — Method

sample(MC::StochACMC, SE::StochACElement, SC::StochACContext)

Perform Monte Carlo sweeps and sample the field configurations.

ACFlow.measure — Method

measure(SE::StochACElement, SC::StochACContext)

Measure the spectral functions and internal energies.

ACFlow.init_mc — Method

init_mc(S::StochACSolver)

Try to create a StochACMC struct.

See also: StochACMC.

ACFlow.init_element — Method

init_element(S::StochACSolver, rng::AbstractRNG, allow::Vector{I64})

Randomize the configurations for future Monte Carlo sampling. It will return a StochACElement object.

Arguments

S -> A StochACSolver object.

rng -> Random number generator.

allow -> Allowed positions for the δ peaks.

See also: StochACElement.

ACFlow.init_iodata — Method

init_iodata(S::StochACSolver, rd::RawData)

Preprocess the input data (rd), then allocate memory for the α-resolved spectral functions.

See also: RawData.

ACFlow.calc_fmesh — Method

calc_fmesh(S::StochACSolver)

Try to calculate very fine (dense) mesh in [wmin, wmax], which is used internally to build the kernel function. Note that this mesh could be non-uniform. If the file fmesh.inp exists, the code will try to load it to initialize the mesh. Or else the code will try to create a linear mesh automatically.

See also: LinearMesh, DynamicMesh.

ACFlow.calc_phi — Function

calc_phi(am::AbstractMesh, model::Vector{F64})

Try to calculate ϕ(ω) function. am is the mesh for calculated spectrum, and model means the default model function.

See also: AbstractMesh, calc_delta.

ACFlow.calc_delta — Function

calc_delta(fmesh::AbstractMesh, ϕ::Vector{F64})

Precompute the δ functions. fmesh is a very dense mesh in [wmin, wmax] and ϕ is the ϕ function.

See also: calc_phi.

ACFlow.calc_hamil — Function

calc_hamil(
    Γₚ::Array{I64,2},
    Γₐ::Array{I64,2},
    kernel::Matrix{F64},
    Gᵥ::Vector{F64}
    )

Initialize h(τ) and H(α) using Eq.(35) and Eq.(36), respectively. Γₚ and Γₐ represent n(x), kernel means the kernel function, Gᵥ is the correlator. Note that kernel and Gᵥ have been rotated into singular space. Please see comments in init() for more details.

See also: calc_htau.

ACFlow.calc_htau — Function

calc_htau(Γₚ, Γₐ, kernel, Gᵥ)

Try to calculate α-dependent h(τ) via Eq.(36). Γₚ and Γₐ represent n(x), kernel means the kernel function, Gᵥ is the correlator. Note that kernel and Gᵥ have been rotated into singular space. Please see comments in init() for more details.

See also: calc_hamil.

ACFlow.calc_alpha — Function

calc_alpha()

Generate a list for the α parameters

ACFlow.constraints — Method

constraints(S::StochACSolver, fmesh::AbstractMesh)

Try to implement the constrained stochastic analytic continuation method. This function will return a collection. It contains all the allowable indices. Be careful, the constrained stochastic analytic continuation method is compatible with the self-adaptive mesh.

Arguments

S -> A StochACSolver object.

fmesh -> Very dense mesh for the δ peaks.

See also: StochACSolver.

ACFlow.try_move_a — Method

try_move_a(
    i::I64,
    MC::StochACMC,
    SE::StochACElement,
    SC::StochACContext
    )

Select two δ functions randomly and then change their weights. Here i means the index for α parameters.

See also: try_move_p.

ACFlow.try_move_p — Method

try_move_p(
    i::I64,
    MC::StochACMC,
    SE::StochACElement,
    SC::StochACContext
    )

Select two δ functions randomly and then change their positions. Here i means the index for α parameters.

See also: try_move_a.

ACFlow.try_move_x — Method

try_move_x(
    MC::StochACMC,
    SE::StochACElement,
    SC::StochACContext
    )

Try to exchange field configurations between two adjacent layers. Because this function involves two layers, so it doesn't need the argument i.

StochSK Solver

Structs

ACFlow.StochSKSolver — Type

StochSKSolver

It represents the analytic continuation solver that implements the stochastic analytic continuation method (A. W. Sandvik's version).

ACFlow.StochSKMC — Type

StochSKMC

Mutable struct. It is used within the StochSK solver. It includes random number generator and some counters.

Members

rng -> Random number generator.

Sacc -> Counter for single-updated operation (accepted).

Stry -> Counter for single-updated operation (tried).

Pacc -> Counter for pair-updated operation (accepted).

Ptry -> Counter for pair-updated operation (tried).

Qacc -> Counter for quadruple-updated operation (accepted).

Qtry -> Counter for quadruple-updated operation (tried).

See also: StochSKSolver.

ACFlow.StochSKElement — Type

StochSKElement

Mutable struct. It is used to record the field configurations, which will be sampled by Monte Carlo sweeping procedure.

In the present implementation of StochSK solver, the amplitudes of the δ functions are fixed. But in principles, they could be sampled in the Monte Carlo procedure.

Members

P -> It means the positions of the δ functions.

A -> It means the weights / amplitudes of the δ functions.

W -> It denotes the window that is used to shift the δ functions.

ACFlow.StochSKContext — Type

StochSKContext

Mutable struct. It is used within the StochSK solver only.

Members

Gᵥ -> Input data for correlator.

Gᵧ -> Generated correlator.

σ¹ -> Actually 1.0 / σ¹.

allow -> Allowable indices.

grid -> Imaginary axis grid for input data.

mesh -> Real frequency mesh for output spectrum.

kernel -> Default kernel function.

Aout -> Calculated spectral function.

χ² -> Current goodness function.

χ²min -> Mininum goodness function.

χ²vec -> Vector of goodness function.

Θ -> Current Θ parameter.

Θvec -> Vector of Θ parameter.

Functions

ACFlow.solve — Method

solve(S::StochSKSolver, rd::RawData)

Solve the analytic continuation problem by the stochastic analytic continuation algorithm (A. W. Sandvik's version).

ACFlow.init — Method

init(S::StochSKSolver, rd::RawData)

Initialize the StochSK solver and return the StochSKMC, StochSKElement, and StochSKContext structs.

ACFlow.run — Method

run(MC::StochSKMC, SE::StochSKElement, SC::StochSKContext)

Perform stochastic analytic continuation simulation, sequential version.

ACFlow.prun — Method

prun(
    S::StochSKSolver,
    p1::Dict{String,Vector{Any}},
    p2::Dict{String,Vector{Any}},
    MC::StochSKMC,
    SE::StochSKElement,
    SC::StochSKContext
    )

Perform stochastic analytic continuation simulation, parallel version. The arguments p1 and p2 are copies of PBASE and PStochSK, respectively.

ACFlow.average — Method

average(step::F64, SC::StochSKContext)

Postprocess the results generated during the stochastic analytic continuation simulations. It will generate the spectral functions.

Arguments

step -> Number of Monte Carlo samplings.

SC -> A StochSKContext object.

ACFlow.last — Method

last(
    SC::StochSKContext,
    Asum::Vector{F64},
    χ²vec::Vector{F64},
    Θvec::Vector{F64}
    )

It will process and write the calculated results by the StochSK solver, including final spectral function and reproduced correlator.

Arguments

SC -> A StochSKContext object.

Asum -> Spectral function.

χ²vec -> Θ-dependent χ².

Θvec -> List of the Θ parameters.

ACFlow.warmup — Method

warmup(MC::StochSKMC, SE::StochSKElement, SC::StochSKContext)

Warmup the Monte Carlo engine to acheieve thermalized equilibrium. Then it will try to figure out the optimized Θ and the corresponding Monte Carlo field configuration.

ACFlow.sample — Method

sample(MC::StochSKMC, SE::StochSKElement, SC::StochSKContext)

Perform Monte Carlo sweeps and sample the field configurations.

ACFlow.measure — Method

measure(SE::StochSKElement, SC::StochSKContext)

Measure the final spectral functions.

See also: nearest.

ACFlow.shuffle — Method

shuffle(MC::StochSKMC, SE::StochSKElement, SC::StochSKContext)

Try to shuffle the Monte Carlo field configuration via the Metropolis algorithm. Then the window for shifting the δ functions is adjusted.

ACFlow.init_mc — Method

init_mc(S::StochSKSolver)

Try to create a StochSKMC struct.

See also: StochSKMC.

ACFlow.init_element — Method

init_element(
    S::StochSKSolver,
    rng::AbstractRNG,
    allow::Vector{I64}
    )

Randomize the configurations for future Monte Carlo sampling. It will return a StochSKElement object.

Arguments

S -> A StochSKSolver object.

rng -> Random number generator.

allow -> Allowed positions for the δ peaks.

See also: StochSKElement.

ACFlow.init_iodata — Method

init_iodata(S::StochSKSolver, rd::RawData)

Preprocess the input data (rd), then allocate memory for the calculated spectral functions.

See also: RawData.

ACFlow.calc_fmesh — Method

calc_fmesh(S::StochSKSolver)

Try to calculate very fine (dense) linear mesh in [wmin, wmax], which is used internally to build the kernel function. Note that the stochastic analytic continuation method (A. W. Sandvik's version) does not support the self-adaptive mesh.

See also: LinearMesh, DynamicMesh.

ACFlow.calc_correlator — Function

calc_correlator(SE::StochSKElement, kernel::Array{F64,2})

Try to calculate correlator with the kernel function and the Monte Carlo field configuration. This correlator will then be used to evaluate the goodness function.

Arguments

SE -> A StochSKElement object.

kernel -> The fermionic or bosonic kernel.

See also: calc_goodness.

ACFlow.calc_goodness — Function

calc_goodness(Gₙ::Vector{F64}, Gᵥ::Vector{F64})

Try to calculate the goodness function (i.e, χ²), which measures the distance between input and regenerated correlators.

Arguments

Gₙ -> Reconstructed correlators.

Gᵥ -> Input (original) correlators.

See also: calc_correlator.

ACFlow.calc_theta — Function

calc_theta(𝒜::Array{APC,3}, ℋ::Array{APC,2}, 𝑎𝑏::Vector{C64})

Try to calculate the contractive function θ(z). 𝒜 is the coefficients matrix abcd, ℋ is the Hardy matrix, and 𝑎𝑏 are complex coefficients for expanding θₘ₊₁. See Eq. (7) in Fei's NAC paper.

Arguments

𝒜 -> Matrix 𝑎𝑏𝑐𝑑.

ℋ -> Hardy matrix.

𝑎𝑏 -> Expansion coefficients 𝑎 and 𝑏 for the contractive function θ.

calc_theta(len::I64, SC::StochSKContext)

Try to locate the optimal Θ and χ². This function implements the chi2min and chi2kink algorithms. Note that the chi2min algorithm is preferred.

ACFlow.constraints — Method

constraints(S::StochSKSolver, fmesh::AbstractMesh)

Try to implement the constrained stochastic analytic continuation method. This function will return a collection. It contains all the allowable indices. Be careful, fmesh should be a fine linear mesh.

Arguments

S -> A StochSKSolver object.

fmesh -> Very dense mesh for the δ peaks.

See also: StochSKSolver.

ACFlow.try_move_s — Method

try_move_s(MC::StochSKMC, SE::StochSKElement, SC::StochSKContext)

Try to update the Monte Carlo field configurations via the Metropolis algorithm. In each update, only single δ function is shifted.

See also: try_move_p.

ACFlow.try_move_p — Method

try_move_p(MC::StochSKMC, SE::StochSKElement, SC::StochSKContext)

Try to update the Monte Carlo field configurations via the Metropolis algorithm. In each update, only a pair of δ functions are shifted.

See also: try_move_s.

ACFlow.try_move_q — Method

try_move_q(MC::StochSKMC, SE::StochSKElement, SC::StochSKContext)

Try to update the Monte Carlo field configurations via the Metropolis algorithm. In each update, four different δ functions are shifted.

See also: try_move_s.

StochOM Solver

Structs

ACFlow.StochOMSolver — Type

StochOMSolver

It represents the analytic continuation solver that implements the stochastic optimization method.

ACFlow.StochOMMC — Type

StochOMMC

Mutable struct. It is used within the StochOM solver. It includes random number generator and some counters.

Members

rng -> Random number generator.

Macc -> Counter for move operation (accepted).

Mtry -> Counter for move operation (tried).

See also: StochOMSolver.

ACFlow.Box — Type

Box

Rectangle. The field configuration consists of many boxes. They exhibit various areas (width × height). We use the Metropolis important sampling algorithm to sample them and evaluate their contributions to the spectrum.

Members

h -> Height of the box.

w -> Width of the box.

c -> Position of the box.

ACFlow.StochOMElement — Type

StochOMElement

Mutable struct. It is used to record the field configurations, which will be sampled by Monte Carlo sweeping procedure.

Members

C -> Field configuration.

Λ -> Contributions of the field configuration to the correlator.

G -> Reproduced correlator.

Δ -> Difference between reproduced and raw correlators.

ACFlow.StochOMContext — Type

StochOMContext

Mutable struct. It is used within the StochOM solver only.

Members

Gᵥ -> Input data for correlator.

σ¹ -> Actually 1.0 / σ¹.

grid -> Grid for input data.

mesh -> Mesh for output spectrum.

Cᵥ -> It is used to record the field configurations for all attempts.

Δᵥ -> It is used to record the errors for all attempts.

𝕊ᵥ -> It is used to interpolate the Λ functions.

Functions

ACFlow.solve — Method

solve(S::StochOMSolver, rd::RawData)

Solve the analytic continuation problem by the stochastic optimization method.

ACFlow.init — Method

init(S::StochOMSolver, rd::RawData)

Initialize the StochOM solver and return the StochOMMC and StochOMContext structs.

ACFlow.run — Method

run(MC::StochOMMC, SC::StochOMContext)

Perform stochastic optimization simulation, sequential version.

ACFlow.prun — Method

prun(
    S::StochOMSolver,
    p1::Dict{String,Vector{Any}},
    p2::Dict{String,Vector{Any}},
    MC::StochOMMC,
    SC::StochOMContext
    )

Perform stochastic optimization simulation, parallel version. The arguments p1 and p2 are copies of PBASE and PStochOM, respectively.

ACFlow.average — Method

average(SC::StochOMContext)

Postprocess the collected results after the stochastic optimization simulations. It will generate the spectral functions.

ACFlow.last — Method

last(SC::StochOMContext, Aout::Vector{F64})

It will process and write the calculated results by the StochOM solver, including final spectral function and reproduced correlator.

Arguments

SC -> A StochOMContext object.

Aout -> Spectral function.

ACFlow.update — Method

update(MC::StochOMMC, SE::StochOMElement, SC::StochOMContext)

Using the Metropolis algorithm to update the field configuration, i.e, a collection of hundreds of boxes.

ACFlow.init_mc — Method

init_mc(S::StochOMSolver)

Try to create a StochOMMC struct.

See also: StochOMMC.

ACFlow.init_element — Method

init_element(MC::StochOMMC, SC::StochOMContext)

Try to initialize a StochOMElement struct.

See also: StochOMElement.

ACFlow.init_iodata — Method

init_iodata(S::StochOMSolver, rd::RawData)

Preprocess the input data (rd).

See also: RawData, GreenData.

ACFlow.init_context — Method

init_context(S::StochOMSolver, grid::AbstractGrid)

Try to initialize the key members of a StochOMContext struct.

See also: StochOMContext.

ACFlow.eval_lambda — Function

eval_lambda(
    r::Box,
    grid::FermionicMatsubaraGrid,
    𝕊::Vector{<:AbstractInterpolation}
    )

Try to calculate the contribution of a given box r to the Λ function. This function works for FermionicMatsubaraGrid only. Because there is an analytic expression for this case, 𝕊 is useless.

Actually, 𝕊 is undefined here. See init_context().

Arguments

r -> A box or rectangle.

grid -> Imaginary axis grid for input data.

𝕊 -> An interpolant.

See also: FermionicMatsubaraGrid.

eval_lambda(
    r::Box,
    grid::FermionicFragmentMatsubaraGrid,
    𝕊::Vector{<:AbstractInterpolation}
    )

Try to calculate the contribution of a given box r to the Λ function. This function works for FermionicFragmentMatsubaraGrid only. Because there is an analytic expression for this case, 𝕊 is useless.

Actually, 𝕊 is undefined here. See init_context().

Arguments

r -> A box or rectangle.

grid -> Imaginary axis grid for input data.

𝕊 -> An interpolant.

See also: FermionicFragmentMatsubaraGrid.

eval_lambda(
    r::Box,
    grid::FermionicImaginaryTimeGrid,
    𝕊::Vector{<:AbstractInterpolation}
    )

Try to calculate the contribution of a given box r to the Λ function. This function works for FermionicImaginaryTimeGrid only. Since there is not analytic expressions for this case, the cubic spline interpolation algorithm is adopted. Here, 𝕊 is initialized in init_context().

Arguments

r -> A box or rectangle.

grid -> Imaginary axis grid for input data.

𝕊 -> An interpolant.

See also: FermionicImaginaryTimeGrid.

eval_lambda(
    r::Box,
    grid::FermionicFragmentTimeGrid,
    𝕊::Vector{<:AbstractInterpolation}
    )

Try to calculate the contribution of a given box r to the Λ function. This function works for FermionicFragmentTimeGrid only. Since there is not analytic expressions for this case, the cubic spline interpolation algorithm is adopted. Here, 𝕊 is initialized in init_context().

Arguments

r -> A box or rectangle.

grid -> Imaginary axis grid for input data.

𝕊 -> An interpolant.

See also: FermionicFragmentTimeGrid.

eval_lambda(
    r::Box,
    grid::BosonicMatsubaraGrid,
    𝕊::Vector{<:AbstractInterpolation}
    )

Try to calculate the contribution of a given box r to the Λ function. This function works for BosonicMatsubaraGrid only. Because there is an analytic expression for this case, 𝕊 is useless.

Actually, 𝕊 is undefined here. See init_context().

Arguments

r -> A box or rectangle.

grid -> Imaginary axis grid for input data.

𝕊 -> An interpolant.

See also: BosonicMatsubaraGrid.

eval_lambda(
    r::Box,
    grid::BosonicFragmentMatsubaraGrid,
    𝕊::Vector{<:AbstractInterpolation}
    )

Try to calculate the contribution of a given box r to the Λ function. This function works for BosonicFragmentMatsubaraGrid only. Because there is an analytic expression for this case, 𝕊 is useless.

Actually, 𝕊 is undefined here. See init_context().

Arguments

r -> A box or rectangle.

grid -> Imaginary axis grid for input data.

𝕊 -> An interpolant.

See also: BosonicFragmentMatsubaraGrid.

eval_lambda(
    r::Box,
    grid::BosonicImaginaryTimeGrid,
    𝕊::Vector{<:AbstractInterpolation}
    )

Try to calculate the contribution of a given box r to the Λ function. This function works for BosonicImaginaryTimeGrid only. Since there is not analytic expressions for this case, the cubic spline interpolation algorithm is adopted. Here, 𝕊 is initialized in init_context().

Arguments

r -> A box or rectangle.

grid -> Imaginary axis grid for input data.

𝕊 -> An interpolant.

See also: BosonicImaginaryTimeGrid.

eval_lambda(
    r::Box,
    grid::BosonicFragmentTimeGrid,
    𝕊::Vector{<:AbstractInterpolation}
    )

Try to calculate the contribution of a given box r to the Λ function. This function works for BosonicFragmentTimeGrid only. Since there is not analytic expressions for this case, the cubic spline interpolation algorithm is adopted. Here, 𝕊 is initialized in init_context().

Arguments

r -> A box or rectangle.

grid -> Imaginary axis grid for input data.

𝕊 -> An interpolant.

See also: BosonicFragmentTimeGrid.

ACFlow.calc_error — Function

calc_error(G::Vector{F64}, Gᵥ::Vector{F64}, σ¹::Vector{F64})

Try to calculate χ². Here Gᵥ and σ¹ denote the raw correlator and related standard deviation. G means the reproduced correlator.

See also: calc_green.

ACFlow.calc_green — Method

calc_green(Λ::Array{F64,2}, nk::I64)

Try to reconstruct the correlator via the field configuration.

See also: calc_error.

ACFlow.calc_norm — Function

calc_norm(C::Vector{Box})

Calculate the total area of all boxes.

ACFlow.constraints — Method

constraints(e₁::F64, e₂::F64)

This function is used to judege whether a given box overlapes with the forbidden zone. Here e₁ and e₂ denote the left and right boundaries of the box.

ACFlow.try_insert — Function

try_insert(
    MC::StochOMMC,
    SE::StochOMElement,
    SC::StochOMContext,
    dacc::F64
    )

Insert a new box into the field configuration.

ACFlow.try_remove — Function

try_remove(
    MC::StochOMMC,
    SE::StochOMElement,
    SC::StochOMContext,
    dacc::F64
    )

Remove an old box from the field configuration.

ACFlow.try_shift — Function

try_shift(
    MC::StochOMMC,
    SE::StochOMElement,
    SC::StochOMContext,
    dacc::F64
    )

Change the position of given box in the field configuration.

ACFlow.try_width — Function

try_width(
    MC::StochOMMC,
    SE::StochOMElement,
    SC::StochOMContext,
    dacc::F64
    )

Change the width and height of given box in the field configuration. Note that the box's area is kept.

ACFlow.try_height — Function

try_height(
    MC::StochOMMC,
    SE::StochOMElement,
    SC::StochOMContext,
    dacc::F64
    )

Change the heights of two given boxes in the field configuration.

ACFlow.try_split — Function

try_split(
    MC::StochOMMC,
    SE::StochOMElement,
    SC::StochOMContext,
    dacc::F64
    )

Split a given box into two boxes in the field configuration.

ACFlow.try_merge — Function

try_merge(
    MC::StochOMMC,
    SE::StochOMElement,
    SC::StochOMContext,
    dacc::F64
    )

Merge two given boxes into one box in the field configuration.

ACFlow.Pdx — Function

Pdx(xmin::F64, xmax::F64, rng::AbstractRNG)

Try to calculate the probability density function.

StochPX Solver

Structs

ACFlow.StochPXSolver — Type

StochPXSolver

It represents the analytic continuation solver that implements the stochastic pole expansion.

ACFlow.StochPXMC — Type

StochPXMC

Mutable struct. It is used within the StochPX solver. It includes random number generator and some counters.

Members

rng -> Random number generator.

Sacc -> Counter for position-updated (type 1) operation (accepted).

Stry -> Counter for position-updated (type 1) operation (tried).

Pacc -> Counter for position-updated (type 2) operation (accepted).

Ptry -> Counter for position-updated (type 2) operation (tried).

Aacc -> Counter for amplitude-updated operation (accepted).

Atry -> Counter for amplitude-updated operation (tried).

Xacc -> Counter for exchange operation (accepted).

Xtry -> Counter for exchange operation (tried).

See also: StochPXSolver.

ACFlow.StochPXElement — Type

StochPXElement

Mutable struct. It is used to record the field configurations, which will be sampled by Monte Carlo sweeping procedure.

For the off-diagonal elements of the matrix-valued Green's function, the signs of the poles (𝕊) could be negative (-1.0). However, for the other cases, 𝕊 is always positive (+1.0).

Members

P -> It means the positions of the poles.

A -> It means the weights / amplitudes of the poles.

𝕊 -> It means the signs of the poles.

ACFlow.StochPXContext — Type

StochPXContext

Mutable struct. It is used within the StochPX solver only.

Members

Gᵥ -> Input data for correlator.

Gᵧ -> Generated correlator.

σ¹ -> Actually 1.0 / σ¹.

allow -> Allowable indices.

grid -> Grid for input data.

mesh -> Mesh for output spectrum.

fmesh -> Very dense mesh for the poles.

Λ -> Precomputed kernel matrix.

Θ -> Artificial inverse temperature.

χ²min -> Minimum of χ²min.

χ² -> Vector of goodness function.

Pᵥ -> Vector of poles' positions.

Aᵥ -> Vector of poles' amplitudes.

𝕊ᵥ -> Vector of poles' signs.

Functions

ACFlow.solve — Method

solve(S::StochPXSolver, rd::RawData)

Solve the analytic continuation problem by the stochastic pole expansion. Note that this solver is still experimental. It is useful for analytic continuation of Matsubara data.

Arguments

S -> A StochPXSolver struct.

rd -> A RawData struct, containing raw data for input correlator.

Returns

mesh -> Real frequency mesh, ω.

Aout -> Spectral function, A(ω).

Gout -> Retarded Green's function, G(ω).

ACFlow.init — Method

init(S::StochPXSolver, rd::RawData)

Initialize the StochPX solver and return the StochPXMC, StochPXElement, and StochPXContext structs. Please don't call this function directly.

Arguments

S -> A StochPXSolver struct.

rd -> A RawData struct, containing raw data for input correlator.

Returns

MC -> A StochPXMC struct.

SE -> A StochPXElement struct.

SC -> A StochPXContext struct.

ACFlow.run — Method

run(MC::StochPXMC, SE::StochPXElement, SC::StochPXContext)

Perform stochastic pole expansion simulation, sequential version.

ACFlow.prun — Method

prun(
    S::StochPXSolver,
    p1::Dict{String,Vector{Any}},
    p2::Dict{String,Vector{Any}},
    MC::StochPXMC,
    SE::StochPXElement,
    SC::StochPXContext
    )

Perform stochastic pole expansion simulation, parallel version. The arguments p1 and p2 are copies of PBASE and PStochPX, respectively.

ACFlow.average — Method

average(SC::StochPXContext)

Postprocess the results generated during the stochastic pole expansion simulations. It will generate the spectral functions, real frequency Green's function, and imaginary frequency Green's function.

ACFlow.last — Method

last(
    SC::StochPXContext,
    Aout::Vector{F64},
    Gout::Vector{C64},
    Gᵣ::Vector{F64}
    )

It will write the calculated results by the StochPX solver, including final spectral function and reproduced correlator.

Arguments

SC -> A StochPXContext object.

Aout -> Spectral function.

Gout -> Retarded Green's function.

Gᵣ -> Reconstructed Green's function.

ACFlow.sample — Method

sample(
    t::I64,
    MC::StochPXMC,
    SE::StochPXElement,
    SC::StochPXContext
    )

Try to search the configuration space to locate the minimum by using the simulated annealing algorithm. Here, t means the t-th attempt.

ACFlow.measure — Method

measure(t::I64, SE::StochPXElement, SC::StochPXContext)

Store Monte Carlo field configurations (positions, amplitudes, and signs of many poles) for the t-th attempt.

ACFlow.init_mc — Method

init_mc(S::StochPXSolver)

Try to create a StochPXMC struct.

See also: StochPXMC.

ACFlow.init_element — Method

init_element(
    S::StochPXSolver,
    rng::AbstractRNG,
    allow::Vector{I64}
    )

Randomize the configurations for future Monte Carlo sampling. It will return a StochPXElement object.

Arguments

S -> A StochPXSolver object.

rng -> Random number generator.

allow -> Allowed positions for the poles.

See also: StochPXElement.

ACFlow.init_iodata — Method

init_iodata(S::StochPXSolver, rd::RawData)

Preprocess the input data (rd).

See also: RawData.

ACFlow.init_context — Method

init_context(S::StochPXSolver)

Try to initialize the key members of a StochPXContext struct.

See also: StochPXContext.

ACFlow.reset_mc — Method

reset_mc(MC::StochPXMC)

Reset the counters in StochPXMC struct.

ACFlow.reset_element — Method

reset_element(
    rng::AbstractRNG,
    allow::Vector{I64},
    SE::StochPXElement
    )

Reset the Monte Carlo field configurations (i.e. positions and amplitudes of the poles). Note that the signs of the poles should not be changed.

ACFlow.reset_context — Method

reset_context(t::I64, SE::StochPXElement, SC::StochPXContext)

Recalculate imaginary frequency Green's function and goodness-of-fit function by new Monte Carlo field configurations for the t-th attempts.

ACFlow.calc_fmesh — Method

calc_fmesh(S::StochPXSolver)

Try to calculate very fine (dense) mesh in [wmin, wmax], which is used internally to represent the possible positions of poles. Note that this mesh could be non-uniform. If the file fmesh.inp exists, the code will try to load it to initialize the mesh. Or else the code will generate a linear mesh automatically.

See also: LinearMesh, DynamicMesh.

ACFlow.calc_lambda — Method

calc_lambda(grid::AbstractGrid, fmesh::AbstractMesh)

Precompute the kernel matrix Λ (Λ ≡ 1 / (iωₙ - ϵ)). It is for the fermionic systems.

Arguments

grid -> Imaginary axis grid for input data.

fmesh -> Very dense mesh in [wmin, wmax].

ACFlow.calc_lambda — Method

calc_lambda(
    grid::AbstractGrid,
    fmesh::AbstractMesh,
    χ₀::F64,
    bsymm::Bool
    )

Precompute the kernel matrix Λ. Here, χ₀ is actually -G(iωₙ = 0). And the argument bsymm is used to distinguish two different bosonic kernels. If bsymm is false, it means that the kernel is boson. If bsymm is true, the kernel is bsymm. This function is for the bosonic systems.

Arguments

grid -> Imaginary axis grid for input data.

fmesh -> Very dense mesh in [wmin, wmax].

χ₀ -> -G(0).

bsymm -> Type of bosonic kernel.

ACFlow.calc_green — Method

calc_green(
    P::Vector{I64},
    A::Vector{F64},
    𝕊::Vector{F64},
    Λ::Array{F64,2}
    )

Reconstruct Green's function at imaginary axis by the pole expansion.

Arguments

P -> Positions of poles.

A -> Amplitudes of poles.

𝕊 -> Signs of poles.

Λ -> Kernel matrix Λ.

ACFlow.calc_green — Method

calc_green(
    P::Vector{I64},
    A::Vector{F64},
    𝕊::Vector{F64},
    mesh::AbstractMesh,
    fmesh::AbstractMesh
    )

Reconstruct Green's function at real axis by the pole expansion. It is for the fermionic systems only.

Arguments

P -> Positions of poles.

A -> Amplitudes of poles.

𝕊 -> Signs of poles.

mesh -> Real frequency mesh for spectral functions.

fmesh -> Very dense real frequency mesh for poles.

ACFlow.calc_green — Method

calc_green(
    P::Vector{I64},
    A::Vector{F64},
    𝕊::Vector{F64},
    mesh::AbstractMesh,
    fmesh::AbstractMesh,
    χ₀::F64,
    bsymm::Bool
    )

Reconstruct Green's function at real axis by the pole expansion. Here, χ₀ is actually -G(iωₙ = 0). And the argument bsymm is used to distinguish two different bosonic kernels. If bsymm is false, it means that the kernel is boson. If bsymm is true, the kernel is bsymm. It is for the bosonic systems only.

Arguments

P -> Positions of poles.

A -> Amplitudes of poles.

𝕊 -> Signs of poles.

mesh -> Real frequency mesh for spectral functions.

fmesh -> Very dense real frequency mesh for poles.

χ₀ -> -G(0).

bsymm -> Type of bosonic kernel.

ACFlow.calc_chi2 — Method

calc_chi2(Gₙ::Vector{F64}, Gᵥ::Vector{F64})

Try to calculate the goodness function (i.e, χ²), which measures the distance between input and regenerated correlators.

Arguments

Gₙ -> Reconstructed Green's function.

Gᵥ -> Original Green's function.

See also: calc_green.

ACFlow.constraints — Method

constraints(S::StochPXSolver, fmesh::AbstractMesh)

Try to implement the constrained stochastic pole expansion. This function will return a collection. It contains all the allowable indices. Be careful, the constrained stochastic pole expansion method is compatible with the self-adaptive mesh.

Arguments

S -> A StochPXSolver object.

fmesh -> Very dense mesh for the poles.

See also: StochPXSolver.

ACFlow.try_move_s — Method

try_move_s(
    t::I64,
    MC::StochPXMC,
    SE::StochPXElement,
    SC::StochPXContext
    )

Change the position of one randomly selected pole.

See also: try_move_p.

ACFlow.try_move_p — Method

try_move_p(
    t::I64,
    MC::StochPXMC,
    SE::StochPXElement,
    SC::StochPXContext
    )

Change the positions of two randomly selected poles.

See also: try_move_s.

ACFlow.try_move_a — Method

try_move_a(
    t::I64,
    MC::StochPXMC,
    SE::StochPXElement,
    SC::StochPXContext
    )

Change the amplitudes of two randomly selected poles.

See also: try_move_x.

ACFlow.try_move_x — Method

try_move_x(
    t::I64,
    MC::StochPXMC,
    SE::StochPXElement,
    SC::StochPXContext
    )

Exchange the amplitudes of two randomly selected poles.

See also: try_move_a.