Types

Define some fundamental types and structs for the ACFlow package.

Contents

Types

Contents

Index

Data Types

Constants

Index

ACFlow.PBASE

ACFlow.PMaxEnt

ACFlow.PStochAC

ACFlow.PStochOM

ACFlow.PStochSK

ACFlow.ADT

ACFlow.AbstractData

ACFlow.DType

ACFlow.GreenData

ACFlow.RawData

Data Types

ACFlow.DType — Type

Customized types. It is used to define the following dicts.

ACFlow.ADT — Type

Customized types. It is used to define the following dicts.

ACFlow.AbstractData — Type

AbstractData

An abstract type representing the input data in imaginary axis. It is used to build the internal type system.

ACFlow.RawData — Type

RawData

Mutable struct. It represent the raw input data. The datatype of raw data may be float or complex.

Members

_grid -> Raw grid for the input data, such as τ or iωₙ.

value -> Raw input data, such as G(τ), G(iωₙ), or Σ(iωₙ).

error -> Error bar (standard deviation) of raw input data, σ.

See also: GreenData.

ACFlow.GreenData — Type

GreenData

Mutable struct. It represents the preprocessed input data.

Members

value -> Preprocessed input data.

error -> Preprocessed error bar.

covar -> Diagonal parts of the covariance matrix, σ².

See also: RawData.

Constants

ACFlow.PBASE — Constant

PBASE

Dictionary for configuration parameters: general setup.

ACFlow.PMaxEnt — Constant

PMaxEnt

Dictionary for configuration parameters: the maximum entropy method.

ACFlow.PStochAC — Constant

PStochAC

Dictionary for configuration parameters: the stochastic analytical continuation method (K. S. D. Beach's version).

ACFlow.PStochSK — Constant

PStochSK

Dictionary for configuration parameters: the stochastic analytical continuation method (A. W. Sandvik's version).

ACFlow.PStochOM — Constant

PStochOM

Dictionary for configuration parameters: the stochastic optimization method.