Read, parse, and extract the configuration parameters.

ACTest._v

ACTest.chk_dict

ACTest.fil_dict

ACTest.get_t

ACTest.inp_toml

ACTest.rev_dict

ACTest.see_dict

Parser

ACTest.inp_toml — Function

inp_toml(f::String, necessary::Bool)

Parse the configuration file (in toml format). It reads the whole file.

Arguments

f -> Filename of configuration.

necessary -> If it is true and configuration is absent, raise an error.

Returns

dict -> A dictionary that contains all the key-value pairs.

ACTest.fil_dict — Function

fil_dict(cfg::Dict{String,Any})

Transfer configurations from dict cfg to internal dict (PTEST). In other words, all the relevant internal dicts should be filled / updated in this function.

Arguments

cfg -> A dict that contains all the configurations (from act.toml).

Returns

N/A

ACTest.rev_dict — Function

rev_dict(TEST::Dict{String,Any})

Setup the configuration dictionary: PTEST.

Arguments

TEST -> A dict that contains configurations from the [Test] block.

Returns

N/A

See also: PTEST.

rev_dict(TEST::Dict{String,Vector{Any}})

Setup the configuration dictionary: PTEST.

Arguments

TEST -> A dict that contains configurations from the [Test] block.

Returns

N/A

See also: PTEST.

ACTest.see_dict — Function

see_dict()

Display all of the relevant configuration parameters to the terminal.

Arguments

N/A

Returns

N/A

See also: fil_dict.

Validation

ACTest.chk_dict — Function

chk_dict()

Validate the correctness and consistency of configurations.

Arguments

N/A

Returns

N/A

See also: fil_dict, _v.

ACTest._v — Function

_v(key::String, val::Array{Any,1})

Verify the value array. Called by chk_dict() function only.

Arguments

key -> Key of parameter.

val -> Value of parameter.

Returns

N/A

See also: chk_dict.

Extraction

ACTest.get_t — Function

get_t(key::String)

Extract configurations from dict: PTEST.

Arguments

key -> Key of parameter.

Returns

value -> Value of parameter.

See also: PTEST.