Intermediate Representation

Tools for the intermediate representation format (adaptor).

Source: ir.jl

Contents

Intermediate Representation

Contents

Index

Functions

Index

ZenCore.ir_adaptor

ZenCore.ir_read

ZenCore.ir_save

ZenCore.irio_charge

ZenCore.irio_eigen

ZenCore.irio_fermi

ZenCore.irio_groups

ZenCore.irio_kmesh

ZenCore.irio_lattice

ZenCore.irio_maps

ZenCore.irio_params

ZenCore.irio_projs

ZenCore.irio_rawcp

ZenCore.irio_tetra

ZenCore.irio_windows

Functions

ZenCore.ir_adaptor — Function

ir_adaptor(D::Dict{Symbol,Any})

Write the Kohn-Sham dataset to specified files using the IR format. Note that the Kohn-Sham dataset are encapsulated in the D dict.

See also: plo_adaptor, wannier_adaptor.

ZenCore.ir_read — Function

ir_read(f::String)

Read and parse maps.ir, groups.ir, windows.ir, lattice.ir, kmesh.ir, eigen.ir, projs.ir, fermi.ir, and rawcp.ir. The data are encapsulated in a dictionary. Here f means the directory where the files as mentioned above are available.

See also: ir_adaptor.

ZenCore.ir_save — Function

ir_save(it::IterInfo)

Backup the files outputed by the adaptor.

See also: ir_adaptor.

ZenCore.irio_params — Function

irio_params(f::String, D::Dict{Symbol,Any})

Write the key parameters extracted from the Kohn-Sham dataset. Here f means only the directory that we want to use.

See also: PrGroup, PrWindow.

ZenCore.irio_maps — Function

irio_maps(f::String, MAP::Mapping)

Write the information contained in Mapping struct. Here f means only the directory that we want to use.

See also: Mapping.

irio_maps(f::String)

Extract the Mapping struct from maps.ir. Here f means the directory that this file exists.

See also: Mapping.

ZenCore.irio_groups — Function

irio_groups(f::String, PG::Array{PrGroup,1})

Write the information contained in PrGroup struct. Here f means only the directory that we want to use.

See also: PrGroup.

irio_groups(f::String)

Extract the PrGroup struct from groups.ir. Here f means the directory that this file exists. Be careful, the returned PrGroup is not completely the same with the true one. For example, its Pr and Tr fields are not correct. But it doesn't matter.

See also: PrGroup.

ZenCore.irio_windows — Function

irio_windows(f::String, PW::Array{PrWindow,1})

Write the information contained in PrWindow struct. Here f means only the directory that we want to use.

See also: PrWindow.

irio_windows(f::String)

Extract the PrWindow struct from windows.ir. Here f means the directory that this file exists.

See also: PrWindow.

ZenCore.irio_lattice — Function

irio_lattice(f::String, latt::Lattice)

Write the lattice information to lattice.ir using the IR format. Here f means only the directory that we want to use.

See also: vaspio_lattice, qeio_lattice.

irio_lattice(f::String)

Extract the lattice information from lattice.ir. Here f means the directory that this file exists.

See also: Lattice.

ZenCore.irio_kmesh — Function

irio_kmesh(f::String, kmesh::Array{F64,2}, weight::Array{F64,1})

Write the kmesh and weight information to kmesh.ir using the IR format. Here f means only the directory that we want to use.

See also: vaspio_kmesh.

irio_kmesh(f::String)

Extract the kmesh and weight information from kmesh.ir. Here f means the directory that this file exists.

ZenCore.irio_tetra — Function

irio_tetra(f::String, volt::F64, itet::Array{I64,2})

Write the tetrahedra information to tetra.ir using the IR format. Here f means only the directory that we want to use.

See also: vaspio_tetra.

irio_tetra(f::String)

Extract the tetrahedra information from tetra.ir. Here f means the directory that this file exists.

ZenCore.irio_eigen — Function

irio_eigen(f::String, enk::Array{F64,3}, occupy::Array{F64,3})

Write the eigenvalues to eigen.ir using the IR format. Here f means only the directory that we want to use.

See also: vaspio_eigen, qeio_eigen.

irio_eigen(f::String)

Extract the eigenvalues from eigen.ir. Here f means the directory that this file exists.

ZenCore.irio_projs — Function

irio_projs(f::String, chipsi::Array{Array{C64,4},1})

Write the projectors to projs.ir using the IR format. Here f means only the directory that we want to use.

The projectors have been processed to fulfill the requirement of the DMFT engine.

See also: vaspio_projs.

irio_projs(f::String)

Extract the normalized projectors from projs.ir. Here f means the directory that this file exists.

ZenCore.irio_fermi — Function

irio_fermi(f::String, fermi::F64)

Write the fermi level to fermi.ir using the IR format. Here f means only the directory that we want to use.

See also: vaspio_fermi, qeio_fermi.

irio_fermi(f::String)

Extract the fermi level from fermi.ir. Here f means the directory that this file exists.

ZenCore.irio_charge — Function

irio_charge(f::String)

Write the charge density to charge.ir using the IR format. Here f means only the directory that we want to use.

See also: vaspio_charge.

ZenCore.irio_rawcp — Function

irio_rawcp(f::String, chipsi::Array{C64,4})

Write the projectors to rawcp.ir using the IR format. Here f means only the directory that we want to use.

The projectors are original data. They have not been modified.

See also: irio_projs.

irio_rawcp(f::String)

Extract the raw projectors from rawcp.ir. Here f means the directory that this file exists.